### 2026-04-04 #765 @Rodotasso
ciecl: International Classification of Diseases ICD-10/ICD-11 for Chile
Submitting Author Name: Rodolfo Tasso Suazo
Submitting Author Github Handle: @Rodotasso
Other Package Authors Github handles: (comma separated, delete if none)
Repository: https://github.com/Rodotasso/ciecl
Version submitted: 0.9.6
Submission type: Standard

Version accepted: TBD
Language: es

---

```
Package: ciecl
Title: International Classification of Diseases 'ICD-10'/'ICD-11' for Chile
Version: 0.9.6
    person("Rodolfo", "Tasso Suazo",
           email = "rtasso@uchile.cl",
Description: Tools for working with the International Classification of
    Diseases ('ICD-10' Chile official 'MINSAL'/'DEIS' v2018). Includes
    optimized 'SQL' search with 'SQLite', fuzzy matching of medical terms
    ('Jaro-Winkler'), Charlson and Elixhauser comorbidity calculation,
    'WHO' 'ICD-11' 'API' integration, and hierarchical code validation.
    Data from Centro FIC Chile 'DEIS' <https://deis.minsal.cl/centrofic/>.
Imports: DBI, RSQLite, stringdist, stringr, dplyr, tibble, tools, utils
Suggests: comorbidity, gt, httr2, readxl, usethis, withr, writexl,
    testthat (>= 3.1.5), knitr, litedown (>= 0.9), rmarkdown
    https://github.com/RodoTasso/ciecl
```

**Explain how/why the package fits into these categories:**

`ciecl` provides tools for working with Chile's official ICD-10 classification (MINSAL/DEIS v2018, 39,877 codes). It fits **data munging** (normalizing, expanding, and validating ICD-10 codes from messy clinical data), **data validation** (validating code format and existence against the official dataset), and **database bindings** (SQLite-backed search with FTS5 full-text indexing for fast lookups).

**Target audience and scientific applications:**

Health researchers, epidemiologists, biostatisticians, and public health analysts working with Chilean hospital discharge records, morbidity data, and clinical registries. The package is designed for the School of Public Health (Universidad de Chile) Data Science for Public Health group and the broader health informatics community in Chile and Latin America.

**Are there other R packages that accomplish the same thing?**

- `icd` (CRAN): US-focused ICD-9/10 comorbidity calculations. Does not include Chilean codes or MINSAL/DEIS classifications.
- `comorbidity` (CRAN): Comorbidity indices from ICD codes. `ciecl` uses it as optional backend and adds Chile-specific code search, fuzzy matching, and the complete MINSAL dataset.
- `ICD10gm` (CRAN): German ICD-10-GM modification. Country-specific like `ciecl` but for Germany.
- No existing R package provides the Chilean ICD-10 classification with Spanish-language fuzzy search and medical abbreviation support.

**Ethics, data privacy, human subjects:**

The package contains only publicly available classification codes from MINSAL/DEIS (government open data). No patient data is included or processed. The WHO ICD-11 API integration uses publicly available endpoints. No ethical concerns apply.

**Pre-submission inquiry link:** N/A

**Explain any `pkgcheck` items which your package is unable to pass:**

The goodpractice linter warnings are for line lengths >80 chars in test files (not in source code) and 95.6% coverage (vs 100%).

- [x] I have read the [author guide](https://devguide.ropensci.org/softwarereview_author.html) and I expect to maintain this package for at least 2 years.

- [x] No Terms of Service violations
- [x] CRAN and OSI accepted license (MIT)
- [x] README with development installation instructions
- [x] Roxygen2 documentation with examples for all exported functions
- [x] Vignettes (5: English guide, Spanish guide, hospital discharge use case, language guide, installation)
- [x] Test suite (1148 tests, 95.6% coverage)
- [x] Continuous integration with coverage reporting (GitHub Actions: R-CMD-check, test-coverage)

## Generative AI disclosure

- [x] Generative AI tools were used in the development of this package.

AI tools (Claude Code / Claude Opus) were used to assist with: test suite expansion, documentation translation (Spanish - English), CI/CD workflow configuration, and code review. All AI-generated code was reviewed and validated by the author. The package repository is at https://github.com/Rodotasso/ciecl.

- [x] This package is currently on CRAN (v0.9.3).
- [ ] This package is on Bioconductor.
- [ ] This package is intended for submission to Methods in Ecology and Evolution.
