### 2026-04-04 #765 @Rodotasso
Hi @emilyriederer, thanks for the quick response! Yes, the pkgdown site is live at https://rodotasso.github.io/ciecl/. It's also listed in the DESCRIPTION URL field, but I notice the link might not be prominently visible from the GitHub repo page. I'll make sure it's easy to find. Thanks!
