### 2026-04-04 55aba19
docs: agregar badge pkgdown en READMEs
