### 2026-04-04 e20370a
docs: agregar badge pkgdown en READMEs
