### 2026-04-10 #765 @Rodotasso
Hi @emilyriederer !, idk if i have to let a msg for the pkgdown badge in the readme on repo, but i've done, sorry if i'm late. :)
