### 2026-04-10 #765 @emilyriederer
Thanks @Rodotasso and apologies for the slow reply! A few more initial questions:

- I notice the package has a very large amount of stored data. Did you consider having a function to let users download this versus shipping it inside of the package?
- I notice the main pkgdown website is bilingual English/Spanish (very nice!) but the function docs are in only Spanish. Is that true? I'm thinking what the best language for an editor would be
