### 2026-04-10 #765 @Rodotasso
Tyvm @emilyriederer for the reply and the questions!

  On the stored data - I think you're referring to data/cie10_cl.rda? That file is just 289 KB zipped. The total number of codes (39,873) might give the impression of a large file, but the compressed .rda is quite small. The tarball on CRAN for v0.9.3 is around 2.6 MB in total. If you were referring to a different file, I'm happy to clarify.

  One thing worth noting: in v0.9.3 I already removed a pre-built SQLite cache (~21 MB) that used to ship in inst/extdata/. It's now built on first use from the .rda into the user cache directory (tools::R_user_dir("ciecl", "data")), so the installed package stays lean.

  On the bilingual question: yes, the pkgdown site and external documentation are bilingual, but the function docs are deliberately in Spanish because of the target users' context in Chile. The package is meant to support interoperability across real-world use cases - research, administrative, clinical, and even health-payment systems (GRD/DRG) - and in all of those domains the canonical CIE-10-CL terminology only exists in Spanish. Going English-first would create a gap between the code and the terminology users actually work with every day.

If it helps, I'm happy to consider adding English titles/descriptions in roxygen, or bilingual comments. Just let me know what works best on your end.
