### 2026-04-19 3e96045
chore: eliminar CLAUDE.md redundante del paquete
Info de agentes release consolidada en el CLAUDE.md del workspace.
Evita duplicacion con D:/MAGISTER/01_Paquete_R/CLAUDE.md.
