+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,25 +1,0 @@
-## Descripcion
-
-Breve descripcion de los cambios y su motivacion.
-
-## Tipo de cambio
-
-- [ ] `fix:` Correccion de bug
-- [ ] `feat:` Nueva funcionalidad
-- [ ] `docs:` Documentacion
-- [ ] `refactor:` Refactorizacion sin cambio de comportamiento
-- [ ] `test:` Agregar o mejorar tests
-- [ ] `chore:` Mantenimiento y configuracion
-
-## Cambios realizados
-
--
--
-
-## Checklist
-
-- [ ] `devtools::test()` pasa sin errores
-- [ ] `devtools::check(error_on = "error")` pasa sin errores ni warnings
-- [ ] Documentacion roxygen2 actualizada (`devtools::document()`)
-- [ ] PR dirigido a la rama `dev` (no a `main`)
-- [ ] NEWS.md actualizado si corresponde
