+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   ciecl: ciecl.html
   idiomas: idiomas.html
   instalacion: instalacion.html
-last_built: 2026-04-04T20:39Z
+last_built: 2026-04-19T00:45Z
 urls:
   reference: https://rodotasso.github.io/ciecl/reference
   article: https://rodotasso.github.io/ciecl/articles
