### 2026-04-21 #765 @maelle
@Rodotasso a good place where to add the pkgdown URL is the repository URL, below the description:
