### 2026-04-22 #765 @maelle
---

Thank you for your submission!

First, some concerns/questions about languages:

- If you expect most users of this package to speak Spanish, why not have the README in Spanish? To be honest, I wouldn't even have expected any translation in English of your package.
- Can you explain your reasoning behind the mix of languages in functions' names and arguments' names? E.g. `cie_lookup()` is in English, then there's a parameter called `expandir` but another one is called `extract` not `extraer`. I totally get working in mixed languages (I'm a French native speaker!) but I wonder whether all function names and arguments should be in Spanish? It'd make it easier to remember what things are called (as opposed to having to remember some argument is in Spanish, in a function named in English). As your package is on CRAN, this would necessitate some deprecations.
- Regarding the section about pending translations, why would you add data from Brasil or in French? Can you explain the use cases and how it fits within the scope of this package?
- Do you want the remainder of this submission's thread to happen in Spanish? I expect your package will get reviews in Spanish, so we might as well be consistent.
- The vignette about languages state "tildes preservades" for words without tildes. Furthermore, "español" is written "espanol" in that vignette. (the feature in the package to be able to use rinon or riñon is very nice!)

##### Documentation

- In the README the "main features" come up twice, it should be streamlined.
- Make sure all code in the README and vignette is executed, so that the reader can see the results.
- The intro in Spanish, and some other vignettes, have very little text accompanying the code chunk (to comment on the code, to comment on the results).
- I like the idea of the "Caso de uso" vignette but I think the introduction should state more clearly what the use case is. Being a researcher who was given some hospital data and wanted to analyse them for what? It could even feature comorbidity.
- In the "Caso de uso" vignette, the codes are normalized but without using `cie_normalizar()`.

##### Features

The "Caso de uso" vignette made me realize it'd be nice to have a function to create the column with the code meaning without doing any left join. Something like this (but with better names!):

```r
datos_filtrados %>%
  mutate(diagnostico = codigo_hasta_diagnostico(DIAG1))
```

The demo of `cie_table()` in a vignette creates a table with two columns full of NA, incluye and escluye, why is that?

Instead of https://github.com/Rodotasso/ciecl/blob/3aec054a252d0eee903e47324fede61aba2da86b/R/cie-data.R#L19 you could use [`rlang::check_installed()`](https://rlang.r-lib.org/reference/is_installed.html) that directly prompts the user to install the missing package (but with a message in English).

The package expects a filename with a space in it? 😰 https://github.com/Rodotasso/ciecl/blob/3aec054a252d0eee903e47324fede61aba2da86b/R/cie-data.R#L108

##### Code

Code like `names(raw)[stringr::str_detect(names(raw), "c[o\u00f3]d|clave")]` could be more direct. I don't know the stringr equivalent to `grepv()`, maybe `stringr::str_match()`? https://github.com/Rodotasso/ciecl/blob/3aec054a252d0eee903e47324fede61aba2da86b/R/cie-data.R#L42 (this pattern comes up several time in ciecl)

I'm a bit surprised at https://github.com/Rodotasso/ciecl/blob/3aec054a252d0eee903e47324fede61aba2da86b/R/cie-data.R#L125. I think it's more frequent to save cached data in an [user directory](https://blog.r-hub.io/2020/03/12/user-preferences/). I've never seen `use_data()` called from package code. This way you could drop the need for the user to install usethis to generate the data.

##### Installation instructions

- In the README and dedicated vignette you can simply show pak and drop the alternatives.
- Are you sure the system dependencies stuff is needed for macOS and Ubuntu? Especially as your GitHub Actions workflow for R CMD check does not have special checks. Is this related to problems users of the packages had when installing it?

##### API key

You could recommend users save the key using the keyring R package. Example: https://docs.ropensci.org/babeldown/#api-key

##### Tests

- I'd suggest using the withr package for setting environment variables locally. https://withr.r-lib.org/reference/with_envvar.html
- https://github.com/Rodotasso/ciecl/blob/3aec054a252d0eee903e47324fede61aba2da86b/tests/testthat/test-cie-api.R#L23 should use `skip_if_not_installed()` (that is used in other instances).
- `expect_false(any(grepl`: see https://testthat.r-lib.org/reference/expect_match.html (same for e.g. https://github.com/Rodotasso/ciecl/blob/3aec054a252d0eee903e47324fede61aba2da86b/tests/testthat/test-cie-search.R#L8, maybe in other places)
- `skip_if_offline()` also skips on CRAN.
- `expect_equal(nrow(resultado), 0)` could use [`expect_shape()`](https://testthat.r-lib.org/reference/expect_length.html) (same for https://github.com/Rodotasso/ciecl/blob/3aec054a252d0eee903e47324fede61aba2da86b/tests/testthat/test-cie-search.R#L15, maybe in other places)
- why use `gc()`? https://github.com/Rodotasso/ciecl/blob/3aec054a252d0eee903e47324fede61aba2da86b/tests/testthat/test-robustness.R#L66

---

Happy to discuss any item. Thanks again for your submission!
