### 2026-04-24 #765 @Rodotasso
Hi, sorry for the late, thanks for all the answers!.

Your suggestions are enrichment all my work making the review through time of all the commits. I' working now on every change to maintain coherence between args and functions. If we can make the review in spanish i'll be grateful.

I'm trying to set all your suggestions according to coherence in code and show in one post :P, but it takes more time than expected, so I choose to post first and ask how we can continue in Spanish.
