### 2026-04-24 #765 @maelle
Hola @Rodotasso! Podemos empezar ahora a hablar español aqui. 😸

La semana que viene intentaré traducir mis comentarios iniciales y enviarlos de nuevo.
