### 2026-04-24 #765 @Rodotasso
Muy agradecido @maelle, no es necesario traducir los comentarios iniciales, podemos trabajar desde aquí. 😄
En estos momentos estoy modificado las funciones y argumentos para que se escriban en inglés (el ecosistema funciona así, es más fácil trabajar entre librerías), pero los comentarios de código y documentación esté en español, así es más fácil entender y reproducir el código. Tu sugerencia ha sido muy útil para encontrar la forma más práctica en escritura y reproducibilidad.

El README en español estaba y era el principal, pero al aplicar para CRAN y al levantar el sitio con pkgdown, necesitaba documentación en Inglés para facilitar a los revisores el trabajo. Pero cometí un error con pkgdown y dejé 2 archivos .Rmd, cuando solo debería ser uno que cree los README en ambos idiomas.

En estos momentos estoy trabajando en una branch específica con tus sugerencias, espero hacer push de esa branch al repo (no al main) durante el día para mantener el registro. 🦖
