### 2026-04-24 3e56978
feat: respuesta a revision editorial rOpenSci (v0.10.0-dev)
Idioma (punto principal del editor)
- Unifica convencion: funciones y argumentos en ingles, documentacion
  y mensajes en espanol. Decision registrada en NEWS.md.
- cie_normalizar() -> cie_normalize(). Alias deprecated via
  lifecycle::deprecate_warn("0.10.0", ...); codigo CRAN no se rompe.
- Argumentos renombrados con lifecycle::deprecated():
    cie_lookup: codigo/expandir/normalizar/descripcion_completa
      -> code/expand/normalize/full_description
    cie_normalize: codigos/buscar_db -> codes/search_db
    cie_validate_vector: c
[…]
