+++ tests/testthat/test-robustness.R
@@ -61,9 +61,7 @@ test_that("queries grandes no causan problemas de memoria", {
   expect_s3_class(resultado, "tbl_df")
   expect_lte(nrow(resultado), 5000)
 
-  # Liberar memoria
   rm(resultado)
-  gc()
 })
 
 test_that("cie_search con muchos resultados funciona", {
@@ -329,7 +327,7 @@ test_that("cie_lookup con expansion de categoria completa", {
   skip_on_cran()
 
   # Expandir categoria E (muy grande)
-  hijos <- cie_lookup("E", expandir = TRUE)
+  hijos <- cie_lookup("E", expand = TRUE)
   expect_s3_class(hijos, "tbl_df")
   expect_gt(nrow(hijos), 100)
 })
