### 2026-04-25 39b238d
feat: renombrar funciones a EN + version 0.9.8 (rev-ropensci Bloques A/C/D)
- cie_normalize -> cie_norm (shim deprecated conservado)
- cie_siglas -> cie_short (shim deprecated conservado)
- cie_guia_busqueda -> cie_guide (shim deprecated conservado)
- Migracion argumentos ES->EN con lifecycle::deprecate_warn("0.9.8")
- cie_norm: elimina simbolo + de codificacion dual (fix)
- cie_normalize shim: emite deprecate_warn para codigos= y buscar_db=
- @rdname cie_normalize -> @rdname cie_norm (jerarquia pkgdown correcta)
- Version bump 0.9.6 -> 0.9.8
