+++ R/cie-search.R
@@ -458,17 +458,17 @@ sigla_to_codigo <- function(sigla) {
 
 #' Extraer codigo CIE-10 de texto con ruido
 #'
-#' @param texto Character vector que puede contener prefijos/sufijos
+#' @param text Character vector que puede contener prefijos/sufijos
 #' @return Character vector con codigo CIE-10 extraido o original
 #' @keywords internal
 #' @noRd
-extract_cie_from_text <- function(texto) {
+extract_cie_from_text <- function(text) {
   # Fix #2: Patron estricto para extraer codigo CIE-10:
   # letra + 2-3 digitos + punto opcional + 0-2 digitos
   # Solo extrae si esta rodeado de no-alfanumericos o en extremos
   patron <- "(?:^|[^A-Z0-9])([A-Z][0-9]{2}[0-9]?\\.?[0-9X]{0,2})(?:$|[^A-Z0-9])"
   
-  extraido <- stringr::str_extract(toupper(texto), patron)
+  extraido <- stringr::str_extract(toupper(text), patron)
   
   # Extraer grupo capturado (quitar prefijos/sufijos)
   if (!is.na(extraido) && extraido != "") {
@@ -479,7 +479,7 @@ extract_cie_from_text <- function(texto) {
   resultado <- ifelse(
     !is.na(extraido) & extraido != "",
     extraido,
-    texto
+    text
   )
   
   return(resultado)
@@ -487,28 +487,42 @@ extract_cie_from_text <- function(texto) {
 
 #' Listar siglas medicas soportadas
 #'
-#' Muestra todas las siglas medicas que pueden usarse en cie_search().
+#' Muestra todas las siglas medicas que pueden usarse en [cie_search()].
+#' "Sigla" se conserva como termino local (concepto medico chileno) en la
+#' columna de salida; el nombre de la funcion usa `cie_short` por
+#' consistencia con el ecosistema R (verbos cortos en ingles).
 #'
-#' @param categoria Character opcional, filtrar por categoria. Valores validos:
+#' @param category Character opcional, filtrar por categoria. Valores validos:
 #'   "cardiovascular", "respiratoria", "metabolica", "gastrointestinal",
 #'   "infecciosa", "oncologica", "reumatologica", "neurologica",
 #'   "psiquiatrica", "traumatologica", "pediatrica", "gineco_obstetrica".
 #'   Si es NULL (default), retorna todas las siglas.
+#' @param categoria `r lifecycle::badge("deprecated")` Use `category`.
 #' @return tibble con columnas: sigla, termino_busqueda, categoria
 #' @family busqueda
 #' @seealso \code{\link{cie_search}}, \code{\link{cie_lookup}}
 #' @export
 #' @examples
 #' # Ver todas las siglas
-#' cie_siglas()
+#' cie_short()
 #'
 #' # Filtrar por categoria
-#' cie_siglas("cardiovascular")
-#' cie_siglas("oncologica")
+#' cie_short("cardiovascular")
+#' cie_short("oncologica")
 #'
 #' # Buscar una sigla especifica
-#' cie_siglas() |> dplyr::filter(sigla == "iam")
-cie_siglas <- function(categoria = NULL) {
+#' cie_short() |> dplyr::filter(sigla == "iam")
+cie_short <- function(category = NULL,
+                      categoria = lifecycle::deprecated()) {
+  if (lifecycle::is_present(categoria)) {
+    lifecycle::deprecate_warn(
+      "0.9.8",
+      "cie_short(categoria = )",
+      "cie_short(category = )"
+    )
+    category <- categoria
+  }
+
   siglas <- get_siglas_medicas()
 
[…]
