### 2026-04-25 332d4a9
docs: reescribir vignettes con narrativa completa (rev-ropensci Bloque B)
- ciecl-es.Rmd: narrativa entre chunks, cie_describe(), version 0.9.8, expand=TRUE
- ciecl.Rmd: idem en ingles
- caso-uso-egresos.Rmd: eliminar referencia a cie_normalize, comorbilidad a eval=FALSE
- idiomas.Rmd: cie_siglas -> cie_short, cie_normalizar -> cie_norm
- instalacion.Rmd: corregir texto sobre disponibilidad en CRAN
