+++ vignettes/ciecl-es.Rmd
@@ -1,10 +1,10 @@
 ---
-title: "Introduccion a ciecl: CIE-10 Chile en R"
+title: "Introducción a ciecl: CIE-10 Chile en R"
 author: "Rodolfo Tasso Suazo"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Introduccion a ciecl: CIE-10 Chile en R}
+  %\VignetteIndexEntry{Introducción a ciecl: CIE-10 Chile en R}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -19,78 +19,131 @@ knitr::opts_chunk$set(
 library(ciecl)
 ```
 
-> **Version 0.9.3**: Disponible en CRAN con optimizaciones SQLite y soporte XLSX.
+> **Versión 0.9.8**: Disponible en CRAN con optimizaciones SQLite, soporte XLSX y argumentos estandarizados en inglés.
 
-## Acerca del dataset
+## El problema: codificar diagnósticos en Chile
 
-El paquete incluye **39,877 codigos CIE-10** del catalogo oficial MINSAL/DEIS v2018,
-incluyendo categorias (3 digitos) y subcategorias (4+ digitos).
+Los sistemas de información sanitaria chilenos —DEIS, GRD, REM— almacenan diagnósticos usando la clasificación CIE-10 en su versión oficial MINSAL/DEIS v2018. Trabajar con estos datos en R sin una referencia local obliga al analista a consultar PDFs, tablas Excel o sitios web, interrumpiendo el flujo de análisis.
 
-## Instalacion
- 
+`ciecl` resuelve este problema: incorpora los **39.877 códigos CIE-10** del catálogo oficial directamente en R, con funciones de búsqueda rápida, expansión jerárquica y cálculo de índices de comorbilidad.
+
+## Instalación
+
+El paquete está disponible en CRAN. Para instalar la versión estable:
+
 ```{r eval=FALSE}
-# Desde GitHub (version beta)
-pak::pak("RodoTasso/ciecl")
-
-# Alternativa con devtools
-devtools::install_github("RodoTasso/ciecl")
+install.packages("ciecl")
 ```
 
-## Busqueda SQL rapida
+Para la versión de desarrollo con las últimas correcciones:
+
+```{r eval=FALSE}
+# Requiere el paquete pak
+pak::pak("RodoTasso/ciecl")
+```
+
+## Consultas SQL directas al catálogo
+
+La función `cie10_sql()` expone el catálogo completo a través de SQLite, lo que permite filtrar con toda la expresividad de SQL. Esto es útil cuando se conoce la lógica del código pero no el texto exacto, por ejemplo para recuperar todos los subcódigos de una categoría.
+
+El siguiente ejemplo extrae los primeros cinco códigos de diabetes tipo 2 (categoría E11):
 
 ```{r}
-# Todos los codigos diabetes tipo 2
 cie10_sql("SELECT codigo, descripcion FROM cie10 WHERE codigo LIKE 'E11%' LIMIT 5")
 ```
 
-## Busqueda por codigo
+La función solo acepta consultas `SELECT` para proteger la integridad del catálogo.
+
+## Búsqueda por código conocido
+
+Cuando el código ya está disponible en los datos —por ejemplo, al preparar un informe con diagnósticos de egreso hospitalario— `cie_lookup()` recupera la descripción oficial a partir de uno o varios códigos.
 
 ```{r}
-# Busqueda de un solo codigo
+# Un solo código
 cie_lookup("E11.0")
-
-# Busqueda vectorizada - multiples codigos a la vez
-codigos <- c("E11.0", "I10", "Z00", "J44.0")
[…]
