+++ vignettes/caso-uso-egresos.Rmd
@@ -1,10 +1,10 @@
 ---
-title: "Caso de uso: Egresos hospitalarios con ciecl"
+title: "Caso de Uso: Análisis de Egresos Hospitalarios (DEIS Chile)"
 author: "Rodolfo Tasso Suazo"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Caso de uso: Egresos hospitalarios con ciecl}
+  %\VignetteIndexEntry{Caso de Uso: Análisis de Egresos Hospitalarios (DEIS Chile)}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -18,198 +18,128 @@ library(ciecl)
 library(dplyr)
 ```
 
-Los registros de egresos hospitalarios en Chile (fuente DEIS) almacenan
-diagnosticos en formato CIE-10 sin punto (`J189`, `O800`) y en algunos casos
-con sufijo `X` (`I10X`, `N40X`) para indicar categorias sin subcategoria
-adicional. Esta convencion de codificacion dificulta unir los datos con
-catalogos externos que usan el formato normalizado con punto (`J18.9`, `O80.0`).
+## Introducción: El desafío del investigador
 
-Esta vignette reproduce el flujo utilizado en investigacion con datos DEIS:
-obtener descripciones CIE-10 para toda la base, unirlas con los registros
-originales, y generar perfiles de diagnostico.
+Imagina que eres un investigador de salud pública en Chile y acabas de descargar las bases de datos de **Egresos Hospitalarios** del Departamento de Estadísticas e Información de Salud (DEIS). Estas bases son fundamentales para entender la carga de enfermedad en el país, pero presentan un desafío técnico común: la codificación de diagnósticos.
 
-## Datos de ejemplo
+En los registros del DEIS, los códigos CIE-10 suelen venir "pegados" (sin el punto decimal, ej: `J189` en lugar de `J18.9`) y a menudo incluyen una `X` al final para indicar categorías sin subcategoría específica (ej: `I10X` para Hipertensión esencial). Esta convención dificulta enormemente el cruce con catálogos internacionales o la creación de reportes legibles para audiencias clínicas.
 
-Simulamos una base con la estructura tipica de los egresos hospitalarios DEIS.
-Las columnas `DIAG1` y `ANO` imitan el formato original:
+El paquete `ciecl` fue diseñado precisamente para resolver estos problemas de forma idiomática y eficiente. En este caso de uso, seguiremos el flujo de trabajo típico para transformar una base bruta en un análisis de comorbilidad listo para publicación.
+
+## 1. Preparación de los datos
+
+Primero, simulamos una muestra de datos con la estructura real que encontrarías en los archivos `.csv` del DEIS. Notarás que la columna `DIAG1` (diagnóstico principal) contiene códigos en diversos formatos informales.
 
 ```{r datos}
 set.seed(42)
 
+# Simulamos 200 egresos con formatos típicos del DEIS Chile
 egresos <- data.frame(
   ID_EGRESO = 1:200,
+  PACIENTE_ID = sample(1:50, 200, replace = TRUE), # 50 pacientes únicos
   ANO       = sample(2018:2022, 200, replace = TRUE),
   DIAG1     = sample(
     c(
-      # Codigos frecuentes en egresos DEIS (sin punto, formato original)
[…]
