### 2026-04-25 47a75e7
test: corregir tests para rev-ropensci (Bloque G parcial)
- expect_empty -> expect_equal(nrow, 0L) (funcion inexistente en testthat)
- Mensajes de error texto -> text (args migrados a EN)
- S72X01A/T84X0XA: expectativa corregida a comportamiento real del normalizador
- Z99.9 NA: reemplazado por test del fallback 3 digitos
- cie_validate_vector strict: codigos ajustados a catalogo MINSAL real
