+++ tests/testthat/test-cie-utils.R
@@ -1,449 +1,82 @@
-test_that("cie_normalizar convierte formatos correctamente", {
+# test-cie-utils.R
+# Pruebas para funciones de utilidad y normalizacion
+
+test_that("cie_norm convierte formatos correctamente", {
+  # Caso base
+  expect_equal(cie_norm("E11.0", search_db = FALSE), "E11.0")
+  
+  # Sin punto
+  expect_equal(cie_norm("E110", search_db = FALSE), "E11.0")
+  
+  # Sufijo X (Chile)
+  expect_equal(cie_norm("I10X", search_db = FALSE), "I10")
+  expect_equal(cie_norm("J00X", search_db = FALSE), "J00")
+  expect_equal(cie_norm("B54X", search_db = FALSE), "B54")
+  expect_equal(cie_norm("i10x", search_db = FALSE), "I10")
+  
+  # Codigos con 7+ chars (sufijos de encuentro ICD-10-PCS) se preservan sin cambio
+  expect_equal(cie_norm("S72X01A", search_db = FALSE), "S72X01A")
+  expect_equal(cie_norm("T84X0XA", search_db = FALSE), "T84X0XA")
+})
+
+test_that("cie_norm es vectorizado", {
+  codigos <- c("E110", "I10X", "Z00")
+  expect_equal(cie_norm(codigos, search_db = FALSE), c("E11.0", "I10", "Z00"))
+  
+  codigos_x <- c("A15X", "B20X", "C50X")
+  expect_equal(cie_norm(codigos_x, search_db = FALSE), c("A15", "B20", "C50"))
+})
+
+test_that("cie_norm maneja NAs y vacios", {
+  expect_true(is.na(cie_norm(NA_character_, search_db = FALSE)))
+  expect_equal(cie_norm("", search_db = FALSE), "")
+})
+
+test_that("cie_norm con search_db=TRUE valida contra catalogo", {
   skip_on_cran()
   
-  # Con punto (ya normalizado)
-  expect_equal(cie_normalize("E11.0", search_db = FALSE), "E11.0")
+  # Codigo que existe despues de normalizar
+  expect_equal(cie_norm("E110", search_db = TRUE), "E11.0")
   
-  # Sin punto (agrega punto)
-  expect_equal(cie_normalize("E110", search_db = FALSE), "E11.0")
-  
-  # Con sufijo X en codigos cortos (elimina X)
-  expect_equal(cie_normalize("I10X", search_db = FALSE), "I10")
-  expect_equal(cie_normalize("J00X", search_db = FALSE), "J00")
-  expect_equal(cie_normalize("B54X", search_db = FALSE), "B54")
-  expect_equal(cie_normalize("i10x", search_db = FALSE), "I10")  # minusculas
-  
-  # Codigos largos con X placeholder (preserva X)
-  # En trauma/lesiones, X es placeholder obligatorio del 7o caracter
-  expect_equal(cie_normalize("S72X01A", search_db = FALSE), "S72X01A")
-  expect_equal(cie_normalize("T84X0XA", search_db = FALSE), "T84X0XA")
-  
-  # Vectorizado
-  codigos <- c("E11.0", "I10.0", "Z00")
-  resultado <- cie_normalize(codigos, search_db = FALSE)
-  expect_equal(resultado, c("E11.0", "I10.0", "Z00"))
-  
-  # Vectorizado con X (cortos y largos)
-  codigos_x <- c("I10X", "J00X", "E110", "S72X01A")
-  resultado_x <- cie_normalize(codigos_x, search_db = FALSE)
-  expect_equal(resultado_x, c("I10", "J00", "E11.0", "S72X01A"))
[…]
