### 2026-04-25 b04ed2e
Revision rOpenSci: internacionalizacion completa de argumentos y unificacion de normalizacion.
- Argumentos en ingles con deprecacion lifecycle para compatibilidad.
- cie_norm() establecida como funcion canonica de limpieza.
- Nuevas vignettes con narrativa mejorada y ejemplos ejecutables.
- Limpieza masiva de tests para eliminar warnings de deprecacion.
- Regeneracion de README.md (ES) y README.en.md (EN).
