### 2026-04-26 a6c8c8d
Refactor: narrativa profesional en caso de uso y reconstruccion de pkgdown.
- Eliminado tono dramatico e imaginativo en vignettes.
- Documentacion alineada con estandares tecnicos de rOpenSci.
