+++ vignettes/caso-uso-egresos.Rmd
@@ -18,25 +18,28 @@ library(ciecl)
 library(dplyr)
 ```
 
-## Introducción: El desafío del investigador
+## Introducción
 
-Imagina que eres un investigador de salud pública en Chile y acabas de descargar las bases de datos de **Egresos Hospitalarios** del Departamento de Estadísticas e Información de Salud (DEIS). Estas bases son fundamentales para entender la carga de enfermedad en el país, pero presentan un desafío técnico común: la codificación de diagnósticos.
+Este caso de uso describe el flujo de procesamiento técnico necesario para transformar bases de datos administrativas de salud en Chile hacia un formato analítico estandarizado. Se utilizan como referencia las bases de **Egresos Hospitalarios** publicadas por el Departamento de Estadísticas e Información de Salud (DEIS).
 
-En los registros del DEIS, los códigos CIE-10 suelen venir "pegados" (sin el punto decimal, ej: `J189` en lugar de `J18.9`) y a menudo incluyen una `X` al final para indicar categorías sin subcategoría específica (ej: `I10X` para Hipertensión esencial). Esta convención dificulta enormemente el cruce con catálogos internacionales o la creación de reportes legibles para audiencias clínicas.
+Los registros administrativos de salud presentan frecuentemente inconsistencias estructurales en la codificación CIE-10. Las dos variaciones más comunes en el contexto chileno son:
 
-El paquete `ciecl` fue diseñado precisamente para resolver estos problemas de forma idiomática y eficiente. En este caso de uso, seguiremos el flujo de trabajo típico para transformar una base bruta en un análisis de comorbilidad listo para publicación.
+1.  **Formatos compactos**: Códigos sin punto decimal (ej: `J189` en lugar de `J18.9`).
+2.  **Sufijos de relleno**: Uso de la letra `X` para completar la longitud del campo en categorías de 3 dígitos (ej: `I10X` para Hipertensión esencial).
 
-## 1. Preparación de los datos
+Estas variaciones dificultan la interoperabilidad y el cruce con estándares internacionales. El paquete `ciecl` automatiza la corrección de estas inconsistencias de forma vectorizada y eficiente.
 
-Primero, simulamos una muestra de datos con la estructura real que encontrarías en los archivos `.csv` del DEIS. Notarás que la columna `DIAG1` (diagnóstico principal) contiene códigos en diversos formatos informales.
+## 1. Estructura de los datos originales
+
+A continuación, se genera un conjunto de datos sintético que replica la estructura y las anomalías típicas encontradas en los archivos `.csv` del DEIS. La columna `DIAG1` representa el diagnóstico principal con los formatos informales mencionados.
 
 ```{r datos}
 set.seed(42)
 
-# Simulamos 200 egresos con formatos típicos del DEIS Chile
+# Simulación de 200 registros con formatos típicos del DEIS Chile
 egresos <- data.frame(
   ID_EGRESO = 1:200,
-  PACIENTE_ID = sample(1:50, 200, replace = TRUE), # 50 pacientes únicos
+  PACIENTE_ID = sample(1:50, 200, replace = TRUE),
   ANO       = sample(2018:2022, 200, replace = TRUE),
   DIAG1     = sample(
     c(
@@ -54,34 +57,34 @@ egresos <- data.frame(
 head(egresos)
 ```
 
-## 2. Limpieza inteligente con `cie_norm()`
+## 2. Normalización de códigos con `cie_norm()`
 
[…]
