### 2026-04-26 16db274
Compliance rOpenSci: mejoras tecnicas en cliente API, documentacion y tests.
- Implementados User-Agent, retry y throttling en cie11_search() siguiendo estandares rOpenSci.
- Mapeo total de deprecaciones a lifecycle::deprecate_warn().
- Suite de tests actualizada para capturar advertencias de lifecycle.
- devtools::check() limpio (0 errores, 0 warnings).
