+++ tests/testthat/test-robustness.R
@@ -507,13 +507,13 @@ test_that("cie10_sql permite punto y coma dentro de strings", {
   expect_s3_class(resultado, "tbl_df")
 })
 
-test_that("cie10_sql con close=FALSE mantiene conexion", {
+test_that("cie10_sql con close=FALSE emite advertencia", {
   skip_on_cran()
 
-  # Este test verifica comportamiento interno
-  # La conexion no debe cerrarse si close=FALSE
-  resultado <- cie10_sql("SELECT COUNT(*) as n FROM cie10", close = TRUE)
-  expect_gt(resultado$n, 0)
+  expect_warning(
+    cie10_sql("SELECT COUNT(*) as n FROM cie10", close = TRUE),
+    class = "lifecycle_warning_deprecated"
+  )
 })
 
 test_that("cie10_sql normaliza espacios al inicio", {
