+++ man/cie_normalizar.Rd
@@ -21,6 +21,12 @@ Character vector con codigos normalizados
 Alias en espanol de \code{\link[=cie_norm]{cie_norm()}}. Se mantiene por compatibilidad
 con codigo existente en CRAN. Usar \code{\link[=cie_norm]{cie_norm()}} en codigo nuevo.
 }
+\examples{
+\dontrun{
+# Deprecated: usar cie_norm()
+cie_normalizar("E110")
+}
+}
 \seealso{
 Other validacion: 
 \code{\link{cie_expand}()},
