### 2026-04-26 be61df7
Homologacion total: documentacion bilingue espejo y configuracion bilingue de pkgdown.
- Creadas vignettes en ingles: case-study-discharges e installation.
- Refactorizado NEWS.md con fecha 2026-04-25 y English Summary.
- Traducida navbar de pkgdown y organizado menu de Guias por idioma.
- Validacion final devtools::check() exitosa (0 errores, 0 warnings).
