+++ NEWS.md
@@ -1,71 +1,78 @@
-# ciecl 0.9.8 (en desarrollo — rev-ropensci)
+# ciecl 0.9.8 (2026-04-25)
 
-Respuesta a la revision editorial de rOpenSci. Unificacion de idioma
-(funciones y argumentos en ingles, documentacion en espanol) sin romper
-compatibilidad con CRAN.
+*English summary below*
 
-## Nueva funcion
+## Respuesta a Revisión rOpenSci (rev-ropensci)
 
-* **`cie_describe(codes)`**: devuelve un vector `character` con la
-  descripcion de cada codigo CIE-10, pensado para `mutate()` sin
-  `left_join`. Acepta el argumento `default` para codigos no encontrados.
+Versión de cumplimiento editorial y técnico tras la revisión de Maëlle Salmon.
+Esta versión unifica la interfaz del paquete bajo estándares internacionales de R,
+manteniendo la documentación pedagógica en español para el contexto local.
 
-## Renombres con deprecacion (compatibilidad CRAN)
+### Cambios que rompen compatibilidad (Deprecaciones)
 
-Todos los argumentos y funciones en espanol siguen funcionando; emiten
-`lifecycle::deprecate_warn("0.10.0", ...)` al invocarse.
+* **Internacionalización de la API**: Todos los argumentos de las funciones
+  públicas han sido migrados al inglés para consistencia con el ecosistema
+  tidyverse (dplyr, httr2). Los argumentos antiguos en español emiten
+  `lifecycle::deprecate_warn()` y serán eliminados en la versión 1.0.0.
+  
+  - `cie_lookup()`: `codigo` -> `code`, `expandir` -> `expand`, `normalizar` -> `normalize`.
+  - `cie_search()`: `texto` -> `text`, `campo` -> `field`, `solo_fuzzy` -> `only_fuzzy`.
+  - `cie_short()`: Reemplaza a `cie_siglas()`. Argumento `category`.
+  - `cie11_search()`: `texto` -> `text`.
 
-* Funcion: `cie_normalizar()` -> `cie_norm()`
-* Funcion: `cie_normalize()` -> `cie_norm()`
-* Argumentos:
-  - `cie_lookup(codigo=, expandir=, normalizar=, descripcion_completa=)` ->
-    `cie_lookup(code=, expand=, normalize=, full_description=)`
-  - `cie_norm(codigos=, buscar_db=)` -> `cie_norm(codes=, search_db=)`
-  - `cie_validate_vector(codigos=)` -> `cie_validate_vector(codes=)`
-  - `cie_expand(codigo=)` -> `cie_expand(code=)`
+* **Motor de Normalización Único**: Se establece `cie_norm()` como la función
+  canónica. `cie_normalize()` y `cie_normalizar()` están ahora deprecadas.
 
-## Documentacion
+### Nuevas funcionalidades
 
-* README unificado paramétrico: un solo `README.Rmd` con `params$lang`
-  genera `README.md` (espanol canonico) y `README.en.md` (ingles
-  internacional). Ya no hay `README.es.Rmd` separado. Build via
-  `tools/build_readmes.R`.
-* `README.md` ahora en espanol (el editor rOpenSci lo pidio por el
-  target hispanohablante del paquete).
-* Vignette `idiomas.Rmd`: reescrita con tildes correctas ("tildes
-  preservadas", "espanol"), narrativa entre chunks, seccion Brasil/
-  Frances eliminada (fuera del scope `ciecl` = CIE Chile).
-* Vignette `instalacion.Rmd`: seccion `keyring` estilo babeldown para
[…]
