+++ vignettes/case-study-discharges.Rmd
@@ -1,0 +1,142 @@
+---
+title: "Case Study: Hospital Discharge Analysis (DEIS Chile)"
+author: "Rodolfo Tasso Suazo"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Case Study: Hospital Discharge Analysis (DEIS Chile)}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+library(ciecl)
+library(dplyr)
+```
+
+## Introduction
+
+This case study describes the technical workflow required to transform administrative health databases in Chile into a standardized analytical format. We use the **Hospital Discharge** databases published by the Department of Health Statistics and Information (DEIS) as a reference.
+
+Health administrative records frequently present structural inconsistencies in ICD-10 coding. The two most common variations in the Chilean context are:
+
+1.  **Compact formats**: Codes without a decimal point (e.g., `J189` instead of `J18.9`).
+2.  **Filler suffixes**: Use of the letter `X` to complete the field length in 3-digit categories (e.g., `I10X` for Essential hypertension).
+
+These variations hinder interoperability and cross-referencing with international standards. The `ciecl` package automates the correction of these inconsistencies in a vectorized and efficient manner.
+
+## 1. Original Data Structure
+
+A synthetic dataset is generated below that replicates the structure and typical anomalies found in DEIS `.csv` files. The `DIAG1` column represents the primary diagnosis with the aforementioned informal coding formats.
+
+```{r datos}
+set.seed(42)
+
+# Simulation of 200 records with typical DEIS Chile formats
+discharges <- data.frame(
+  DISCHARGE_ID = 1:200,
+  PATIENT_ID   = sample(1:50, 200, replace = TRUE),
+  YEAR         = sample(2018:2022, 200, replace = TRUE),
+  DIAG1        = sample(
+    c(
+      "J189", "O800", "Z380", "K359", "N390",
+      "I10X", "J449", "E119", "O829", "J069",
+      "K922", "N185", "I509", "C509", "A099",
+      "N40X", "K800", "I259", "J180", "E149"
+    ),
+    size    = 200,
+    replace = TRUE
+  ),
+  stringsAsFactors = FALSE
+)
+
+head(discharges)
+```
+
+## 2. Code Normalization with `cie_norm()`
+
+Normalization is the critical first step to ensure analysis integrity. The `cie_norm()` function processes codes by applying official MINSAL coding rules:
+
+*   **Suffix removal**: Identifies and removes the trailing `X`.
+*   **Punctuation formatting**: Inserts the decimal point in the standard position according to the ICD-10 hierarchy.
+*   **String cleaning**: Removes whitespace and non-printable characters.
+
+```{r normalizacion}
+# Cleaning and standardization of diagnoses
+discharges <- discharges %>%
+  mutate(
+    DIAG1_NORM = cie_norm(codes = DIAG1)
+  )
+
+# Comparison between original and normalized formats
+discharges %>% 
+  select(DIAG1, DIAG1_NORM) %>% 
+  distinct() %>% 
+  head(5)
+```
+
+## 3. Semantic Enrichment with `cie_lookup()`
+
+After normalization, standardized clinical descriptions are assigned to facilitate result interpretation. `cie_lookup()` performs this task through an optimized query to the integrated catalog, returning full metadata for each code.
+
+```{r lookup}
+# Extracting descriptions for unique codes
+unique_codes <- unique(discharges$DIAG1_NORM)
+
[…]

[…]
