### 2026-04-26 27df70a
test: completar G3 y G1 — expect_no_match y withr::local_envvar
- expect_false(grepl()) -> expect_no_match() en test-api-mock (x3) y test-cie-data (x1)
- Sys.setenv manual -> withr::local_envvar en test-api-mock (x2) e integration (x1)
- .Rbuildignore: excluir vignettes/ciecl-en.html del tarball
