### 2026-04-26 #765 @Rodotasso
Hola @maelle!, ya puedo responder a todas tus sugerencias con respecto al código y he aplicado los cambios necesarios.
Nuevamente agradezco todas tus sugerencias consejos para la mejora del paquete 😸

Todas estas sugerencias y cambios estan en la rama  [`rev-ropensci`](https://github.com/RodoTasso/ciecl/tree/rev-ropensci)

**Idiomas y la nomenclatura**

-Como comentaba antes, al hacer la postulación a CRAN es necesaria la documentación en Inglés, pero cuando comencé la estructura de pkgdown para RopenSci, quedaron dos archivos .rmd que separaban las traducciones. Ahora: El README principal está en español. El sitio `pkgdown` incluye una versión espejo en inglés para usuarios internacionales, con el merge a main me aseguraré de que también pase con el repo.

-Debido a que en algunos momentos temporales de la creación predominaba la reproducibilidad, o en otro predominaba la adaptación al ecosistema, provoqué inconsistencias en las funciones y argumentos como notaste, es por eso que para facilitar su adopción en el ecosistema los argumentos y funciones ahora usan nombres en inglés (`text`, `code`, `expand`, `field`). Los alias en español emiten advertencias de deprecación vía `lifecycle::deprecate_warn()`, para no romper código existente. `cie_normalize()` se consolidó en `cie_norm()`. Los comentarios del código, variables siguen en español para que gente en este idioma pueda replicar el paquete.

-Sobre el idioma BRA y FR: El paquete se circunscribe a la CIE-10 Chile (MINSAL/DEIS), como indica su nombre. Las variantes de otros países (Brasil, Francia u otras) quedan fuera del alcance declarado y no forman parte del roadmap. La sección correspondiente fue eliminada de la vignette.

-Tildes: Corregido `espanol → español` y `tildes preservadas → tildes preservadas`.

**Documentación**

-Eliminadas las duplicaciones en el README.
-La vignette de introducción (`ciecl-es.Rmd`) fue reescrita con texto narrativo entre bloques de código, en lugar de bloques consecutivos sin contexto. Lo mismo con 'Caso de uso', ahora se entiende el flujo completo usando las comorbilidades.

-Todos los ejemplos corren con `eval = TRUE`. El bloque de `cie_comorbid()` requiere `comorbidity` y está marcado con una nota.

-Las columnas `incluye` y `excluye` corresponden a las notas clínicas del estándar CIE-10 internacional: sinónimos y equivalentes diagnósticos para la primera, y diagnósticos similares que deben codificarse en otro lugar para la segunda. Están en `NA` porque el catálogo MINSAL/DEIS v2018 no publica esos campos en su versión tabular o no existen para todas las patologías; es una limitación de la fuente oficial, no un bug del paquete. `cie_table()` ahora los oculta automáticamente cuando están vacíos.

**Nuevas funcionalidades y código**
Agregué algunas cosas a partir de tus sugerencias y funciones que antes eran parte de un flujo se separaron.

-`cie_describe()`: Nueva función para enriquecer vectores de códigos dentro de `dplyr::mutate()`, sin joins manuales.

-`rlang::check_installed()`: Reemplaza los mensajes de error escritos a mano para dependencias opcionales (`gt`, `readxl`, `comorbidity`).

-Cache del usuario: `use_data()` fue eliminado del código de ejecución. El cache ahora se construye con `save()` de R base.

-Archivos con espacios en el nombre: Los archivos fuente oficiales MINSAL no fueron diseñados para uso programático, sus nombres contienen espacios por convención de distribución. El parser ahora los maneja independientemente del nombre del archivo.

-`stringr::str_subset()`: Los patrones `names(raw)[str_detect(...)]` fueron reemplazados por `str_subset(names(raw), ...)`. Es equivalente a `grep(value = TRUE)`: filtra directamente en lugar de detectar y luego indexar, lo que hace el código más legible. Se descartó `str_match()` porque retorna una matriz de capturas, no un vector filtrado.

**Instalación y API externa**

- Como sugeriste, `pak` es ahora el método de instalación recomendado en la documentación.

- Para el caso de `SystemRequirements`, fue eliminado de `DESCRIPTION`. Estaban presentes por la misma razón que `gc()`: Son el remanente de un flujo de trabajo con bases de ~20M de filas que ralentizaba mucho los tiempos de trabajo de acuerdo al hardware o software, eran medidas de protección para reproducibilidad en otros equipos. El paquete publicado no los necesita.

-Se añadió una sección sobre gestión de credenciales de la API CIE-11 con `keyring`, siguiendo el modelo de babeldown.
-`cie11_search()` fue refactorizado con `httr2`: incluye `req_user_agent()`, `req_retry(max_tries = 3)` y `req_throttle(rate = 10/60)`.

**Tests**

-Variables de entorno gestionadas con `withr::local_envvar()`.

-`expect_false(any(grepl(...)))` reemplazado por `expect_no_match()`.

-`skip_if_not_installed()` aplicado de forma consistente y no redundante.

-Llamadas a `gc()` eliminadas. Era una costumbre por lo mencionado anteriormente, eliminar los datos cargados.

-`skip_if_offline()` en tests de API: Los mantuve en los tests de `cie11_search()` precisamente porque cubre ambos escenarios, sin conexión y en CRAN, sin necesitar un `skip_on_cran()` adicional, el `skip_if_offline()` los salta por tener el 'skip_on_cran()' internamente, además está agregado un test de api-mock para simular el contexto de conexión, las pruebas se realizaron cubriendo el escenario de los skip, API y api-mock.

-`expect_shape()` vs. `expect_equal(nrow(...))`: En este caso revisé la documentación para entender la sugerencia y preferí mantener `expect_equal(nrow(resultado), 0)` porque deja explícitamente que solo se verifica el número de filas, sin presuponer ni verificar columnas. .  En funciones como `cie11_search()` el resultado puede tener columnas cambiantes según el diagnóstico o la fuente (CIE-10 vs CIE-11), por lo que un test agnóstico a las columnas es más robusto, esa es hasta el momento mi decisión, pero con gusto puedo recibir más feedback.

Espero todos estos cambios logren cumplir con todas las sugerencias que propusiste, y si hay más propuestas estoy feliz de continuar con el proceso 😃.

Saludos!
