### 2026-04-26 #765 @Rodotasso
Pensé que lo había cubierto todo, pero me faltó algo 🥲 , en el sitio pkgdown y el caso de uso, `cie_lookuo()` está usando la ahora llamada `cie_norm()` de forma interna (normalizar = TRUE), el argumento existe para simplificar el flujo cuando el código puede venir en formato irregular (sin punto, con sufijo X, capitalización variable). `cie_norm()` está por separado para formatear según CIE-10 sin realizar una búsqueda, por ejemplo para preparar columnas antes de un join. Actualicé el texto narrativo de la vignette para dejarlo explícito y que sea limpia su forma de uso.
