### 2026-04-26 310f5c7
docs: integrate cie_describe and clean up English documentation
- Updated README.Rmd and vignettes with cie_describe() as preferred method.
- Standardized English vignettes to use English arguments (codes, search_db).
- Regenerated README.md and manual files.
- Verified with R CMD check (0 errors, 0 warnings).
