+++ README.md
@@ -97,50 +97,27 @@ cie_lookup(c("E11.0", "I10", "Z00"))
 #> 3 Z00    Examen general e… Z00 EXAM… Z00-Z1… Cap.21  FACTOR… <NA>      <NA>     
 #> # ℹ 3 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>
 
-# Descripcion directa para usar en mutate()
+# Descripcion directa (vectorizada) para usar en mutate()
 cie_describe(c("E11.0", "I10"))
 #> [1] "Diabetes mellitus tipo 2 con coma" "Hipertensión esencial (primaria)"
 
+# Ejemplo en flujo dplyr
+# egresos %>% mutate(desc = cie_describe(codigo))
+```
+
 # Busqueda fuzzy tolerante a errores
-cie_search("diabetis mellitus")
-#> # A tibble: 50 × 4
-#>    codigo descripcion                                            score categoria
-#>    <chr>  <chr>                                                  <dbl> <chr>    
-#>  1 E10    Diabetes mellitus insulinodependiente                    0.5 E10 DIAB…
-#>  2 E10.0  Diabetes mellitus tipo 1 con coma                        0.5 E10 DIAB…
-#>  3 E10.1  Diabetes mellitus tipo 1 con cetoacidosis                0.5 E10 DIAB…
-#>  4 E10.2  Diabetes mellitus tipo 1 con complicaciones renales      0.5 E10 DIAB…
-#>  5 E10.3  Diabetes mellitus tipo 1 con complicaciones oftálmicas   0.5 E10 DIAB…
-#>  6 E10.4  Diabetes mellitus tipo 1 con complicaciones neurológi…   0.5 E10 DIAB…
-#>  7 E10.5  Diabetes mellitus tipo 1 con complicaciones  circulat…   0.5 E10 DIAB…
-#>  8 E10.6  Diabetes mellitus tipo 1 con otras complicaciones esp…   0.5 E10 DIAB…
-#>  9 E10.7  Diabetes mellitus tipo 1 con complicaciones múltiples    0.5 E10 DIAB…
-#> 10 E10.8  Diabetes mellitus tipo 1 con complicaciones no especi…   0.5 E10 DIAB…
-#> # ℹ 40 more rows
+
+cie_search(“diabetis mellitus”)
 
 # Siglas medicas chilenas
-cie_search("IAM")
-#> i Sigla detectada: IAM -> infarto agudo miocardio
-#> # A tibble: 50 × 4
-#>    codigo descripcion                                            score categoria
-#>    <chr>  <chr>                                                  <dbl> <chr>    
-#>  1 I21    Infarto agudo del miocardio                                1 I21 INFA…
-#>  2 I21.0  Infarto transmural agudo del miocardio de la pared an…     1 I21 INFA…
-#>  3 I21.1  Infarto transmural agudo del miocardio de la pared in…     1 I21 INFA…
-#>  4 I21.2  Infarto agudo transmural del miocardio de otros sitios     1 I21 INFA…
-#>  5 I21.3  Infarto transmural agudo del miocardio, de sitio no e…     1 I21 INFA…
-#>  6 I21.4  Infarto subendocárdico agudo del miocardio                 1 I21 INFA…
-#>  7 I21.9  Infarto agudo del miocardio, sin otra especificación       1 I21 INFA…
-#>  8 I23    Ciertas complicaciones presentes posteriores al infar…     1 I23 CIER…
-#>  9 I23.0  Hemopericardio como complicación presente posterior a…     1 I23 CIER…
-#> 10 I23.3  Ruptura de la pared cardíaca sin hemopericardio como …     1 I23 CIER…
[…]
