### 2026-04-27 aba8c80
fijo: revertir cambios de seguridad nativa SQL para restaurar integridad de tests
- Se revirtió el uso de SQLITE_RO en get_cie10_db() por incompatibilidad con el sistema de auto-sanación.
- Se restauraron los tests originales en test-cie-sql.R.
- Se limpió NEWS.md y README.md de referencias a la característica revertida.
- El paquete vuelve a estado estable (0 errores, 0 warnings).
