### 2026-04-27 #765 @maelle
👋 @Rodotasso! Muchas gracias para tu repuesta y para los cambios. No traduciré mis comentarios, dado que has respuesto con bastante contexto para que cualquier persona pueda entender lo que se preguntó.

No veo los cambios en el repo, cual rama tendria que mirar?
