### 2026-04-27 #765 @Rodotasso
Hola @maelle! Me disculpo porque no sea visible el link en la respuesta. Lo dejo acá para que se vea claro, de todas maneras, puedo hacer el merge a main si es necesario o si lo prefieres. 🫰

https://github.com/RodoTasso/ciecl/tree/rev-ropensci
