### 2026-04-27 #765 @maelle
Creo que para mi y para l@s revisores será más facil si los cambios estan en main 😸

Porqué tuviste que crear documentación en inglés para CRAN?
