+++ llms.txt
@@ -1,305 +1,183 @@
 # ciecl
 
-> **Language / Idioma:** **English** \|
-> [Español](https://rodotasso.github.io/ciecl/articles/ciecl-es.html)
+> **Idioma / Language:** **Español** \|
+> [English](https://github.com/RodoTasso/ciecl/blob/main/README.en.md)
 
-**Data Science for Public Health Group** \| University of Chile
+**Grupo de Ciencia de Datos para Salud Pública** \| Universidad de Chile
 
-**Official Chilean ICD-10 Classification (CIE-10) for R**.
+**Clasificación Internacional de Enfermedades (CIE-10) oficial de Chile
+para R**.
 
-Specialized package for searching, validating, and analyzing ICD-10
-codes in the Chilean context. Includes **39,877 codes** (categories and
-subcategories) from the official MINSAL/DEIS v2018 catalog, with
-optimized search, comorbidity calculation, and WHO ICD-11 API access.
+Paquete especializado para búsqueda, validación y análisis de códigos
+CIE-10 en el contexto chileno. Incluye el catálogo oficial MINSAL/DEIS
+v2018 con búsqueda optimizada, cálculo de comorbilidades y acceso a la
+API CIE-11 de la OMS.
 
-## Purpose
+## Finalidad
 
-`ciecl` facilitates working with ICD-10 codes in health research and
-data analysis in Chile, eliminating the need to manually manipulate
-Excel files and providing specialized tools for:
+`ciecl` facilita el trabajo con códigos CIE-10 en investigación y
+análisis de datos de salud en Chile, eliminando la necesidad de
+manipular archivos Excel manualmente y proporcionando herramientas
+especializadas para:
 
-- Fast validation of diagnostic codes
-- Error-tolerant search (fuzzy search)
-- Automatic calculation of comorbidity indices (Charlson, Elixhauser)
-- Optimized SQL queries over more than 39,000 codes
-- Hierarchical expansion of categories to subcategories (e.g., E11 -\>
-  E11.0, E11.1, …, E11.9)
+- Validación rápida de códigos diagnósticos
+- Búsqueda tolerante a errores tipográficos (fuzzy search Jaro-Winkler)
+- Cálculo automático de índices de comorbilidad (Charlson, Elixhauser)
+- Consultas SQL optimizadas sobre el catálogo completo
+- Expansión jerárquica de categorías (ej: `E11` → `E11.0`, `E11.1`, …,
+  `E11.9`)
+- Acceso a la API CIE-11 de la OMS
 
-## Why use ciecl
+## Características
 
-### Advantages:
+- **Catálogo oficial CIE-10 Chile** (MINSAL/DEIS v2018) embebido como
+  dataset
+- **Búsqueda fuzzy Jaro-Winkler** tolera errores tipográficos
+- **Siglas médicas chilenas** (IAM, EPOC, DM2, HTA, TBC, …)
+- **Consultas SQL directas** con SQLite + FTS5
+- **Comorbilidades Charlson/Elixhauser** con `comorbidity`
+- **API CIE-11 OMS** vía
+  [`cie11_search()`](https://rodotasso.github.io/ciecl/reference/cie11_search.md)
+- **Normalización robusta**: acepta `E110`, `E11.0`, `e 11 0`, `I10-0`,
+  etc.
+- **Dependencias mínimas** (8 paquetes core; el resto en Suggests)
 
-1.  **Performance**: SQLite indexed database with 10-100x faster
-    searches than Excel
-2.  **Native integration**: Works directly in R without heavy external
-    dependencies
[…]
