+++ articles/idiomas.md
@@ -1,163 +1,259 @@
-# Soporte de Idiomas e Internacionalizacion
+# Soporte de idiomas e internacionalización
 
-## Idiomas Soportados
+## Idiomas soportados
 
-El paquete `ciecl` soporta busquedas en multiples idiomas:
+El paquete `ciecl` trabaja principalmente en español de Chile, pero
+ofrece búsqueda multilingüe cuando la fuente de datos lo permite. La
+tabla resume el comportamiento por función.
 
-| Funcion                                                                         | Idioma Dataset  | Idioma Busqueda | Notas                                 |
-|---------------------------------------------------------------------------------|-----------------|-----------------|---------------------------------------|
-| [`cie_lookup()`](https://rodotasso.github.io/ciecl/reference/cie_lookup.md)     | Espanol (Chile) | \-              | Busqueda por codigo, idioma no aplica |
-| [`cie_search()`](https://rodotasso.github.io/ciecl/reference/cie_search.md)     | Espanol (Chile) | Espanol         | Descripciones en espanol chileno      |
-| [`cie11_search()`](https://rodotasso.github.io/ciecl/reference/cie11_search.md) | Espanol/Ingles  | Espanol/Ingles  | Configurable via parametro `lang`     |
-| [`cie10_sql()`](https://rodotasso.github.io/ciecl/reference/cie10_sql.md)       | Espanol (Chile) | SQL             | Descripciones en espanol              |
+| Función                                                                         | Idioma dataset   | Idioma búsqueda  | Notas                                 |
+|---------------------------------------------------------------------------------|------------------|------------------|---------------------------------------|
+| [`cie_lookup()`](https://rodotasso.github.io/ciecl/reference/cie_lookup.md)     | Español (Chile)  | —                | Búsqueda por código; idioma no aplica |
+| [`cie_search()`](https://rodotasso.github.io/ciecl/reference/cie_search.md)     | Español (Chile)  | Español          | Descripciones en español chileno      |
+| [`cie11_search()`](https://rodotasso.github.io/ciecl/reference/cie11_search.md) | Español / Inglés | Español / Inglés | Configurable vía parámetro `lang`     |
+| [`cie10_sql()`](https://rodotasso.github.io/ciecl/reference/cie10_sql.md)       | Español (Chile)  | SQL              | Columna `descripcion` en español      |
 
-## Dataset CIE-10 Chile (Espanol)
+## Dataset CIE-10 Chile
 
-El dataset `cie10_cl` contiene **39,877 codigos** con descripciones en
-espanol de Chile, basado en el catalogo oficial MINSAL/DEIS v2018.
+El dataset `cie10_cl` contiene los códigos vigentes con descripciones en
+español chileno según el catálogo oficial MINSAL/DEIS v2018. Ver
+[`?cie10_cl`](https://rodotasso.github.io/ciecl/reference/cie10_cl.md)
+para el detalle de columnas.
 
 ``` r
 library(ciecl)
 
-# Ver estructura del dataset
-head(cie10_cl)
-#> # A tibble: 6 x 8
-#>   codigo descripcion                    categoria inclusion exclusion capitulo
-#>   <chr>  <chr>                          <chr>     <chr>     <chr>     <chr>
-#> 1 A00    Colera                         A00       ...       ...       01
[…]

[…]
