### 2026-04-27 ec62219
ci: update coverage badge [skip ci]
