### 2026-04-27 #765 @Rodotasso
Perfecto!, ya está en main, esperaba que los gh actions dieran verde.

Cree la documentación en inglés porque cuando hice submit de la versión 0.9.0 a CRAN, aparte de algunos errores para corregir en los test, me pidieron si podía dejar alguna descripción en inglés, por lo que escogí dejar algunas partes de la documentación. 🐈
