### 2026-04-28 #765 @maelle
Gracias @Rodotasso!

Miré los cambios :rocket: Tengo un par de pequeños comentarios antes de empezar a buscar revisores:

- Puedes usar `expect_shape()` sin dar un `ncol`:

``` r
dim(mtcars)
#> [1] 32 11
testthat::expect_shape(mtcars, nrow = 32)
testthat::expect_shape(mtcars, nrow = 11)
#> Error:
#> ! Expected `mtcars` to have 11 rows.
#> Actual rows: 32.
```

- En vez de `\{code{\link{` para añadir un enlace hasta otra funciones, puedes usar https://roxygen2.r-lib.org/articles/rd-formatting.html#function-links o usar el tag `@family` https://roxygen2.r-lib.org/articles/index-crossref.html?q=family#family

- En vez de `\donttest` creo que tendria sentido usar `@examplesIf`, tal vez `@examplesIf interactive()`: https://roxygen2.r-lib.org/articles/rd.html?q=examplesif#examples Mismo comentario para `\dontrun`, pero creo que lo usas para funciones cuyo nombre cambió: creo que no hace falta tener un ejemplo para la versión vieja de una función.

- En `cie_describe()` usas `isTRUE(normalize)`: en vez de esto podrias comprobar más temprano si el usuario o la usuaria pasó un logical a la función como se debe.

- En `cie_describe()` en vez de https://github.com/Rodotasso/ciecl/blob/ec622191ff470480e23ba527aa415b39047cc844/R/cie-describe.R#L54-L58 porqué no escribes

```
  if (normalize) {
    codes <- cie_norm(codes, search_db = FALSE)
  }
```

Y después usas `codes` en vez de `codes_to_search`? O hay una diferencia de length?

- En archivos Rmd te aconsejeria empezar una nueva linea después de cada frase, asi las diffs de Git son más informativas. (pero es una preferencia personal, no lo tienes que hacer!) Cf https://panache.bz/ o https://docs.ropensci.org/aeolus/

- Sigo pensando que los datos no se deberian guardar en `data/`, sino en un user directory. Eso significa crear el path con `tools::R_user_dir("ciecl", "data")`.

- En este test https://github.com/Rodotasso/ciecl/blob/27df70aa583c64500ab88d8e2318e51109f2e127/tests/testthat/test-cie-data.R#L128 en  vez de usar `tryCatch()` porqué no usas `expect_error(`?

```r
# no hace falta el tempdir?
expect_error(generar_cie10_cl("no-existe.xlsx"), "usethis")
```

- En los tests no se debe poner `ciecl:::` antes de las funciones, testthat carga todo el codigo del paquete.
