### 2026-04-28 #765 @maelle
Acabo de ver https://github.com/dewittpe/medicalcoder/ en el programa de R/Medicine (https://rconsortium.github.io/RMedicine_website/Abstracts.html#medicalcoder-a-unified-and-longitudinally-aware-framework-fo). Lo pongo aqui por si a caso es interesante para ciecl.
