### 2026-04-29 #765 @Rodotasso
Hola @maelle! Muchas gracias por tus nuevas sugerencias, pero tengo una pregunta antes de implementar el cambio más radical.

-  ¿El patrón que sugieres es mover `cie10_cl.rda` a `inst/extdata/` y copiarlo a `tools::R_user_dir("ciecl", "data")` en el primer uso, dejando de exponerlo como dataset R documentado? O hay otro patrón en mente? Pregunto porque la implementación cambia bastante según si el .rda sigue siendo parte del paquete instalado o si necesita descargarse externamente.

- Medicalcoder lo conozco! dentro de mi investigación para hacer mi paquete lo encontré y revisé, es una de las razones porque me motivo a desarrollar el paquete; uno de los problemas con los códigos ICD, es que cambian según cada país y territorio de acuerdo a sus necesidades, y hasta el momento de la publicación de ciecl, no había ninguna herramienta en español publicada que permitiera esto, solo herramientas en inglés que se debían traducir y no contenían los códigos utilizados en este territorio, además dentro de Chile existe heterogeneidad de cómo se codifican los diagnósticos, por eso la herramienta se centra en normalizar los datos y categorizarlos, que ayuda mucho en las investigación y en la publicación de datos sanitarios. Para expandir el paquete a toda latinoamérica, el código ICD debería distinguir cuales son los que se utilizan en cada país, o estar en comunicación con OPS para marcar los códigos según la necesidad territorial.
