### 2026-04-29 #765 @maelle
Hablo de https://github.com/Rodotasso/ciecl/blob/ec622191ff470480e23ba527aa415b39047cc844/R/cie-data.R#L116, esos datos se deberian guardar en un user directory la primera vez que el usuario o la usuaria intenta usarlos -- podrias documentar los datos como una función `cie10_cl()` (en vez de datos `cie10_cl`).

```r
cie10_cl <- function() {
# si file.exists(userdir blabla)
# return the data
# si no error que indica que se debe usar la función que genera los datos (y esa función guardaria los datos en el userdir.
}
```

No sé si lo explico bien?
