### 2026-04-29 #765 @Rodotasso
Ahora entiendo, esas líneas son de la formación del `.rda` de `cie-data.R`, ese archivo en esa ubicación es un error mío, ya que ese es un archivo de desarrollo que preferí dejar en caso de que alguien quisiera armar el dataset para comprobar su consistencia o utilizar otro dataset, pero no participa de ninguna forma en la instalación o funcionalidad del paquete. 😶‍🌫️

Mi error y que está causando esta confusión, dejarlo en el directorio `/R`, ya que aparenta ser un archivo de uso en el paquete, cuando debería estar en `data-raw` o simplemente no estar, lo mismo pasa con el archivo `test-cie-data.R`, que también usa `ciecl::`. Pensaba que te referías a que el .rda ya no fuera una base estática. Eso me hizo preguntar. Creo que desplazar de sus ubicaciones a ambos archivos podría solucionar todo esto. 😸.

De todas formas, si sigo sin entender, estaré muy agradecido de que me continues avisando 😄
