### 2026-04-30 #765 @maelle
Aaaah ahora entiendo! Pensaba que era algo de "cache". Mejor mover esto a inst/extdata. https://r-pkgs.org/data.html#sec-data-extdata

Mis [otros comentarios](https://github.com/ropensci/software-review/issues/765#issuecomment-4333211278) son más claros?
