### 2026-04-30 b0152d5
fix: aplicar sugerencias revision rOpenSci (Maelle Salmon)
Cambios alineados con plan_respuesta_maelle_2026-04-28.md:

- Migrar enlaces \code{\link{}} a sintaxis markdown [fn()] y @family
- Reemplazar \donttest{} por @examplesIf interactive() (9 ocurrencias)
- Eliminar bloques \dontrun{} en funciones deprecadas (4 ocurrencias)
- cie_describe(): validacion temprana de normalize y simplificacion
  eliminando alias codes_to_search
- Migrar generador del dataset a data-raw/ siguiendo convencion usethis:
  R/cie-data.R -> data-raw/generate_cie10_cl.R, test movido a sanity-check
- Crear R/data.R con
[…]
