+++ tests/testthat/test-utils-internal.R
@@ -7,7 +7,7 @@
 
 test_that("normalizar_tildes remueve tildes correctamente", {
   # Acceder a funcion interna
-  normalizar_tildes <- ciecl:::normalizar_tildes
+  normalizar_tildes <- normalizar_tildes
 
   # Vocales minusculas con tilde
   expect_equal(normalizar_tildes("caf\u00e9"), "cafe")
@@ -26,7 +26,7 @@ test_that("normalizar_tildes remueve tildes correctamente", {
 })
 
 test_that("normalizar_tildes maneja vector vacio", {
-  normalizar_tildes <- ciecl:::normalizar_tildes
+  normalizar_tildes <- normalizar_tildes
 
   resultado <- normalizar_tildes(character(0))
   expect_length(resultado, 0)
@@ -34,7 +34,7 @@ test_that("normalizar_tildes maneja vector vacio", {
 })
 
 test_that("normalizar_tildes maneja texto sin tildes", {
-  normalizar_tildes <- ciecl:::normalizar_tildes
+  normalizar_tildes <- normalizar_tildes
 
   # Texto ya normalizado
   expect_equal(normalizar_tildes("diabetes"), "diabetes")
@@ -42,7 +42,7 @@ test_that("normalizar_tildes maneja texto sin tildes", {
 })
 
 test_that("normalizar_tildes es vectorizado", {
-  normalizar_tildes <- ciecl:::normalizar_tildes
+  normalizar_tildes <- normalizar_tildes
 
   entrada <- c("caf\u00e9", "ri\u00f1\u00f3n", "normal")
   esperado <- c("cafe", "rinon", "normal")
@@ -55,7 +55,7 @@ test_that("normalizar_tildes es vectorizado", {
 # ============================================================
 
 test_that("get_siglas_medicas retorna lista completa", {
-  get_siglas_medicas <- ciecl:::get_siglas_medicas
+  get_siglas_medicas <- get_siglas_medicas
 
   siglas <- get_siglas_medicas()
 
@@ -70,7 +70,7 @@ test_that("get_siglas_medicas retorna lista completa", {
 })
 
 test_that("get_siglas_medicas contiene siglas comunes", {
-  get_siglas_medicas <- ciecl:::get_siglas_medicas
+  get_siglas_medicas <- get_siglas_medicas
 
   siglas <- get_siglas_medicas()
 
@@ -84,7 +84,7 @@ test_that("get_siglas_medicas contiene siglas comunes", {
 })
 
 test_that("get_siglas_medicas tiene categorias validas", {
-  get_siglas_medicas <- ciecl:::get_siglas_medicas
+  get_siglas_medicas <- get_siglas_medicas
 
   siglas <- get_siglas_medicas()
   categorias <- unique(vapply(siglas, function(x) x$categoria, character(1)))
@@ -104,7 +104,7 @@ test_that("get_siglas_medicas tiene categorias validas", {
 # ============================================================
 
 test_that("expandir_sigla expande siglas conocidas", {
-  expandir_sigla <- ciecl:::expandir_sigla
+  expandir_sigla <- expandir_sigla
 
   # Siglas comunes
   expect_equal(expandir_sigla("iam"), "infarto agudo miocardio")
@@ -115,7 +115,7 @@ test_that("expandir_sigla expande siglas conocidas", {
 })
 
 test_that("expandir_sigla retorna NULL para no-siglas", {
-  expandir_sigla <- ciecl:::expandir_sigla
+  expandir_sigla <- expandir_sigla
 
   expect_null(expandir_sigla("diabetes"))
   expect_null(expandir_sigla("xyz123"))
[…]
