+++ data-raw/DATASET.R
@@ -1,0 +1,3 @@
+# code to prepare `cie10_cl` dataset goes here
+source("data-raw/generate_cie10_cl.R")
+generar_cie10_cl()  # produce data/cie10_cl.rda
