+++ reference/cie_lookup.md
@@ -74,9 +74,9 @@ tibble con codigo(s) matcheado(s)
 
 ## See also
 
-[`cie_search`](https://rodotasso.github.io/ciecl/reference/cie_search.md),
-[`cie_norm`](https://rodotasso.github.io/ciecl/reference/cie_norm.md),
-[`cie_expand`](https://rodotasso.github.io/ciecl/reference/cie_expand.md)
+[`cie_search()`](https://rodotasso.github.io/ciecl/reference/cie_search.md),
+[`cie_norm()`](https://rodotasso.github.io/ciecl/reference/cie_norm.md),
+[`cie_expand()`](https://rodotasso.github.io/ciecl/reference/cie_expand.md)
 
 Other busqueda:
 [`cie_describe()`](https://rodotasso.github.io/ciecl/reference/cie_describe.md),
@@ -91,83 +91,26 @@ Other busqueda:
 ``` r
 # Busqueda directa por codigo
 cie_lookup("E11.0")
-#> # A tibble: 1 × 10
-#>   codigo descripcion       categoria seccion capitulo_nombre inclusion exclusion
-#>   <chr>  <chr>             <chr>     <chr>   <chr>           <chr>     <chr>    
-#> 1 E11.0  Diabetes mellitu… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
-#> # ℹ 3 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>
-
-# \donttest{
-cie_lookup("E110")        # Sin punto
-#> # A tibble: 1 × 10
-#>   codigo descripcion       categoria seccion capitulo_nombre inclusion exclusion
-#>   <chr>  <chr>             <chr>     <chr>   <chr>           <chr>     <chr>    
-#> 1 E11.0  Diabetes mellitu… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
-#> # ℹ 3 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>
-cie_lookup("E11")         # Solo categoria
-#> # A tibble: 1 × 10
-#>   codigo descripcion       categoria seccion capitulo_nombre inclusion exclusion
-#>   <chr>  <chr>             <chr>     <chr>   <chr>           <chr>     <chr>    
-#> 1 E11    Diabetes mellitu… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
-#> # ℹ 3 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>
-cie_lookup("E11", expand = TRUE)  # Todos E11.x
-#> # A tibble: 11 × 10
-#>    codigo descripcion      categoria seccion capitulo_nombre inclusion exclusion
-#>    <chr>  <chr>            <chr>     <chr>   <chr>           <chr>     <chr>    
-#>  1 E11    Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
-#>  2 E11.0  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
-#>  3 E11.1  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
-#>  4 E11.2  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
-#>  5 E11.3  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
-#>  6 E11.4  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
[…]
