+++ index.md
@@ -50,6 +50,7 @@ solo se actualiza por decreto institucional.
 ## Instalación
 
 ``` r
+
 # CRAN
 install.packages("ciecl")
 
@@ -61,6 +62,7 @@ pak::pak("RodoTasso/ciecl")
 ## Uso rápido
 
 ``` r
+
 library(ciecl)
 
 # Busqueda exacta por codigo
@@ -112,6 +114,7 @@ necesitas credenciales gratuitas de la OMS
 `keyring`:
 
 ``` r
+
 keyring::key_set("ciecl_icd11")  # client_id:client_secret
 Sys.setenv(ICD_API_KEY = keyring::key_get("ciecl_icd11"))
 cie11_search("diabetes mellitus")
