### 2026-05-01 #765 @Rodotasso
Hola @maelle! ✋

Sí, todos quedaron claros, muchas gracias por el detalle de cada comentario.

Tuve un par de tropiezos con `expect_shape()`: al aplicar el cambio, las llamadas quedaban huérfanas porque la función no resolvía. Bajé el código de:
 [`r-lib/testthat`](https://github.com/r-lib/testthat) para entender por qué y descubrí que `expect_shape()` recién está en la rama de desarrollo (`3.3.2.9000`, [r-lib/testthat#1423](https://github.com/r-lib/testthat/issues/1423)) y todavía no se libera en CRAN.

 En lugar de exigir una versión de `testthat` no publicada, dejé un *helper polyfill* local en `tests/testthat/helper-expect-shape.R` que reproduce la API oficial usando solo símbolos exportados estables (`testthat::quasi_label`, `fail`, `succeed`, `rlang::check_dots_empty`, `enquo`). Cuando `expect_shape()` salga liberada en CRAN, basta con borrar el helper y subir el mínimo de `testthat` en `DESCRIPTION`, los tests no cambian. Aprendí a hacer algo nuevo con eso 🙌

Los demás cambios también están aplicados; el detalle por punto y los archivos/líneas afectados está en el cuerpo del PR de arriba (aprendí también a referenciar PRs por acá 😄).

Sobre `devtools::test()`: aproveché de limpiar 2 warnings que venían de `try(dbDisconnect(...), silent = TRUE)`. `try()` no captura *warnings*, solo errores, así que el warning "Already disconnected" de RSQLite escapaba. Cambié a `if (DBI::dbIsValid(con)) suppressWarnings(DBI::dbDisconnect(con))` y quedaron en cero.

Quedaron 5 SKIPs y de eso quería preguntarte tu opinión:

  - Cuatro son tests de fallback graceful para cuando `httr2`, `comorbidity` o `gt` (todos en `Suggests`) **no** están instalados. Usan `skip_if(requireNamespace(...))` en sentido inverso para validar el mensaje de error cuando la dependencia opcional falta. En mi entorno local todas están instaladas, así que se saltan; pero en CI con Suggests ausentes (p. ej. CRAN minimal) sí corren.
  - Uno es un test de escala opt-in vía `CIECL_LARGE_TESTS=true` (>10 s, carga los 39K códigos).

La alternativa para que esos cuatro corrieran siempre sería mockear `rlang::check_installed` con `local_mocked_bindings(.package = "rlang")`, pero eso acopla los tests a internals de `rlang` y deja de validar el comportamiento real ante dependencia ausente.

  ¿Te parecen aceptables como están, o prefieres que los reformule? Quedo atento a tu consejo.

  Estado final:

  ```
  devtools::test()  → FAIL 0 | WARN 0 | SKIP 5 | PASS 885
  devtools::check() → 0 errors | 0 warnings | 1 NOTE (timestamp benigno)
  ```
