### 2026-05-01 #765 @maelle
> ¿Te parecen aceptables como están, o prefieres que los reformule?

Gracias! Si, y verremos también lo que piensan l@s revisores de las pruebas -- mejor no solo tener mi opinión 😸
