### 2026-05-02 #765 @Rodotasso
Muchas gracias!, feliz de continuar con las fases del proceso, de hecho, he estado aprovechando las sugerencias para revisar otras partes del paquete que puedan ser actualizadas.

Pero esperaré la opinión de todos 😸

Muy agradecido por todo el feedback @maelle, ha sido muy enriquecedor. 🤲
