### 2026-05-02 4844610
refactor: completar P0 y P1 plan remediacion r-lib
P0 - metadata sincronizada
- DESCRIPTION alineado a 0.9.8
- codemeta.json: 0.9.6 -> 0.9.8
- inst/CITATION: 0.9.3 -> 0.9.8
- tests/testthat/setup.R: patrones withr::defer

P1 - migracion mensajeria base R a cli
- 7 archivos R/ migrados a cli::cli_abort/cli_warn/cli_inform
- cero llamadas residuales stop/warning/message en R/
- bugfix asignacion redundante en cie_search() expansion sigla

Saneamiento ASCII en strings de codigo
- strings con tildes/n convertidos a ASCII plano (sin tildes) en
  cie-api.R, cie-comorbid.R, cie-search.R, cie-table.R,
[…]
