### 2026-05-03 787fc20
refactor: completar plan remediacion r-lib P2-P5
- Migracion total de mensajeria base R a cli en todos los modulos:
  cie-api, cie-comorbid, cie-describe, cie-search, cie-sql, cie-table,
  cie-utils, globals. Cero llamadas residuales stop/warning/message.
  cli_abort con clases de condicion (ciecl_invalid_input, ciecl_unsafe_query,
  ciecl_api_error) para captura programatica.

- Refactor estructural P5: cie-search.R dividido en 4 modulos para mejorar
  mantenibilidad: cie-search (busqueda fuzzy), cie-siglas (diccionario y
  expansion), cie-lookup (busqueda exacta), cie-guide (guia comparativa
[…]
