+++ R/cie-search.R
@@ -1,12 +1,14 @@
-#' @importFrom stringr str_trim str_replace_all str_detect str_extract str_split
-#' @importFrom dplyr select mutate filter all_of any_of bind_rows distinct %>%
+#' @importFrom stringr str_trim str_split fixed
+#' @importFrom dplyr mutate filter arrange desc slice_head select everything
 #' @importFrom tibble as_tibble
+#' @importFrom stringdist stringsim
+#' @importFrom DBI dbGetQuery
 NULL
 
 #' Normalizar texto removiendo tildes y caracteres especiales
 #'
 #' @param texto Character vector a normalizar
-#' @return Character vector sin tildes ni caracteres especiales
+#' @returns Character vector sin tildes ni caracteres especiales
 #' @keywords internal
 #' @noRd
 normalizar_tildes <- function(texto) {
@@ -23,548 +25,6 @@ normalizar_tildes <- function(texto) {
     "aeiouunAEIOUUN",
     texto
   )
-}
-
-#' Diccionario de siglas medicas comunes en Chile
-#'
-#' @return Named list con siglas como keys y terminos de busqueda como values
-#' @keywords internal
-#' @noRd
-get_siglas_medicas <- function() {
-  # Retorna lista con categoria incluida: list(sigla = list(termino, categoria))
-  list(
-    # Cardiovasculares
-    "iam" = list(
-      termino = "infarto agudo miocardio",
-      categoria = "cardiovascular"
-    ),
-    "iamcest" = list(
-      termino = "infarto agudo miocardio",
-      categoria = "cardiovascular"
-    ),
-    "iamsest" = list(
-      termino = "infarto agudo miocardio",
-      categoria = "cardiovascular"
-    ),
-    "sca" = list(
-      termino = "sindrome coronario agudo",
-      categoria = "cardiovascular"
-    ),
-    "hta" = list(
-      termino = "hipertension arterial",
-      categoria = "cardiovascular"
-    ),
-    "aha" = list(
-      termino = "hipertension arterial",
-      categoria = "cardiovascular"
-    ),
-    "icc" = list(
-      termino = "insuficiencia cardiaca",
-      categoria = "cardiovascular"
-    ),
-    "ic" = list(
-      termino = "insuficiencia cardiaca",
-      categoria = "cardiovascular"
-    ),
-    "fa" = list(
-      termino = "fibrilacion auricular",
-      categoria = "cardiovascular"
-    ),
-    "tep" = list(
-      termino = "embolia pulmonar",
-      categoria = "cardiovascular"
-    ),
-    "tvp" = list(
-      termino = "trombosis venosa profunda",
-      categoria = "cardiovascular"
-    ),
-    "eap" = list(
-      termino = "edema agudo pulmon",
-      categoria = "cardiovascular"
-    ),
-    "acv" = list(
-      termino = "accidente cerebrovascular",
-      categoria = "cardiovascular"
-    ),
-    "ave" = list(
-      termino = "accidente vascular encefalico",
-      categoria = "cardiovascular"
-    ),
-    "ait" = list(
-      termino = "isquemico transitorio",
-      categoria = "cardiovascular"
-    ),
-
-    # Respiratorias
-    "tbc" = list(
-      termino = "tuberculosis",
-      categoria = "respiratoria"
-    ),
-    "tb" = list(
-      termino = "tuberculosis",
-      categoria = "respiratoria"
-    ),
-    "epoc" = list(
-      termino = "enfermedad pulmonar obstructiva cronica",
-      categoria = "respiratoria"
-    ),
[…]
