### 2026-05-03 01e6fc0
fix: restaurar ejemplo ejecutable en chunk de comorbilidades del README
El chunk comorb-example tenia eval condicional sobre rlang::is_installed
con la linea ejecutable comentada, por lo que no demostraba nada
incluso con comorbidity instalado. Restaurar eval=FALSE con la
sentencia visible como ejemplo de uso.

Tambien limpiar badge R-CMD-check duplicado: dejar solo el que apunta
al workflow renombrado R-CMD-check.yaml.
