+++ README.Rmd
@@ -39,9 +39,9 @@ if (lang == "es") {
 [![Lifecycle: stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://lifecycle.r-lib.org/articles/stages.html#stable)
 [![CRAN status](https://www.r-pkg.org/badges/version/ciecl)](https://CRAN.R-project.org/package=ciecl)
 [![GitHub release](https://img.shields.io/github/v/release/RodoTasso/ciecl)](https://github.com/RodoTasso/ciecl/releases)
-[![R-CMD-check](https://github.com/RodoTasso/ciecl/actions/workflows/r.yml/badge.svg)](https://github.com/RodoTasso/ciecl/actions/workflows/r.yml)
+[![R-CMD-check](https://github.com/Rodotasso/ciecl/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/Rodotasso/ciecl/actions/workflows/R-CMD-check.yaml)
 [![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/ciecl)](https://cran.r-project.org/package=ciecl)
-[![Coverage](https://img.shields.io/badge/coverage-96%25-brightgreen)](https://github.com/RodoTasso/ciecl/actions/workflows/test-coverage.yaml)
+[![Coverage](https://img.shields.io/badge/coverage-96%25-brightgreen)](https://github.com/Rodotasso/ciecl/actions/workflows/test-coverage.yaml)
 [![pkgdown](https://img.shields.io/badge/docs-pkgdown-blue)](https://rodotasso.github.io/ciecl/)
 <!-- badges: end -->
 
@@ -141,8 +141,8 @@ cie_lookup(c("E11.0", "I10", "Z00"))
 cie_describe(c("E11.0", "I10"))
 
 # Ejemplo en flujo dplyr
-# egresos %>% mutate(desc = cie_describe(codigo))
-```
+# egresos |> mutate(desc = cie_describe(codigo))
+
 # Busqueda fuzzy tolerante a errores
 cie_search("diabetis mellitus")
 
