### 2026-05-03 7a1336c
chore: normalizar metadata DESCRIPTION y guia de contribucion
- DESCRIPTION: limpiar Suggests no usados (usethis, litedown).
  Normalizar URL canonica del sitio pkgdown a minusculas.
- CONTRIBUTING.md: simplificar guia de encoding (UTF-8 declarado en
  DESCRIPTION; escapes \uXXXX solo donde R CMD check lo exige).
