### 2026-05-03 #765 @maurolepore
Hola @Rodotasso, this is to mark the start of my new EiC rotation.

I see @maelle has guided this through editor checks and just assigned @yabellini and @JDConejeros as reviewers. Looking forward to seeing this move through review.

Thanks for sharing your work with the rOpenSci community.
