### 2026-05-07 6d72267
docs: agregar badge de rOpenSci Software Peer Review (#765)
