### 2026-05-07 3e86b1e
docs: reconocer autoria del logo (@fje1) en README
