+++ index.md
@@ -67,23 +67,21 @@ library(ciecl)
 
 # Busqueda exacta por codigo
 cie_lookup("E11.0")
-#> # A tibble: 1 × 11
+#> # A tibble: 1 × 10
 #>   codigo descripcion       categoria seccion capitulo_nombre inclusion exclusion
 #>   <chr>  <chr>             <chr>     <chr>   <chr>           <chr>     <chr>    
 #> 1 E11.0  Diabetes mellitu… E11 DIAB… E08-E1… Cap.04  ENFERM… <NA>      <NA>     
-#> # ℹ 4 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>,
-#> #   uso_cl <chr>
+#> # ℹ 3 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>
 
 # Multiples codigos
 cie_lookup(c("E11.0", "I10", "Z00"))
-#> # A tibble: 3 × 11
+#> # A tibble: 3 × 10
 #>   codigo descripcion       categoria seccion capitulo_nombre inclusion exclusion
 #>   <chr>  <chr>             <chr>     <chr>   <chr>           <chr>     <chr>    
 #> 1 E11.0  Diabetes mellitu… E11 DIAB… E08-E1… Cap.04  ENFERM… <NA>      <NA>     
 #> 2 I10    Hipertensión ese… I10 HIPE… I10-I1… Cap.09  ENFERM… <NA>      <NA>     
 #> 3 Z00    Examen general e… Z00 EXAM… Z00-Z1… Cap.21  FACTOR… <NA>      <NA>     
-#> # ℹ 4 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>,
-#> #   uso_cl <chr>
+#> # ℹ 3 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>
 
 # Descripcion directa (vectorizada) para usar en mutate()
 cie_describe(c("E11.0", "I10"))
@@ -147,6 +145,10 @@ MIT + datos MINSAL de dominio público.
 **Rodolfo Tasso Suazo** — <rtasso@uchile.cl> Grupo de Ciencia de Datos
 para Salud Pública, Escuela de Salud Pública, Facultad de Medicina,
 Universidad de Chile.
+
+## Agradecimientos
+
+- Diseño del logotipo del paquete: [@fje1](https://github.com/fje1).
 
 ## Enlaces
 
