### 2026-05-08 d4b975c
ci: agregar workflow de lintr y modernizar configuracion
- Agrega .github/workflows/lint.yaml (template r-lib upstream limpio)
- Migra .lintr a sintaxis lintr 3.x (linters_with_defaults)
- Remueve with_defaults (deprecado) y cyclocomp_linter (inexistente)
- 0 lints reportados por lintr::lint_package()
