### 2026-05-08 cb488ab
style: limpiar whitespace y partir lineas largas
- Elimina trailing whitespace en archivos R/, tests/, vignettes/, tools/, data-raw/
- Parte 4 lineas >120 chars en cli::cli_abort() (cie-api.R, cie-comorbid.R, test-cie-sql.R)
- 0 cambios funcionales
