### 2026-05-08 54c84f5
fix: desambiguar sigla IRA en cie_search()
La sigla IRA es ambigua en contexto clinico chileno (infeccion respiratoria
aguda vs insuficiencia renal aguda). Antes se resolvia silenciosamente a
respiratoria, lo que podia confundir a usuarios buscando casos renales.

Cambios:
- siglas[['ira']] marcada con ambiguo=TRUE y aviso clinico
- alias explicitos: IRA_RESP (respiratoria) y IRA_RENAL (renal, ya existia)
- expandir_sigla() ahora retorna lista (termino, ambiguo, aviso)
- cie_search() emite cli::cli_warn cuando detecta sigla ambigua (verbose=TRUE)
- mantiene default respiratorio (uso clinico ma
[…]
