### 2026-05-08 954a1f4
docs: agregar badge de rOpenSci Software Peer Review (#765)
