### 2026-05-08 b60a24e
feat: agregar a Francisca Jofre como contribuidora del logo
- DESCRIPTION: Authors@R incluye a Francisca Jofre con rol ctb y comentario "Package logo design"
- inst/CITATION: agrega segunda autora ctb en bibentry
- man/ciecl-package.Rd: regenerado por roxygen2 reflejando Authors@R

Aparecera en CRAN como:
  Rodolfo Tasso Suazo [aut, cre] (ORCID), Francisca Jofre [ctb] (Package logo design)
