+++ authors.md
@@ -5,19 +5,22 @@
 - **Rodolfo Tasso Suazo**. Author, maintainer.
   [](https://orcid.org/0000-0002-1264-4933)
 
+- **Francisca Jofre**. Contributor.  
+  Package logo design
+
 ## Citation
 
 Source:
 [`inst/CITATION`](https://github.com/RodoTasso/ciecl/blob/main/inst/CITATION)
 
-Tasso Suazo R (2026). *ciecl: Clasificacion Internacional de
+Tasso Suazo R, Jofre F (2026). *ciecl: Clasificacion Internacional de
 Enfermedades CIE-10/11 para Chile*. R package version 0.9.3. Grupo de
 Ciencia de Datos para Salud Publica, Escuela de Salud Publica, U. de
 Chile. Datos: MINSAL/DEIS v2018, <https://github.com/Rodotasso/ciecl>.
 
     @Manual{,
       title = {ciecl: Clasificacion Internacional de Enfermedades CIE-10/11 para Chile},
-      author = {Rodolfo {Tasso Suazo}},
+      author = {Rodolfo {Tasso Suazo} and Francisca Jofre},
       year = {2026},
       note = {R package version 0.9.3. Grupo de Ciencia de Datos para Salud Publica, Escuela de Salud Publica, U. de Chile. Datos: MINSAL/DEIS v2018},
       url = {https://github.com/Rodotasso/ciecl},
