+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   idiomas: idiomas.html
   instalacion: instalacion.html
   installation: installation.html
-last_built: 2026-05-07T19:56Z
+last_built: 2026-05-08T20:56Z
 urls:
   reference: https://rodotasso.github.io/ciecl/reference
   article: https://rodotasso.github.io/ciecl/articles
