### 2026-05-08 d25d3e2
fix: corregir asociacion roxygen en cie_guia_busqueda
- Elimina linea en blanco entre roxygen y firma (rompia la asociacion de la doc)
- Compacta @seealso multilinea con coma colgante a una sola linea
- Doc cie_guia_busqueda() ahora se renderiza correctamente en help
