+++ man/cie_guide.Rd
@@ -16,8 +16,7 @@ Muestra tabla comparativa de cuando usar cada funcion de busqueda.
 cie_guide()
 }
 \seealso{
-\code{\link[=cie_search]{cie_search()}},
-\code{\link[=cie_lookup]{cie_lookup()}}, \code{\link[=cie_short]{cie_short()}}
+\code{\link[=cie_search]{cie_search()}}, \code{\link[=cie_lookup]{cie_lookup()}}, \code{\link[=cie_short]{cie_short()}}
 
 Other search: 
 \code{\link{cie_describe}()},
