### 2026-05-08 48eec18
refactor: eliminar @importFrom duplicado en cie-search.R
- Lineas 60-61 repetian imports ya declarados a nivel de archivo (1-5)
- Compacta @seealso multilinea con coma colgante
