### 2026-05-08 3b7e839
refactor: usar vapply en lugar de sapply en cie_table
vapply(..., logical(1)) es type-safe y mas explicito que sapply en
cie-table.R:42 al detectar columnas de notas vacias.

Sin cambio de comportamiento observable.
