### 2026-05-08 02c1d21
test: agregar tests faltantes para cie_describe, cie_guide y cie_guia_busqueda
- Nuevo archivo test-cie-describe.R con 13 tests cubriendo:
  - tipo y largo del retorno
  - descripciones conocidas (E11.0, I10)
  - flujo Auditoria (normalize=FALSE) vs Rescate (normalize=TRUE)
  - argumento default custom
  - vector vacio, NA en input, duplicados, vectorizacion en mutate
  - validacion de normalize (debe ser logico length-1 no-NA)
  - argumento espanol deprecado emite lifecycle warning

- Ampliado test-cie-search.R con tests para cie_guide y cie_guia_busqueda:
  - columnas esperadas (escenario + 
[…]
