### 2026-05-08 8748a50
ci: agregar workflows de tooling rOpenSci/CRAN
Tres workflows automatizados durante revision rOpenSci #765:

- pkgcheck.yaml: corre rOpenSci pkgcheck en push/PR (fail-on-error=false
  para que no bloquee pero reporte regresiones de checks comunitarios)
- urlchecker.yaml: scheduled mensual para detectar link rot antes que
  CRAN/rOpenSci lo flaggee
- codemeta.yaml: regenera codemeta.json automaticamente cuando cambia
  DESCRIPTION en main, manteniendo sync sin intervencion manual

Bloque E del plan P7 (parcial). E3 (desc package en scripts) y E5
(spelling con WORDLIST) quedan pendientes.
