+++ .github/workflows/urlchecker.yaml
@@ -1,0 +1,41 @@
+# Workflow scheduled mensual para detectar link rot en docs/vignettes/README
+# antes que CRAN/rOpenSci lo flaggee.
+# Documentacion: https://github.com/r-lib/urlchecker
+on:
+  schedule:
+    # Primer dia de cada mes 06:00 UTC (~02:00 GMT-4 Chile)
+    - cron: '0 6 1 * *'
+  workflow_dispatch:
+
+name: urlchecker.yaml
+
+permissions: read-all
+
+jobs:
+  urlchecker:
+    runs-on: ubuntu-latest
+    env:
+      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: r-lib/actions/setup-r@v2
+        with:
+          use-public-rspm: true
+
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        with:
+          extra-packages: any::urlchecker, local::.
+          needs: check
+
+      - name: Check URLs
+        run: |
+          res <- urlchecker::url_check()
+          if (nrow(res) > 0) {
+            print(res)
+            cat("\n[urlchecker] ", nrow(res), " URLs con problemas. ",
+                "Revisar y corregir antes de la siguiente release.\n", sep = "")
+          } else {
+            cat("[urlchecker] OK - todas las URLs validan correctamente.\n")
+          }
+        shell: Rscript {0}
