### 2026-05-09 9b36b87
test: subir cobertura de 86% a 96% con patrones r-lib
Cuatro archivos nuevos de tests cubriendo paths previamente sin cobertura:

- test-deprecated-args.R (44 tests): paths de lifecycle::deprecate_warn()
  para args en espanol en cie_norm, cie_normalizar, cie_normalize,
  cie_validate_vector, cie_expand, cie_search, cie_lookup, cie11_search,
  cie_table, cie_map_comorbid

- test-coverage-paths.R (22 tests): ramas secundarias en cie_search
  (field='inclusion', palabras cortas en FTS, sigla ambigua con verbose)
  y cie_lookup (extract=TRUE, check_siglas=TRUE, vector mixto con rangos,
  vector v
[…]
