+++ tests/testthat/test-cie-sql-cache.R
@@ -1,0 +1,169 @@
+# ============================================================
+# COBERTURA: ciclo de vida del cache SQLite en cie-sql.R
+# Cubre build_cache_atomic (creacion de cache_dir, cleanup
+# de .tmp residual, error handling) + build_fts (failsafe)
+# + cache_is_current (sin metadata, version mismatch).
+# Usa patrones r-lib: withr::local_tempdir, local_envvar,
+# local_mocked_bindings, defer.
+# ============================================================
+
+# Helper: crear conexion in-memory aislada para tests de helpers internos
+make_test_con <- function() {
+  con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
+  withr::defer(DBI::dbDisconnect(con), envir = parent.frame())
+  con
+}
+
+# --- build_fts: failsafe FTS5 missing -------------------------------------
+
+test_that("build_fts crea tabla FTS5 sobre conexion existente", {
+  con <- make_test_con()
+
+  # Crear tabla cie10 minima (build_fts requiere que exista)
+  DBI::dbExecute(con, "CREATE TABLE cie10 (codigo TEXT, descripcion TEXT,
+                                            inclusion TEXT, exclusion TEXT)")
+  DBI::dbExecute(con, "INSERT INTO cie10 VALUES
+                       ('E11.0', 'diabetes con coma', NULL, NULL)")
+
+  build_fts <- getFromNamespace("build_fts", "ciecl")
+  expect_silent(build_fts(con))
+
+  expect_true(DBI::dbExistsTable(con, "cie10_fts"))
+})
+
+# --- cache_is_current ------------------------------------------------------
+
+test_that("cache_is_current retorna FALSE si no existe tabla cie10_meta", {
+  con <- make_test_con()
+
+  cache_is_current <- getFromNamespace("cache_is_current", "ciecl")
+  expect_false(cache_is_current(con))
+})
+
+test_that("cache_is_current retorna FALSE si tabla cie10_meta esta vacia", {
+  con <- make_test_con()
+  DBI::dbExecute(con, "CREATE TABLE cie10_meta (key TEXT PRIMARY KEY,
+                                                  value TEXT)")
+
+  cache_is_current <- getFromNamespace("cache_is_current", "ciecl")
+  expect_false(cache_is_current(con))
+})
+
+test_that("cache_is_current retorna FALSE si version no coincide", {
+  con <- make_test_con()
+  DBI::dbExecute(con, "CREATE TABLE cie10_meta (key TEXT PRIMARY KEY,
+                                                  value TEXT)")
+  DBI::dbExecute(
+    con,
+    "INSERT INTO cie10_meta (key, value) VALUES ('cache_version', '0.0.0')"
+  )
+
+  cache_is_current <- getFromNamespace("cache_is_current", "ciecl")
+  expect_false(cache_is_current(con))
+})
+
+test_that("cache_is_current retorna TRUE cuando version coincide", {
+  con <- make_test_con()
+  DBI::dbExecute(con, "CREATE TABLE cie10_meta (key TEXT PRIMARY KEY,
+                                                  value TEXT)")
+  pkg_v <- as.character(utils::packageVersion("ciecl"))
+  DBI::dbExecute(
+    con,
+    sprintf("INSERT INTO cie10_meta (key, value)
+             VALUES ('cache_version', '%s')", pkg_v)
+  )
+
+  cache_is_current <- getFromNamespace("cache_is_current", "ciecl")
+  expect_true(cache_is_current(con))
+})
+
[…]
