+++ tests/testthat/test-deprecated-args.R
@@ -1,0 +1,205 @@
+# ============================================================
+# COBERTURA: paths de deprecacion de argumentos espanol -> ingles
+# Cierra los paths `lifecycle::deprecate_warn()` que estaban
+# sin cobertura en cie-utils.R, cie-search.R, cie-lookup.R.
+# ============================================================
+
+# --- cie_norm ---------------------------------------------------------------
+
+test_that("cie_norm acepta arg deprecado 'codigos' con warning", {
+  expect_warning(
+    res <- cie_norm(codigos = "E110"),
+    class = "lifecycle_warning_deprecated"
+  )
+  expect_equal(res, cie_norm("E110"))
+})
+
+test_that("cie_norm acepta arg deprecado 'buscar_db' con warning", {
+  expect_warning(
+    res <- cie_norm("E11.0", buscar_db = FALSE),
+    class = "lifecycle_warning_deprecated"
+  )
+  expect_equal(res, cie_norm("E11.0", search_db = FALSE))
+})
+
+test_that("cie_norm convierte factor a character", {
+  fac <- factor("E11.0")
+  res <- cie_norm(fac, search_db = FALSE)
+  expect_type(res, "character")
+  expect_equal(res, "E11.0")
+})
+
+# --- cie_normalizar (deprecada completa) ------------------------------------
+
+test_that("cie_normalizar emite warning de deprecacion completa", {
+  expect_warning(
+    res <- cie_normalizar("E110"),
+    class = "lifecycle_warning_deprecated"
+  )
+  expect_equal(res, cie_norm("E110"))
+})
+
+test_that("cie_normalizar pasa buscar_db al backend", {
+  expect_warning(
+    res <- cie_normalizar("E11.0", buscar_db = FALSE),
+    class = "lifecycle_warning_deprecated"
+  )
+  expect_equal(res, "E11.0")
+})
+
+# --- cie_normalize (deprecada completa, alias EN) ---------------------------
+
+test_that("cie_normalize emite warning de deprecacion completa", {
+  expect_warning(
+    res <- cie_normalize("E110"),
+    class = "lifecycle_warning_deprecated"
+  )
+  expect_equal(res, cie_norm("E110"))
+})
+
+test_that("cie_normalize acepta arg deprecado 'codigos'", {
+  expect_warning(
+    expect_warning(
+      res <- cie_normalize(codigos = "E110"),
+      class = "lifecycle_warning_deprecated"
+    ),
+    class = "lifecycle_warning_deprecated"
+  )
+  expect_equal(res, cie_norm("E110"))
+})
+
+test_that("cie_normalize acepta arg deprecado 'buscar_db'", {
+  expect_warning(
+    expect_warning(
+      res <- cie_normalize("E11.0", buscar_db = FALSE),
+      class = "lifecycle_warning_deprecated"
+    ),
+    class = "lifecycle_warning_deprecated"
+  )
+  expect_equal(res, "E11.0")
+})
+
+# --- cie_validate_vector ---------------------------------------------------
+
+test_that("cie_validate_vector acepta arg deprecado 'codigos'", {
+  expect_warning(
+    res <- cie_validate_vector(codigos = c("E11.0", "INVALIDO")),
+    class = "lifecycle_warning_deprecated"
+  )
+  expect_length(res, 2)
+  expect_true(res[1])
+  expect_false(res[2])
+})
+
+# --- cie_expand ------------------------------------------------------------
+
+test_that("cie_expand acepta arg deprecado 'codigo'", {
+  expect_warning(
+    res <- cie_expand(codigo = "E11"),
+    class = "lifecycle_warning_deprecated"
+  )
[…]

[…]
