### 2026-05-09 db57c46
ci: corregir input invalido en pkgcheck workflow
El input 'fail-on-error' no existe en ropensci-review-tools/pkgcheck-action.
Inputs validos (verificado en log de fallo 25581631365):
  ref, post-to-issue, issue-title, summary-only, append-to-issue

Cambio a 'summary-only: true' que postea el reporte al summary del workflow
(visible en pestana Actions) en vez de intentar postear a un Issue. Esto
ademas elimina la necesidad de permisos issues:write que el workflow no
tiene declarados.

Resultado del job al ejecutarse correctamente: pkgcheck reporta
- coverage 86.5% (sera 96.61% tras tests recien
[…]
