### 2026-05-09 16d1f2c
ci: desactivar post-to-issue en pkgcheck workflow
El input 'summary-only: true' agrega el reporte al workflow summary
pero NO inhibe el POST a /issues por defecto. Eso causaba el error
HTTP 403 'Resource not accessible by integration' (necesita issues:write).

Fix: agregar 'post-to-issue: false' explicito.

Verificado en run 25591247637: el job CORRIO bien, pkgcheck reporto
coverage 96.6% y todos los checks rOpenSci en verde (excepto goodpractice
marcado optional). El job completo solo fallaba al final por el POST a
issues fallido.
