### 2026-05-09 f53515f
docs: usar badge dinamico de coverage en lugar de hardcoded 96%
Antes: badge estatico apuntaba a img.shields.io/badge/coverage-96%25-...
con el texto '96%' literal en el URL, independiente del coverage real.

Ahora: badge dinamico via shields.io endpoint que lee badges/coverage.json
(el JSON que ya genera el workflow test-coverage.yaml en cada push).

Resultado: el badge refleja el coverage real del paquete en main en cada
momento. Cuando dev se merge a main, el badge salta del 86% (last main)
al 96.6% (current dev) sin intervencion manual.

URL nuevo:
  https://img.shields.io/endpoint?url=htt
[…]
