### 2026-05-09 e9e10ec
refactor(tests): adoptar helpers canonicos r-lib
Tres refactors basados en auditoria contra D:\01_PROYECTOS\r-lib:

1. lifecycle::expect_deprecated() en lugar de
   expect_warning(class = 'lifecycle_warning_deprecated').
   El helper canonico setea lifecycle_verbosity = 'warning' antes
   del test, evitando la deduplicacion 'once per session' de
   lifecycle::deprecate_warn que produciria falsos negativos
   silenciosos si los tests se reordenan o duplican.
   Aplicado a 13 ocurrencias en test-deprecated-args.R.

2. withr::local_db_connection() en lugar de helper custom
   make_test_con() + wi
[…]
