+++ tests/testthat/test-deprecated-args.R
@@ -2,22 +2,25 @@
 # COBERTURA: paths de deprecacion de argumentos espanol -> ingles
 # Cierra los paths `lifecycle::deprecate_warn()` que estaban
 # sin cobertura en cie-utils.R, cie-search.R, cie-lookup.R.
+#
+# Usa lifecycle::expect_deprecated() (helper canonico r-lib) en
+# lugar de expect_warning(class = "lifecycle_warning_deprecated").
+# expect_deprecated setea lifecycle_verbosity = "warning" antes
+# del test, evitando la deduplicacion "once per session" de
+# lifecycle::deprecate_warn que produciria falsos negativos si
+# tests se reordenan o duplican.
 # ============================================================
 
 # --- cie_norm ---------------------------------------------------------------
 
 test_that("cie_norm acepta arg deprecado 'codigos' con warning", {
-  expect_warning(
-    res <- cie_norm(codigos = "E110"),
-    class = "lifecycle_warning_deprecated"
-  )
+  lifecycle::expect_deprecated(res <- cie_norm(codigos = "E110"))
   expect_equal(res, cie_norm("E110"))
 })
 
 test_that("cie_norm acepta arg deprecado 'buscar_db' con warning", {
-  expect_warning(
-    res <- cie_norm("E11.0", buscar_db = FALSE),
-    class = "lifecycle_warning_deprecated"
+  lifecycle::expect_deprecated(
+    res <- cie_norm("E11.0", buscar_db = FALSE)
   )
   expect_equal(res, cie_norm("E11.0", search_db = FALSE))
 })
@@ -32,17 +35,13 @@ test_that("cie_norm convierte factor a character", {
 # --- cie_normalizar (deprecada completa) ------------------------------------
 
 test_that("cie_normalizar emite warning de deprecacion completa", {
-  expect_warning(
-    res <- cie_normalizar("E110"),
-    class = "lifecycle_warning_deprecated"
-  )
+  lifecycle::expect_deprecated(res <- cie_normalizar("E110"))
   expect_equal(res, cie_norm("E110"))
 })
 
 test_that("cie_normalizar pasa buscar_db al backend", {
-  expect_warning(
-    res <- cie_normalizar("E11.0", buscar_db = FALSE),
-    class = "lifecycle_warning_deprecated"
+  lifecycle::expect_deprecated(
+    res <- cie_normalizar("E11.0", buscar_db = FALSE)
   )
   expect_equal(res, "E11.0")
 })
@@ -50,31 +49,25 @@ test_that("cie_normalizar pasa buscar_db al backend", {
 # --- cie_normalize (deprecada completa, alias EN) ---------------------------
 
 test_that("cie_normalize emite warning de deprecacion completa", {
-  expect_warning(
-    res <- cie_normalize("E110"),
-    class = "lifecycle_warning_deprecated"
-  )
+  lifecycle::expect_deprecated(res <- cie_normalize("E110"))
   expect_equal(res, cie_norm("E110"))
 })
 
 test_that("cie_normalize acepta arg deprecado 'codigos'", {
-  expect_warning(
-    expect_warning(
-      res <- cie_normalize(codigos = "E110"),
-      class = "lifecycle_warning_deprecated"
-    ),
-    class = "lifecycle_warning_deprecated"
+  # Doble warning: uno por cie_normalize completa + uno por arg codigos
+  lifecycle::expect_deprecated(
+    lifecycle::expect_deprecated(
+      res <- cie_normalize(codigos = "E110")
+    )
   )
   expect_equal(res, cie_norm("E110"))
 })
 
[…]
