### 2026-05-18 001b453
test(a5): aislar snapshots con testthat::local_reproducible_output
Agrega testthat::local_reproducible_output() al inicio de cada test_that()
que usa expect_snapshot() en cie-search, cie-api, cie-table y robustness.
Aisla ancho de terminal, color y locale para evitar falsos positivos
cross-platform en snapshots.

Plan P7 Bloque A item A5.
