+++ tests/testthat/test-cie-api.R
@@ -1,12 +1,14 @@
 # Tests para cie11_search (API CIE-11 OMS)
 
 test_that("cie11_search requiere API key", {
+  testthat::local_reproducible_output()
   withr::with_envvar(c(ICD_API_KEY = NA), {
     expect_snapshot(cie11_search("diabetes"), error = TRUE)
   })
 })
 
 test_that("cie11_search valida formato de API key", {
+  testthat::local_reproducible_output()
   skip_if_not_installed("httr2")
 
   # API key sin separador ":"
@@ -17,6 +19,7 @@ test_that("cie11_search valida formato de API key", {
 })
 
 test_that("cie11_search requiere httr2", {
+  testthat::local_reproducible_output()
   # Solo corre si httr2 NO esta disponible para verificar el error
   skip_if(requireNamespace("httr2", quietly = TRUE), "httr2 esta instalado")
 
@@ -36,8 +39,10 @@ test_that("cie11_search con API real retorna resultados", {
 
   # Verificar que existe API key en environment
   api_key <- Sys.getenv("ICD_API_KEY", unset = NA)
-  skip_if(is.na(api_key) || api_key == "",
-          "ICD_API_KEY no configurada para tests reales")
+  skip_if(
+    is.na(api_key) || api_key == "",
+    "ICD_API_KEY no configurada para tests reales"
+  )
 
   # Buscar termino comun que debe existir
   resultado <- cie11_search("diabetes", max_results = 2)
@@ -59,8 +64,10 @@ test_that("cie11_search con API real maneja busqueda sin resultados", {
   skip_if_not_installed("httr2")
 
   api_key <- Sys.getenv("ICD_API_KEY", unset = NA)
-  skip_if(is.na(api_key) || api_key == "",
-          "ICD_API_KEY no configurada para tests reales")
+  skip_if(
+    is.na(api_key) || api_key == "",
+    "ICD_API_KEY no configurada para tests reales"
+  )
 
   # Buscar termino que probablemente no existe
   expect_message(
@@ -86,6 +93,7 @@ test_that("cie11_search retorna tibble con tipos correctos incluso en error", {
 })
 
 test_that("cie11_search valida tipos de input", {
+  testthat::local_reproducible_output()
   skip_if_not_installed("httr2")
 
   expect_snapshot(cie11_search(123), error = TRUE)
