### 2026-05-18 a14a72b
refactor(a4): usar rlang::check_required() en args obligatorios
Sustituye fallos implicitos de R ("argument missing, with no default")
por mensaje uniforme via rlang::check_required() en args sin default:

- cie_search(): text
- cie_comorbid(): data, id, code
- cie_map_comorbid(): codes

Los check_required() se posicionan despues del manejo de args en espanol
(deprecated) para no gatillarse cuando se llama con texto=/codigos=.

Plan P7 Bloque A item A4 (sin cie-api.R, que va en commit posterior).
