### 2026-05-18 bcfecfd
feat(a3): cli_progress_step en build_cache_atomic y build_fts
Reemplaza cli_inform() aislados por una secuencia visible de progreso
durante la construccion inicial del cache SQLite (~16s):

  1. Cargando dataset cie10_cl
  2. Escribiendo tabla cie10
  3. Construyendo indices
  4. Creando tabla FTS5

build_fts() acepta nuevo flag .progress = TRUE para suprimirlo cuando
build_cache_atomic() ya muestra el progreso (evita anidacion).

En sesiones no interactivas (CI, tests) se mantiene completamente
silencioso. expect_silent(build_fts(con)) sigue verde.

Nota: cobertura de R/cie-sql.R baja de ~89.
[…]
