+++ R/cie-sql.R
@@ -33,10 +33,9 @@ get_cie10_db <- function() {
   # Pooling: reutilizar conexion existente si es valida y apunta al mismo path
 
   if (!is.null(.ciecl_env$con) &&
-      inherits(.ciecl_env$con, "SQLiteConnection") &&
-      DBI::dbIsValid(.ciecl_env$con) &&
-      identical(.ciecl_env$db_path, db_path)) {
-
+    inherits(.ciecl_env$con, "SQLiteConnection") &&
+    DBI::dbIsValid(.ciecl_env$con) &&
+    identical(.ciecl_env$db_path, db_path)) {
     # Failsafe: verificar que FTS5 existe (fix cache parcial)
     if (!DBI::dbExistsTable(.ciecl_env$con, "cie10_fts")) {
       build_fts(.ciecl_env$con)
@@ -120,57 +119,101 @@ build_cache_atomic <- function(cache_dir, db_path) {
   }
 
   con <- DBI::dbConnect(RSQLite::SQLite(), tmp_path)
-  on.exit({
-    if (DBI::dbIsValid(con)) DBI::dbDisconnect(con)
-  }, add = TRUE)
+  on.exit(
+    {
+      if (DBI::dbIsValid(con)) DBI::dbDisconnect(con)
+    },
+    add = TRUE
+  )
 
-  tryCatch({
-    # Cargar datos
-    utils::data(cie10_cl, package = "ciecl", envir = environment())
-    DBI::dbWriteTable(con, "cie10", cie10_cl, overwrite = TRUE)
+  # Progress steps solo en sesion interactiva para no contaminar tests/CI.
+  # Usamos un id propio para gestionar el ciclo de vida manualmente y poder
+  # cerrar el progreso en caso de error.
+  show_progress <- interactive()
+  progress_id <- NULL
 
-    # Indices
-    DBI::dbExecute(con, "CREATE INDEX idx_codigo ON cie10(codigo)")
-    DBI::dbExecute(con, "CREATE INDEX idx_desc ON cie10(descripcion)")
+  tryCatch(
+    {
+      # Etapa 1: Cargar dataset cie10_cl
+      if (show_progress) {
+        progress_id <- cli::cli_progress_step(
+          "Cargando dataset {.field cie10_cl}",
+          msg_done = "Dataset {.field cie10_cl} cargado"
+        )
+      }
+      utils::data(cie10_cl, package = "ciecl", envir = environment())
 
-    # FTS5
-    build_fts(con)
+      # Etapa 2: Escribir tabla cie10
+      if (show_progress) {
+        cli::cli_progress_step(
+          "Escribiendo tabla {.field cie10}",
+          msg_done = "Tabla {.field cie10} escrita"
+        )
+      }
+      DBI::dbWriteTable(con, "cie10", cie10_cl, overwrite = TRUE)
 
-    # Metadata con version
-    DBI::dbExecute(con, "
+      # Etapa 3: Construir indices
+      if (show_progress) {
+        cli::cli_progress_step(
+          "Construyendo indices",
+          msg_done = "Indices construidos"
+        )
+      }
+      DBI::dbExecute(con, "CREATE INDEX idx_codigo ON cie10(codigo)")
+      DBI::dbExecute(con, "CREATE INDEX idx_desc ON cie10(descripcion)")
+
+      # Etapa 4: Crear FTS5
+      if (show_progress) {
+        cli::cli_progress_step(
+          "Creando tabla {.field FTS5}",
+          msg_done = "Tabla {.field FTS5} creada"
+        )
+      }
+      build_fts(con, .progress = FALSE)
+
+      # Metadata con version (sin step propio, parte del cierre)
+      DBI::dbExecute(con, "
[…]
