### 2026-05-18 3942266
feat(a1,a2): modernizar OAuth y manejo de errores en cie11_search
A1: Migracion del flujo OAuth manual a httr2 nativo.
  - httr2::oauth_client() construye el cliente desde "client_id:secret".
  - httr2::req_oauth_client_credentials() obtiene, cachea y refresca el
    token de acceso automaticamente en cada peticion. Elimina el bloque
    manual de POST al token endpoint + parseo JSON + header Authorization.

A2: Errores HTTP de la OMS ahora son tipados httr2_http_*.
  - httr2::resp_check_status() despues de req_perform() asegura el tipado.
  - httr2::req_error(body = who_error_body) extrae err
[…]
