+++ R/cie-api.R
@@ -36,6 +36,7 @@ cie11_search <- function(text, api_key = NULL, lang = c("es", "en"),
     text <- texto
   }
 
+  rlang::check_required(text)
   lang <- rlang::arg_match(lang)
 
   # Validacion de inputs
@@ -49,11 +50,11 @@ cie11_search <- function(text, api_key = NULL, lang = c("es", "en"),
     cli::cli_abort("{.arg text} no puede estar vacio.", class = "ciecl_invalid_input")
   }
   if (!is.numeric(max_results) || length(max_results) != 1 ||
-      max_results < 1 || max_results != as.integer(max_results)) {
+    max_results < 1 || max_results != as.integer(max_results)) {
     cli::cli_abort("{.arg max_results} debe ser un entero positivo.", class = "ciecl_invalid_input")
   }
   if (!is.character(release) || length(release) != 1 ||
-      !grepl("^\\d{4}-\\d{2}$", release)) {
+    !grepl("^\\d{4}-\\d{2}$", release)) {
     cli::cli_abort(
       "{.arg release} debe ser formato {.val YYYY-MM} (ej. {.val 2024-01}).",
       class = "ciecl_invalid_input"
@@ -71,92 +72,111 @@ cie11_search <- function(text, api_key = NULL, lang = c("es", "en"),
     }
   }
 
-  # Separar client_id y client_secret
+  # Construir cliente OAuth (parsea "client_id:client_secret" del api_key)
   credentials <- strsplit(api_key, ":")[[1]]
   if (length(credentials) != 2) {
-    cli::cli_abort("API key debe tener formato {.val client_id:client_secret}", class = "ciecl_invalid_input")
-  }
-  client_id <- credentials[1]
-  client_secret <- credentials[2]
-
-  tryCatch({
-    # Paso 1: Obtener token OAuth
-    token_url <- "https://icdaccessmanagement.who.int/connect/token"
-    token_req <- httr2::request(token_url) |>
-      httr2::req_method("POST") |>
-      httr2::req_user_agent("ciecl (https://github.com/Rodotasso/ciecl)") |>
-      httr2::req_timeout(30) |>
-      httr2::req_retry(max_tries = 3) |>
-      httr2::req_body_form(
-        client_id = client_id,
-        client_secret = client_secret,
-        scope = "icdapi_access",
-        grant_type = "client_credentials"
-      )
-
-    token_resp <- httr2::req_perform(token_req)
-    token_data <- httr2::resp_body_json(token_resp)
-    access_token <- token_data$access_token
-
-    # Paso 2: Buscar en CIE-11 con el token
-    search_url <- paste0(
-      "https://id.who.int/icd/release/11/", release, "/mms/search"
+    cli::cli_abort(
+      "API key debe tener formato {.val client_id:client_secret}",
+      class = "ciecl_invalid_input"
     )
+  }
 
-    search_req <- httr2::request(search_url) |>
-      httr2::req_user_agent("ciecl (https://github.com/Rodotasso/ciecl)") |>
-      httr2::req_timeout(30) |>
-      httr2::req_retry(max_tries = 3) |>
-      httr2::req_throttle(rate = 10 / 60) |> # 10 req/min para ser conservador
-      httr2::req_url_query(
-        q = text,
-        flatResults = "true",
-        useFlexisearch = "true"
-      ) |>
-      httr2::req_headers(
-        Authorization = paste("Bearer", access_token),
-        `API-Version` = "v2",
-        `Accept-Language` = lang
+  who_client <- httr2::oauth_client(
+    id = credentials[1],
+    secret = credentials[2],
+    token_url = "https://icdaccessmanagement.who.int/connect/token",
+    name = "ciecl"
+  )
+
+  # Extractor de detalle desde body de error OMS (para condiciones tipadas
[…]
