### 2026-05-18 bd53f7c
docs: registrar Bloque A y sprint cobertura de mayo en NEWS
Agrega al inicio de la entrada 0.9.8 dos secciones fechadas:

  - Modernizacion r-lib - Bloque A (2026-05-17): httr2 OAuth,
    resp_check_status, cli_progress_step, check_required,
    local_reproducible_output.
  - Sprint cobertura y tooling rOpenSci (2026-05-08/09): subida
    de 86.45% a 96.61% de coverage, workflows pkgcheck/urlchecker/
    codemeta, badge dinamico, fix IRA, limpieza importFrom/vapply.

La entrada 0.9.8 pasa a marcarse como "en desarrollo, 2026-04-25 ->
2026-05-17" hasta que se cierre la version.
