### 2026-05-20 02f70ae
test(cie-sql): cubrir cli_progress en sesion interactiva
Reemplaza interactive() por rlang::is_interactive() en build_cache_atomic
y build_fts para que rlang::local_interactive(TRUE) pueda activar el path
cli_progress en tests sin TTY real.

Suma 4 tests a test-cie-sql-cache.R que cubren:
- build_fts emite cli_progress_step con .progress = TRUE
- build_fts permanece silencioso con .progress = FALSE
- build_cache_atomic emite los 4 steps (dataset, tabla, indices, FTS5)
- handler de error cierra el progress y limpia .tmp residual

Cobertura R/cie-sql.R: 81.07% -> 96.60%
