### 2026-05-20 86cdc1f
refactor(cie-api): extraer who_error_body y adoptar local_mocked_responses
Cambios en R/cie-api.R:
- who_error_body() sale del cuerpo de cie11_search() como helper interno
  @noRd. Misma logica de extraccion (error_description > error > message)
  pero ahora unit-testeable directamente.
- El warning del tryCatch externo pasa de {e$message} a {conditionMessage(e)}
  para propagar el detalle del body OMS cuando la API responde 4xx
  (antes mostraba solo "HTTP 401 Unauthorized").

Refactor de tests/testthat/test-api-mock.R al patron canonico httr2:
- httr2::local_mocked_responses(mock_who_api(...
[…]
