+++ tests/testthat/test-api-mock.R
@@ -139,45 +139,54 @@ test_that("cie11_search informa sobre httr2 faltante", {
 })
 
 # ============================================================
-# PRUEBAS CON HTTP MOCKING (local_mocked_bindings)
-# Cubren lineas 41-95 de cie-api.R: OAuth token + search + parsing
+# HELPER de mock canonico httr2 (local_mocked_responses)
+# Distingue por URL el endpoint OAuth de token del endpoint de
+# busqueda OMS para que el flow real de httr2 corra hasta el
+# final (incluido el callback req_error(body = who_error_body)).
+# ============================================================
+
+mock_who_api <- function(search_body = list(destinationEntities = list()),
+                         search_status = 200L) {
+  function(req) {
+    if (grepl("/connect/token", req$url, fixed = TRUE)) {
+      httr2::response_json(body = list(
+        access_token = "fake_token",
+        token_type = "Bearer",
+        expires_in = 3600L
+      ))
+    } else {
+      httr2::response_json(status_code = search_status, body = search_body)
+    }
+  }
+}
+
+# ============================================================
+# PRUEBAS CON HTTP MOCKING (httr2::local_mocked_responses)
+# El flow real de httr2 corre hasta el final: OAuth token +
+# search + parsing JSON + handle_resp.
 # ============================================================
 
 test_that("cie11_search retorna resultados con mock HTTP exitoso", {
   skip_if_not_installed("httr2")
 
-  call_count <- 0L
-
-  local_mocked_bindings(
-    req_perform = function(req, ...) {
-      call_count <<- call_count + 1L
-      structure(list(call_id = call_count, status_code = 200L), class = "httr2_response")
-    },
-    resp_body_json = function(resp, ...) {
-      # Con req_oauth_client_credentials(), la mock de req_perform reemplaza
-      # el flow completo (OAuth + search), por lo que solo se recibe la
-      # llamada de busqueda. Devolvemos la search response directamente.
-      list(destinationEntities = data.frame(
-        theCode = c("5A00", "5A01", "5A02"),
-        title = c(
-          "<em class='found'>Diabetes</em> mellitus tipo 1",
-          "<em class='found'>Diabetes</em> mellitus tipo 2",
-          "Otra <em class='found'>diabetes</em> mellitus"
-        ),
-        chapter = c("05", "05", "05"),
-        stringsAsFactors = FALSE
-      ))
-    },
-    .package = "httr2"
-  )
+  httr2::local_mocked_responses(mock_who_api(
+    search_body = list(destinationEntities = data.frame(
+      theCode = c("5A00", "5A01", "5A02"),
+      title = c(
+        "<em class='found'>Diabetes</em> mellitus tipo 1",
+        "<em class='found'>Diabetes</em> mellitus tipo 2",
+        "Otra <em class='found'>diabetes</em> mellitus"
+      ),
+      chapter = c("05", "05", "05"),
+      stringsAsFactors = FALSE
+    ))
+  ))
 
   resultado <- cie11_search("diabetes", api_key = "test_id:test_secret")
 
   expect_s3_class(resultado, "tbl_df")
   expect_equal(nrow(resultado), 3)
   expect_equal(resultado$codigo, c("5A00", "5A01", "5A02"))
-  # Verificar que HTML tags fueron limpiados
-
   expect_no_match(resultado$titulo[1], "<em", fixed = TRUE)
   expect_match(resultado$titulo[1], "Diabetes mellitus tipo 1")
   expect_equal(resultado$capitulo, c("05", "05", "05"))
@@ -186,23 +195,14 @@ test_that("cie11_search retorna resultados con mock HTTP exitoso", {
[…]
