### 2026-05-20 cbf1253
docs: registrar sprint F1 (cobertura 94.20 -> 97.31) en NEWS
Documenta los dos commits previos del sprint F1:
- test(cie-sql): cli_progress en sesion interactiva
- refactor(cie-api): who_error_body + local_mocked_responses

Sin cambios de API publica. Detalla los nuevos patrones r-lib aplicados
(rlang::is_interactive, httr2::local_mocked_responses, response_json).
