### 2026-05-21 a905147
feat(uso_cl): exponer include_uso_cl y only_uso_cl en cie_lookup y cie_search
Agrega dos flags a las dos funciones de busqueda principales para
controlar la columna uso_cl del dataset CIE-10 MINSAL/DEIS v2018:

- include_uso_cl: incluir la columna en el output
- only_uso_cl: filtrar a codigos vigentes (excluye uso_cl == "legado")

Defaults asimetricos por backward compat:
- cie_lookup(include_uso_cl = TRUE): preserva el output historico
- cie_search(include_uso_cl = FALSE): preserva el output historico

La semantica de uso_cl es textual (no booleano): principal, legado,
causa_externa, etiologi
[…]
