### 2026-05-21 afaaa3e
docs: registrar sprint F2 (include_uso_cl / only_uso_cl) en NEWS
Documenta el commit anterior: flags de exposicion y filtrado por uso
clinico vigente en Chile en cie_lookup() y cie_search(). Incluye
defaults asimetricos por backward compat, semantica textual de
uso_cl, y patrones r-lib usados.
