### 2026-05-21 a1205b8
refactor(b3): lifecycle::deprecated en args de cie_normalizar
Formaliza la deprecacion de cie_normalizar() con sentinel
lifecycle::deprecated() en ambos args (codigos, buscar_db) siguiendo
el patron canonico de r-lib (lifecycle/R/arg.R, dplyr/R/group-split.R).

- Args con default lifecycle::deprecated() en lugar de TRUE/missing
- rlang::check_required(codigos) para arg obligatorio (rlang/R/arg.R:199)
- lifecycle::is_present(buscar_db) para extraer valor opcional
- Roxygen @param con lifecycle::badge('deprecated') visible en HTML/PDF
- Mantiene UN solo deprecate_warn por funcion (no por arg) co
[…]
