### 2026-05-21 7c57081
chore(docs): regenerar .Rd con roxygen2 8.0.0
Re-render mecanico de la documentacion tras actualizar roxygen2 a
8.0.0 (vino con R 4.6 + devtools). Cambios cosmeticos:

- DESCRIPTION: RoxygenNote: 7.3.2 -> Config/roxygen2/version: 8.0.0
- .Rd: \link{foo}() -> \link[=foo]{foo()} (links mas explicitos)
- .Rd: trailing space removido tras 'Other ...:'
- .Rd: \dontshow{} simplificado en cie_lookup

Sin cambios de comportamiento. Mantiene consistencia para que
devtools::document() en sesiones futuras no re-genere lo mismo.
