### 2026-05-24 5ee5a63
fix: corregir Authors@R — afiliacion en comment del autor, ctb solo para diseno del logo
