+++ man/ciecl-package.Rd
@@ -20,17 +20,16 @@ Useful links:
 
 }
 \author{
-\strong{Maintainer}: Rodolfo Tasso Suazo \email{rtasso@uchile.cl} (\href{https://orcid.org/0000-0002-1264-4933}{ORCID})
+\strong{Maintainer}: Rodolfo Tasso Suazo \email{rtasso@uchile.cl} (\href{https://orcid.org/0000-0002-1264-4933}{ORCID}) (Grupo de Ciencia de Datos para Salud Pública, Escuela de Salud Pública, U. de Chile)
 
 Authors:
 \itemize{
-  \item Rodolfo Tasso Suazo \email{rtasso@uchile.cl} (\href{https://orcid.org/0000-0002-1264-4933}{ORCID})
+  \item Rodolfo Tasso Suazo \email{rtasso@uchile.cl} (\href{https://orcid.org/0000-0002-1264-4933}{ORCID}) (Grupo de Ciencia de Datos para Salud Pública, Escuela de Salud Pública, U. de Chile)
 }
 
 Other contributors:
 \itemize{
-  \item Grupo de Ciencia de Datos para Salud Pública (Escuela de Salud Pública, U. de Chile) [contributor]
-  \item Centro FIC (DEIS / MINSAL Chile) [contributor]
+  \item Francisca Jofre (Package logo design) [contributor]
 }
 
 }
