### 2026-05-24 d40ac6d
docs: logo CDSP transparente y reposicion de seccion desarrollo + pie institucional en README
