+++ README.Rmd
@@ -187,6 +187,10 @@ Catálogo oficial **CIE-10 MINSAL/DEIS v2018**:
 - Fuente: [DEIS](https://deis.minsal.cl) — [Centro FIC Chile](https://deis.minsal.cl/centrofic/)
 - Dominio público según [Decreto 356/2017](https://www.bcn.cl/leychile/navegar?i=1112064)
 
+## Desarrollo
+
+Este paquete fue desarrollado con asistencia de Claude (Anthropic), con verificación y validación humana de todo el código y la documentación.
+
 ## Contribuir
 
 - Reportar bugs: <https://github.com/RodoTasso/ciecl/issues>
@@ -202,11 +206,23 @@ MIT + datos MINSAL de dominio público.
 Grupo de Ciencia de Datos para Salud Pública, Escuela de Salud Pública,
 Facultad de Medicina, Universidad de Chile.
 
+## Agradecimientos
+
+- Diseño del logotipo del paquete: [@fje1](https://github.com/fje1).
+
 ## Enlaces
 
 - Repositorio: <https://github.com/RodoTasso/ciecl>
 - DEIS MINSAL: <https://deis.minsal.cl>
 - API CIE-11: <https://icd.who.int/icdapi>
+
+---
+
+<img src=\"man/figures/logo-CDSP_color.png\" height=\"60\" alt=\"Grupo de Ciencia de Datos para Salud Pública\">
+
+**Grupo de Ciencia de Datos para Salud Pública**<br>
+Escuela de Salud Pública, Facultad de Medicina<br>
+Universidad de Chile
 ")
 } else {
 cat("## Data
@@ -215,6 +231,10 @@ Official **CIE-10 MINSAL/DEIS v2018** catalogue:
 
 - Source: [DEIS](https://deis.minsal.cl) — [Centro FIC Chile](https://deis.minsal.cl/centrofic/)
 - Public domain under [Decree 356/2017](https://www.bcn.cl/leychile/navegar?i=1112064)
+
+## Development
+
+This package was developed with assistance from Claude (Anthropic), with human verification and validation of all code and documentation.
 
 ## Contributing
 
@@ -231,11 +251,23 @@ MIT + MINSAL public-domain data.
 Data Science for Public Health Group, School of Public Health,
 Faculty of Medicine, University of Chile.
 
+## Acknowledgements
+
+- Package logo design: [@fje1](https://github.com/fje1).
+
 ## Links
 
 - Repository: <https://github.com/RodoTasso/ciecl>
 - DEIS MINSAL: <https://deis.minsal.cl>
 - ICD-11 API: <https://icd.who.int/icdapi>
+
+---
+
+<img src=\"man/figures/logo-CDSP_color.png\" height=\"60\" alt=\"Data Science for Public Health Group\">
+
+**Data Science for Public Health Group**<br>
+School of Public Health, Faculty of Medicine<br>
+University of Chile
 ")
 }
 ```
