### 2026-05-25 d4076ac
docs: logo CDSP transparente (pie a height 120), reponer seccion desarrollo + pie institucional y afiliacion en README/DESCRIPTION
