+++ README.Rmd
@@ -187,6 +187,10 @@ Catálogo oficial **CIE-10 MINSAL/DEIS v2018**:
 - Fuente: [DEIS](https://deis.minsal.cl) — [Centro FIC Chile](https://deis.minsal.cl/centrofic/)
 - Dominio público según [Decreto 356/2017](https://www.bcn.cl/leychile/navegar?i=1112064)
 
+## Desarrollo
+
+Este paquete fue desarrollado con asistencia de Claude (Anthropic), con verificación y validación humana de todo el código y la documentación.
+
 ## Contribuir
 
 - Reportar bugs: <https://github.com/RodoTasso/ciecl/issues>
@@ -211,6 +215,14 @@ Facultad de Medicina, Universidad de Chile.
 - Repositorio: <https://github.com/RodoTasso/ciecl>
 - DEIS MINSAL: <https://deis.minsal.cl>
 - API CIE-11: <https://icd.who.int/icdapi>
+
+---
+
+<img src=\"man/figures/logo-CDSP_color.png\" height=\"120\" alt=\"Grupo de Ciencia de Datos para Salud Pública\">
+
+**Grupo de Ciencia de Datos para Salud Pública**<br>
+Escuela de Salud Pública, Facultad de Medicina<br>
+Universidad de Chile
 ")
 } else {
 cat("## Data
@@ -219,6 +231,10 @@ Official **CIE-10 MINSAL/DEIS v2018** catalogue:
 
 - Source: [DEIS](https://deis.minsal.cl) — [Centro FIC Chile](https://deis.minsal.cl/centrofic/)
 - Public domain under [Decree 356/2017](https://www.bcn.cl/leychile/navegar?i=1112064)
+
+## Development
+
+This package was developed with assistance from Claude (Anthropic), with human verification and validation of all code and documentation.
 
 ## Contributing
 
@@ -244,6 +260,14 @@ Faculty of Medicine, University of Chile.
 - Repository: <https://github.com/RodoTasso/ciecl>
 - DEIS MINSAL: <https://deis.minsal.cl>
 - ICD-11 API: <https://icd.who.int/icdapi>
+
+---
+
+<img src=\"man/figures/logo-CDSP_color.png\" height=\"120\" alt=\"Data Science for Public Health Group\">
+
+**Data Science for Public Health Group**<br>
+School of Public Health, Faculty of Medicine<br>
+University of Chile
 ")
 }
 ```
