### 2026-05-25 8cbf1d0
docs: agrandar logo CDSP del pie (height 60->120)
