+++ authors.md
@@ -3,7 +3,9 @@
 ## Authors
 
 - **Rodolfo Tasso Suazo**. Author, maintainer.
-  [](https://orcid.org/0000-0002-1264-4933)
+  [](https://orcid.org/0000-0002-1264-4933)  
+  Grupo de Ciencia de Datos para Salud Pública, Escuela de Salud
+  Pública, U. de Chile
 
 - **Francisca Jofre**. Contributor.  
   Package logo design
