### 2026-05-25 9c1a54e
ci: serializar pkgcheck tras R-CMD-check y test-coverage para evitar falso negativo del check de CI
