+++ .github/workflows/pkgcheck.yaml
@@ -22,6 +22,39 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
+      # El check 'ci' de pkgcheck consulta via API el estado de los workflows
+      # 'cmd|coverage' del commit y exige que TODOS esten en 'success'. Si pkgcheck
+      # corre en paralelo, puede consultarlos cuando aun estan 'in_progress' y dar
+      # un falso negativo. Este paso lo serializa: espera a R-CMD-check.yaml y
+      # test-coverage del mismo commit antes de continuar. Sin dependencias externas.
+      - name: Esperar a R-CMD-check y test-coverage
+        env:
+          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          targets="R-CMD-check.yaml test-coverage"
+          end=$((SECONDS+1500))      # timeout global: 25 min
+          discover=$((SECONDS+180))  # si no aparecen runs en 3 min, no bloquear (p.ej. PR con SHA distinto)
+          while true; do
+            json=$(gh api "repos/${GITHUB_REPOSITORY}/actions/runs?head_sha=${GITHUB_SHA}&per_page=100" 2>/dev/null)
+            pending=""; anyseen=0
+            for t in $targets; do
+              row=$(printf '%s' "$json" | jq -r --arg n "$t" 'first(.workflow_runs[] | select(.name==$n)) | "\(.status // "none") \(.conclusion // "none")"' 2>/dev/null)
+              [ -z "$row" ] && row="none none"
+              status=${row%% *}; concl=${row##* }
+              if [ "$status" = "none" ]; then pending="$pending $t(sin-run)"; continue; fi
+              anyseen=1
+              if [ "$status" != "completed" ]; then
+                pending="$pending $t($status)"
+              elif [ "$concl" != "success" ]; then
+                echo "::error::$t concluyo en '$concl'; pkgcheck fallaria igual."; exit 1
+              fi
+            done
+            if [ -z "$pending" ]; then echo "OK: R-CMD-check y test-coverage en success."; break; fi
+            if [ "$anyseen" = 0 ] && [ "$SECONDS" -gt "$discover" ]; then echo "Aviso: sin runs target para este SHA; continuo."; break; fi
+            if [ "$SECONDS" -gt "$end" ]; then echo "Aviso: timeout esperando; continuo y dejo que pkgcheck evalue."; break; fi
+            echo "Esperando:$pending"; sleep 15
+          done
+
       - uses: ropensci-review-tools/pkgcheck-action@main
         with:
           # Desactivar post a Issues (requiere issues:write que no tenemos).
