### 2026-06-09 fc25be2
docs(f4): completar documentación cie10_cl y limpieza interna
- data.R + cie10_cl.Rd: documentar las 11 columnas (incluye uso_cl),
  tildes correctas, @family datasets, @seealso y ejemplos adicionales
- cie-search.R, cie-siglas.R: renombrar arg interno texto → text en
  normalizar_tildes() y expandir_sigla() (funciones no exportadas)
- cie-siglas.R, cie-table.R: lambdas de una línea migradas a \(x)
- NEWS.md: registrar sprint F4
