### 2026-06-10 e3c65c1
chore(github): reemplazar templates monolinguales por versiones bilingues ES/EN
