+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,38 +1,0 @@
----
-name: Bug report
-about: Reportar un error en ciecl / Report a bug in ciecl
-title: '[BUG] '
-labels: 'bug'
-assignees: ''
-
----
-
-**Descripcion del bug / Describe the bug**
-Descripcion clara y concisa del error.
-
-**Ejemplo reproducible / Reprex**
-Proporciona un ejemplo minimo reproducible ([reprex](https://reprex.tidyverse.org/)):
-
-```r
-library(ciecl)
-
-# Codigo que produce el error
-```
-
-**Comportamiento esperado / Expected behavior**
-Que esperabas que ocurriera.
-
-**Informacion del sistema / Session info**
-Pega la salida de:
-
-```r
-sessionInfo()
-packageVersion("ciecl")
-```
-
-- **SO / OS**: [ej: Windows 11, macOS 14, Ubuntu 22.04]
-- **Version de R / R version**: [ej: 4.4.1]
-- **Version de ciecl**: [ej: 0.9.3]
-
-**Contexto adicional / Additional context**
-Cualquier informacion adicional relevante.
