+++ .github/ISSUE_TEMPLATE/bug_report_en.md
@@ -1,0 +1,38 @@
+---
+name: "🐛 Report a bug (English)"
+about: Report a bug in ciecl
+title: '[BUG] '
+labels: 'bug'
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of the error.
+
+**Reproducible example (reprex)**
+Provide a minimal reproducible example ([reprex](https://reprex.tidyverse.org/)):
+
+```r
+library(ciecl)
+
+# Code that produces the error
+```
+
+**Expected behavior**
+What you expected to happen.
+
+**Session info**
+Paste the output of:
+
+```r
+sessionInfo()
+packageVersion("ciecl")
+```
+
+- **OS**: [e.g. Windows 11, macOS 14, Ubuntu 22.04]
+- **R version**: [e.g. 4.4.1]
+- **ciecl version**: [e.g. 0.9.3]
+
+**Additional context**
+Any other relevant information.
