+++ .github/ISSUE_TEMPLATE/question_en.md
@@ -1,0 +1,23 @@
+---
+name: "❓ Ask a question (English)"
+about: Ask a question about using ciecl
+title: '[QUESTION] '
+labels: 'question'
+assignees: ''
+
+---
+
+**Question**
+Clearly describe your question about using ciecl.
+
+**Context**
+Tell us what you are trying to achieve and in what context (type of data, workflow, etc.).
+
+**What you tried**
+Describe what you have already tried and, if applicable, include the code:
+
+```r
+library(ciecl)
+
+# Code you tried
+```
