+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,5 @@
+blank_issues_enabled: false
+contact_links:
+  - name: Consultas generales / General questions
+    url: https://github.com/RodoTasso/ciecl/discussions
+    about: Para dudas de uso usa Discussions o la plantilla Consulta/Question. / For usage questions use Discussions or the Question template.
