### 2026-06-18 a1b5aa3
ci: serializar push de codemeta y test-coverage con concurrency group
Ambos workflows hacen commit+push a main (codemeta.json y badges/coverage.json
respectivamente) en el mismo evento de push, causando una carrera donde el
segundo en terminar recibe 'rejected (fetch first)' y rompe en cascada el
gate de pkgcheck que espera a test-coverage en success.

Fix: concurrency group compartido por ref para que GitHub encole el segundo
job en vez de dejarlos correr en paralelo.
