+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   idiomas: idiomas.html
   instalacion: instalacion.html
   installation: installation.html
-last_built: 2026-06-18T15:52Z
+last_built: 2026-06-18T19:44Z
 urls:
   reference: https://rodotasso.github.io/ciecl/reference
   article: https://rodotasso.github.io/ciecl/articles
