### 2026-06-20 2150c65
docs: aplicar correcciones de documentacion de revision rOpenSci (yabellini)
Recupera trabajo preparado el 2026-05-29 que quedo guardado en stash
desde el 2026-06-09 sin commitear. Resuelve ybs01, ybs02, ybs03, ybs05,
ybs06 y ybs07 del informe de revision de documentacion.
