+++ CONTRIBUTING.md
@@ -11,6 +11,14 @@ If you find a bug, please open an [issue on GitHub](https://github.com/RodoTasso
 - Output of `sessionInfo()`
 - Your version of `ciecl` (`packageVersion("ciecl")`)
 
+## Types of issues
+
+When you open a new issue, GitHub will offer you three templates. Please choose the one that matches your situation:
+
+- **Bug report**: something is broken or behaves differently than documented. Include a reprex and your session info.
+- **Feature request**: you would like a new capability or an improvement to an existing one. Describe the problem it solves and the behaviour you expect.
+- **Question**: you have a doubt about how to use the package. Use the **Question** template, which applies the `question` label automatically.
+
 ## Proposing changes
 
 1. Fork the repository
@@ -27,9 +35,13 @@ If you find a bug, please open an [issue on GitHub](https://github.com/RodoTasso
 ## Code style
 
 - Follow the [tidyverse style guide](https://style.tidyverse.org/)
-- Use `%>%` from dplyr for pipes
-- Document functions with roxygen2
-- Encoding UTF-8, no accented characters in source code (use `\uXXXX`)
+- Use the native pipe `|>` for pipes (requires R >= 4.1.0)
+- **Language policy**:
+  - Exported functions and arguments: **English**.
+  - Documentation (roxygen2) and user guides: **Spanish** (primary) and English.
+  - Error and informative messages: **Spanish**.
+- Document functions with roxygen2 using markdown syntax
+- Encoding: keep R source code in ASCII, using `\uXXXX` escapes for accented characters; write documents and prose in UTF-8 with correct accents
 - Declare NSE variables in `R/globals.R`
 
 ## Tests
@@ -44,22 +56,30 @@ By participating in this project, you agree to abide by the [Code of Conduct](CO
 
 ## Questions
 
-For general questions, open an [issue](https://github.com/RodoTasso/ciecl/issues) with the "question" label.
+For general questions, open an [issue](https://github.com/RodoTasso/ciecl/issues) using the **Question** template, which applies the `question` label automatically.
 
 ---
 
 # Contribuir a ciecl
 
-Gracias por tu interes en contribuir a `ciecl`. Este documento describe como participar en el desarrollo del paquete.
+Gracias por tu interés en contribuir a `ciecl`. Este documento describe cómo participar en el desarrollo del paquete.
 
 ## Reportar bugs
 
 Si encuentras un bug, por favor abre un [issue en GitHub](https://github.com/RodoTasso/ciecl/issues) con:
 
-- Descripcion clara del problema
-- Codigo reproducible minimo ([reprex](https://reprex.tidyverse.org/))
-- Output de `sessionInfo()`
-- Version de `ciecl` (`packageVersion("ciecl")`)
+- Descripción clara del problema
+- Código reproducible mínimo ([reprex](https://reprex.tidyverse.org/))
+- Salida de `sessionInfo()`
+- Versión de `ciecl` (`packageVersion("ciecl")`)
+
+## Tipos de issue
+
+Al abrir un nuevo issue, GitHub te ofrecerá tres plantillas. Elige la que corresponda a tu caso:
+
+- **Reporte de error (Bug report)**: algo está roto o se comporta distinto a lo documentado. Incluye un reprex y la información de tu sesión.
+- **Sugerencia de mejora (Feature request)**: te gustaría una nueva funcionalidad o una mejora a una existente. Describe el problema que resuelve y el comportamiento que esperas.
[…]
