+++ CONTRIBUTING.md
@@ -14,6 +14,20 @@ GitHub](https://github.com/RodoTasso/ciecl/issues) with:
 - Output of [`sessionInfo()`](https://rdrr.io/r/utils/sessionInfo.html)
 - Your version of `ciecl` (`packageVersion("ciecl")`)
 
+## Types of issues
+
+When you open a new issue, GitHub will offer you three templates. Please
+choose the one that matches your situation:
+
+- **Bug report**: something is broken or behaves differently than
+  documented. Include a reprex and your session info.
+- **Feature request**: you would like a new capability or an improvement
+  to an existing one. Describe the problem it solves and the behaviour
+  you expect.
+- **Question**: you have a doubt about how to use the package. Use the
+  **Question** template, which applies the `question` label
+  automatically.
+
 ## Proposing changes
 
 1.  Fork the repository
@@ -38,9 +52,16 @@ GitHub](https://github.com/RodoTasso/ciecl/issues) with:
 ## Code style
 
 - Follow the [tidyverse style guide](https://style.tidyverse.org/)
-- Use `%>%` from dplyr for pipes
-- Document functions with roxygen2
-- Encoding UTF-8, no accented characters in source code (use `\uXXXX`)
+- Use the native pipe `|>` for pipes (requires R \>= 4.1.0)
+- **Language policy**:
+  - Exported functions and arguments: **English**.
+  - Documentation (roxygen2) and user guides: **Spanish** (primary) and
+    English.
+  - Error and informative messages: **Spanish**.
+- Document functions with roxygen2 using markdown syntax
+- Encoding: keep R source code in ASCII, using `\uXXXX` escapes for
+  accented characters; write documents and prose in UTF-8 with correct
+  accents
 - Declare NSE variables in `R/globals.R`
 
 ## Tests
@@ -57,25 +78,40 @@ Conduct](https://rodotasso.github.io/ciecl/CODE_OF_CONDUCT.md).
 ## Questions
 
 For general questions, open an
-[issue](https://github.com/RodoTasso/ciecl/issues) with the “question”
-label.
+[issue](https://github.com/RodoTasso/ciecl/issues) using the
+**Question** template, which applies the `question` label automatically.
 
 ------------------------------------------------------------------------
 
 # Contribuir a ciecl
 
-Gracias por tu interes en contribuir a `ciecl`. Este documento describe
-como participar en el desarrollo del paquete.
+Gracias por tu interés en contribuir a `ciecl`. Este documento describe
+cómo participar en el desarrollo del paquete.
 
 ## Reportar bugs
 
 Si encuentras un bug, por favor abre un [issue en
 GitHub](https://github.com/RodoTasso/ciecl/issues) con:
 
-- Descripcion clara del problema
-- Codigo reproducible minimo ([reprex](https://reprex.tidyverse.org/))
-- Output de [`sessionInfo()`](https://rdrr.io/r/utils/sessionInfo.html)
-- Version de `ciecl` (`packageVersion("ciecl")`)
+- Descripción clara del problema
+- Código reproducible mínimo ([reprex](https://reprex.tidyverse.org/))
+- Salida de [`sessionInfo()`](https://rdrr.io/r/utils/sessionInfo.html)
+- Versión de `ciecl` (`packageVersion("ciecl")`)
+
+## Tipos de issue
+
+Al abrir un nuevo issue, GitHub te ofrecerá tres plantillas. Elige la
+que corresponda a tu caso:
+
+- **Reporte de error (Bug report)**: algo está roto o se comporta
[…]
