### 2026-08-01 8c181f7
fix: correcciones revision rOpenSci (ybs08-36)
- ybs08/ybs09/ybs36: documentacion de uso_cl en cie10_cl y columna
  uso_cl en el tibble vacio interno
- ybs10-ybs17: cie10_sql con validacion de query, errores relanzados
  como clase ciecl_sql_error, terminologia de mensajes en castellano
  (ybs14) y ayudas de cie10_clear_cache/cie10_disconnect ampliadas
- ybs19-ybs30: errores de argumento obligatorio en castellano via
  check_required_es, conector 'y' en listados cli, tildes en la ayuda
  (incl. cie_siglas ybs25, cie_short ybs26, cie_search ybs27)
- ybs31: cie_guide enlazada desde README y @see
[…]
