+++ R/cie-lookup.R
@@ -33,32 +33,32 @@ extract_cie_from_text <- function(text) {
   return(resultado)
 }
 
-#' Busqueda exacta por codigo CIE-10
+#' Búsqueda exacta por código CIE-10
 #'
-#' @param code Character vector de codigos
+#' @param code Character vector de códigos
 #'   (ej. "E11", "E11.0", c("E11.0", "Z00"))
 #'   o rango (ej. "E10-E14"). Acepta vectores.
 #'   Soporta formatos: con punto (E11.0),
-#'   sin punto (E110), o solo categoria (E11).
-#' @param expand Logical, expandir jerarquia completa (default FALSE)
-#' @param normalize Logical, normalizar formato de codigos
-#'   automaticamente (default TRUE)
+#'   sin punto (E110), o solo categoría (E11).
+#' @param expand Logical, expandir jerarquía completa (default FALSE)
+#' @param normalize Logical, normalizar formato de códigos
+#'   automáticamente (default TRUE)
 #' @param full_description Logical, agregar columna `descripcion_completa`
 #'   con formato "CODIGO - DESCRIPCION" (default FALSE)
-#' @param extract Logical, extraer codigo CIE-10 de texto con
+#' @param extract Logical, extraer código CIE-10 de texto con
 #'   prefijos/sufijos (default FALSE).
-#'   IMPORTANTE: Solo usar con codigo ESCALAR (longitud 1).
+#'   IMPORTANTE: Solo usar con código ESCALAR (longitud 1).
 #'   Ejemplo: "CIE:E11.0" -> "E11.0", "E11.0-confirmado" -> "E11.0".
-#'   Para vectores multiples usar extract=FALSE (default).
-#' @param check_siglas Logical, buscar siglas medicas comunes (default FALSE).
+#'   Para vectores múltiples usar extract=FALSE (default).
+#' @param check_siglas Logical, buscar siglas médicas comunes (default FALSE).
 #'   Ejemplo: "IAM" -> I21.0 (Infarto agudo miocardio)
 #' @param include_uso_cl Logical, incluir columna `uso_cl` en el output
 #'   (default TRUE). El default difiere de [cie_search()] (FALSE) para
-#'   preservar el contrato historico de cada funcion. Valores posibles:
+#'   preservar el contrato histórico de cada función. Valores posibles:
 #'   `"principal"`, `"legado"`, `"causa_externa"`, `"etiologico"`,
 #'   `"causa_externa | principal"`.
-#' @param only_uso_cl Logical, filtrar a codigos vigentes de uso clinico
-#'   en Chile (default FALSE). Cuando es TRUE, excluye los codigos con
+#' @param only_uso_cl Logical, filtrar a códigos vigentes de uso clínico
+#'   en Chile (default FALSE). Cuando es TRUE, excluye los códigos con
 #'   `uso_cl == "legado"` (es decir, conserva `principal`,
 #'   `causa_externa`, `etiologico` y sus combinaciones).
 #' @param codigo `r lifecycle::badge("deprecated")` Use `code`.
@@ -67,27 +67,27 @@ extract_cie_from_text <- function(text) {
 #' @param descripcion_completa `r lifecycle::badge("deprecated")` Use `full_description`.
 #' @returns tibble con codigo(s) matcheado(s)
 #' @family search
-#' @seealso [cie_search()], [cie_norm()], [cie_expand()]
+#' @seealso [cie_search()], [cie_norm()], [cie_expand()], [cie_guide()]
 #' @export
 #' @examples
-#' # Busqueda directa por codigo
+#' # Búsqueda directa por código
[…]
