### 2026-08-10 bc102cf
fix footer
https://pkgdown.r-lib.org/articles/customise.html#footer
