### 2026-08-17 eb3ab6e
feat: implementa revision rOpenSci ms01-ms20 y corrige workflow pkgcheck
