+++ tests/testthat/test-cie-sql.R
@@ -182,7 +182,7 @@ test_that("get_cie10_db tabla tiene indices", {
 test_that("get_cie10_db usa directorio cache correcto", {
   skip_on_cran()
 
-  cache_dir <- ciecl:::get_cache_dir()
+  cache_dir <- get_cache_dir()
   db_path <- file.path(cache_dir, "cie10.db")
 
   con <- get_cie10_db()
@@ -205,7 +205,7 @@ test_that("get_cie10_db tabla tiene columnas esperadas", {
 test_that("cie10_clear_cache elimina archivo db", {
   skip_on_cran()
 
-  cache_dir <- ciecl:::get_cache_dir()
+  cache_dir <- get_cache_dir()
   db_path <- file.path(cache_dir, "cie10.db")
 
   # Asegurar que existe
@@ -337,7 +337,7 @@ test_that("get_cie10_db crea directorio cache si no existe", {
   # Limpiar cache para forzar recreacion
   suppressMessages(cie10_clear_cache())
 
-  cache_dir <- ciecl:::get_cache_dir()
+  cache_dir <- get_cache_dir()
 
   # Conectar - debe crear directorio si no existe
   get_cie10_db()
@@ -487,7 +487,7 @@ test_that("build_cache_atomic crea cache completo", {
   # Limpiar
   suppressMessages(cie10_clear_cache())
 
-  cache_dir <- ciecl:::get_cache_dir()
+  cache_dir <- get_cache_dir()
   db_path <- file.path(cache_dir, "cie10.db")
 
   # No debe existir .tmp residual despues de build exitoso
@@ -505,7 +505,7 @@ test_that("build_cache_atomic crea cache completo", {
 
 test_that(".ciecl_env tiene estructura correcta", {
   env <- .ciecl_env
-  expect_true(is.environment(env))
+  expect_type(env, "environment")
   expect_true("con" %in% ls(env, all.names = TRUE))
   expect_true("db_path" %in% ls(env, all.names = TRUE))
 })
@@ -584,7 +584,7 @@ test_that("get_cie10_db reconstruye si tabla cie10 falta", {
   skip_on_cran()
   ciecl::cie10_disconnect()
   con_direct <- DBI::dbConnect(RSQLite::SQLite(),
-    file.path(ciecl:::get_cache_dir(), "cie10.db"))
+    file.path(get_cache_dir(), "cie10.db"))
   on.exit(
     if (DBI::dbIsValid(con_direct)) suppressWarnings(DBI::dbDisconnect(con_direct)),
     add = TRUE
@@ -654,7 +654,7 @@ test_that("build_cache_atomic limpia .tmp residual", {
   skip_on_cran()
   withr::defer(ciecl::cie10_disconnect())
   ciecl::cie10_disconnect()
-  cache_dir <- ciecl:::get_cache_dir()
+  cache_dir <- get_cache_dir()
   tmp_file <- file.path(cache_dir, "cie10.db.tmp")
   if (!dir.exists(cache_dir)) dir.create(cache_dir, recursive = TRUE)
   file.create(tmp_file)
