+++ R/cie-api.R
@@ -1,14 +1,30 @@
 #' Buscar códigos CIE-11 vía API OMS
 #'
 #' @param text String término búsqueda español/inglés
-#' @param api_key String opcional, Client ID + Secret OMS separados ":"
-#'   Obtener en: https://icd.who.int/icdapi
+#' @param api_key String opcional, Client ID + Secret OMS separados ":".
+#'   Por defecto se lee desde la variable de entorno `ICD_API_KEY` vía
+#'   [get_icd_api_key()]. Obtener credenciales en: https://icd.who.int/icdapi
 #' @param lang Character, idioma respuesta ("es" o "en")
 #' @param max_results Integer, máximo resultados (default 10)
 #' @param release Character, versión de release CIE-11 a consultar
 #'   (default "2024-01"). Ver releases disponibles en la API OMS.
 #' @param texto `r lifecycle::badge("deprecated")` Use `text`.
 #' @returns tibble con códigos CIE-11 + títulos o vacío si error
+#' @details
+#' ## Seguridad de la API key
+#'
+#' NUNCA escribas la llave literal en tus scripts (p. ej.
+#' `api_key = "tu_id:tu_secret"`): si compartes el código (repositorios,
+#' correos, capturas) expondrías tus credenciales de forma accidental.
+#'
+#' La vía recomendada es guardar la llave en la variable de entorno
+#' `ICD_API_KEY`, por ejemplo editando `~/.Renviron` con
+#' `usethis::edit_r_environ()`; [get_icd_api_key()] la lee
+#' automáticamente.
+#'
+#' El argumento `api_key` existe solo para circunstancias excepcionales
+#' (p. ej. manejar múltiples llaves, o entornos donde no es posible fijar
+#' una variable de entorno).
 #' @family api_who
 #' @seealso [cie_search()], [cie_lookup()], [cie_guide()]
 #' @export
@@ -18,12 +34,23 @@
 #' # Ver parámetros disponibles
 #' args(cie11_search)
 #'
-#' @examplesIf interactive()
+#' \dontshow{
+#' # El cassette grabado (inst/_vcr/) reproduce la respuesta de la API sin
+#' # conexión ni credenciales reales; la llave ficticia solo satisface el
+#' # parseo interno del argumento api_key y se restaura al final
+#' .key_prev <- Sys.getenv("ICD_API_KEY", unset = NA)
+#' if (is.na(.key_prev)) Sys.setenv(ICD_API_KEY = "client_id:client_secret")
+#' vcr::insert_example_cassette("cie11_search", package = "ciecl",
+#'                              match_requests_on = c("method", "uri"))
+#' }
 #' # Requiere credenciales OMS gratuitas (https://icd.who.int/icdapi)
-#' withr::with_envvar(c(ICD_API_KEY = "client_id:client_secret"), {
-#'   cie11_search("depresion mayor")
-#' })
-cie11_search <- function(text, api_key = NULL, lang = c("es", "en"),
+#' cie11_search("depresion mayor")
+#' \dontshow{
+#' vcr::eject_cassette()
+#' if (is.na(.key_prev)) Sys.unsetenv("ICD_API_KEY")
+#' }
+cie11_search <- function(text, api_key = get_icd_api_key(),
+                         lang = c("es", "en"),
                          max_results = 10, release = "2024-01",
                          texto = lifecycle::deprecated()) {
   # Deprecation: argumento en espanol -> ingles
@@ -46,11 +73,14 @@ cie11_search <- function(text, api_key = NULL, lang = c("es", "en"),
       class = "ciecl_invalid_input"
     )
   }
-  if (nchar(trimws(text)) == 0) {
[…]
