### 2026-08-17 77a0470
fix: piso testthat 3.3.0, guard vcr en ejemplos y simetria de docs ES/EN
- DESCRIPTION: testthat (>= 3.3.0) por expect_all_false en tests
- cie-api.R: identificadores en ingles, guard requireNamespace("vcr")
  en ejemplos dontshow, tildes en comentarios
- CONTRIBUTING.md: seccion Tests espejo ES/EN (testthat >= 3.3.0,
  withr, expect_snapshot)
- CODE_OF_CONDUCT.md: tildes UTF-8 restauradas en seccion ES
- vignettes: instalacion/installation restauradas (dependencias,
  requisitos, verificacion API, cache, troubleshooting) espejo 1:1;
  languages.Rmd nueva (par EN de idiomas.Rmd);
  case-study-
[…]
