### 2026-08-18 639efe2
fix: bloquear load_extension en cie10_sql y deshabilitar extensiones SQLite
El guard de cie10_sql() no cubria load_extension() y la conexion pooled
se abria con loadable.extensions = TRUE (default RSQLite). Se cierra el
hueco a nivel de driver (loadable.extensions = FALSE en las 3 conexiones
del pool) y se agrega el keyword al guard (defensa en profundidad).
Test nuevo verifica el abort con clase ciecl_unsafe_query.
Hallazgo Medium de auditoria de seguridad 2026-08-18.
