### 2026-08-18 #765 @Rodotasso
Hola @maelle y @yabellini!, muchas gracias por sus correcciones y sugerencias. Dejo aqui los cambios realizados a partir de ellos.
Son un poco extensos, así que iré poco a poco. 😸
(Lamento la demora, gh y su unicornio no me dejaban postear ayer jaja)..

## Parto con las correcciones de Yabellini (Sé su nombre pero no figura aquí, así que no lo uso).
Comienzo desde ybs08 en adelante

### Conjunto de datos `cie10_cl`
<summary> (ybs08, ybs09, ybs36)</summary>
- **ybs08:** columna `uso_cl` documentada como lógica siendo `character` categórica. Se corrigió el bloque `@format` en `R/data.R` (`"principal"`, `"legado"`, `"etiologico"`, `"causa_externa"`, `"causa_externa | principal"`).
- **ybs09:** ejemplo `subset(cie10_cl, uso_cl == TRUE)` devolvía 0 filas. Reemplazado por `uso_cl != "legado"`, consistente con el filtro interno `only_uso_cl`.
- **ybs36:** tibble vacío de respaldo sin columna `uso_cl`. Agregada como `character(0)` en `cie10_empty_tibble()`.
- Debo agregar más documentación que no toque el código sobre `uso_cl`, es algo nuevo y es para aclarar manejo de retrocompatibilidad de años y las listas tabulares de CIE-10 en Chile.

### Validación de parámetros obligatorios
<summary> (ybs19, ybs22, ybs24, ybs30, ybs27)</summary>
- **ybs19, ybs22, ybs24, ybs30:** mensajes en inglés de `rlang`/R base al omitir argumentos obligatorios. Helper interno `check_required_es()` en `R/globals.R` (lanza `cli::cli_abort()` en castellano con clase `ciecl_invalid_input`), aplicado en `cie_norm()`, `cie_validate_vector()`, `cie_expand()`, `cie_table()` y `cie_lookup()`. Detalle técnico: `rlang::check_required()` no se dispara con defaults `lifecycle::deprecated()`; el helper usa `missing()`.
- **ybs27:** conector "and" → "y" en `cie_short()` + error en inglés al faltar `text` en `cie_search()`. Resuelta: el conector castellano (`cli_vec(..., "vec-last" = " y ")`) quedó en `cie_short()`, y `cie_search()` ahora usa el helper `check_required_es()` (`R/cie-search.R:123`), que lanza `cli::cli_abort()` en castellano con clase `ciecl_invalid_input`. `NEWS.md` declara exactamente esto (conector en `cie_short()` + `check_required_es()` en `cie_search()`), así que el texto y el código ahora coinciden.

### `cie10_sql()`
<summary> (ybs10–ybs17)</summary>
- **ybs12, ybs13, ybs16:** documentación insuficiente. Descripción ampliada (SQL de solo lectura sobre el dataset embebido), instrucciones permitidas documentadas, `@seealso [cie10_cl]`.
- **ybs14:** terminología "query/keyword/statements" sin traducir en mensajes. Resuelta: los cuatro mensajes de `cie10_sql()` quedaron en castellano correcto ("Solo consultas `SELECT` permitidas" en `R/cie-sql.R:328`, "La consulta contiene una palabra clave no permitida" en :347, "Múltiples sentencias SQL no permitidas" con escape `ú` en :362 y "Error al ejecutar la consulta SQL" en :371), y el `@returns` (:294) dice "tibble con el resultado de la consulta". Snapshots y tests que fijaban los textos antiguos ya reflejan los nuevos (`tests/testthat/_snaps/robustness.md`, `tests/testthat/test-cie-sql.R:265`).
- **ybs15:** errores de `DBI::dbGetQuery()` no capturados. `tryCatch()` relanzando con clase `ciecl_sql_error` y el error original como `parent`.
- **ybs17:** validación faltante de `query`. Validación de tipo `character`, longitud 1 y no `NA` antes de ejecutar.
- **ybs10, ybs11:** ayudas de caché escuetas. Ampliadas en `cie10_clear_cache()` y `cie10_disconnect()` (qué es la caché SQLite, cuándo reconstruir, bloqueo del `.db`). Como nota menor que sigue vigente, queda spanglish residual en sus `@returns` ("No return value, called for side effects…"), para corregir si se toca el archivo.

### Ortografía de la ayuda (tildes)
<summary> (ybs18, ybs20, ybs21, ybs23, ybs25, ybs26, ybs29, ybs32, ybs33, ybs35)</summary>
- **ybs18, ybs20, ybs21, ybs23, ybs26, ybs29, ybs32, ybs33, ybs35:** tildes faltantes en bloques roxygen. Corregidas en `R/cie-utils.R`, `cie-search.R`, `cie-lookup.R`, `cie-table.R`, `cie-siglas.R`, `cie-guide.R`, `cie-describe.R`, `cie-api.R` (+ `.Rd` regenerados). En mensajes ejecutables, tildes vía `\uXXXX` (método sugerido por la propia revisora con `stringi::stri_escape_unicode()`).
- **ybs25:** tildes en `cie_siglas()` (deprecated pero exportada). Resuelta: "Listar siglas médicas (deprecated)" (`R/cie-siglas.R:522`) y "filtrar por categoría" (:526) corregidos en roxygen UTF-8, con `man/cie_siglas.Rd` regenerado y sincronizado. La atribución en `NEWS.md` también quedó corregida: ybs25 corresponde a `cie_siglas()` y ybs26 a `cie_short()`/`cie_search()`.
- Los escapes acá no los usaba solo porque no me gustaban jaja,. Y aún queda trabajo para hacer un buen match de siglas y búsquedas 😄

### Mensajes en un solo idioma
<summary> (ybs20, ybs27)</summary>
- **ybs20 (conector), ybs27 (conector):** *spanglish* del conector final de `cli`. Forzado `cli::cli_vec(x, style = list("vec-last" = " y "))` en `cie_short()` y `cie_validate_vector()` (en esta última también `"DB"` → `"base de datos"` y tilde en `"Códigos"`).

### `cie_map_comorbid()` y `cie_guide()`
<summary> (ybs31, ybs34)</summary>
- **ybs34:** entradas sin formato CIE-10 clasificadas como "Otra" sin aviso. Nuevo `cli::cli_warn()` con conteo de entradas malformadas (sin cambiar la clasificación de códigos válidos no mapeados).
- **ybs31:** `cie_guide()` poco visible. Enlace cruzado bilingüe desde el README y listada en `_pkgdown.yml`. Nota: la auditoría sugiere además `@seealso` a `cie_guide()` desde otras funciones (opcional, evaluar).

## Acá comienzan las respuestas a lo solicitado por Maëlle.

### CONTRIBUTING.md
<summary>Detalle de los cambios (ms01)</summary>
**ms01.** *"¿Por qué no poner primero la versión en español del texto y después la versión en inglés?"* Se reordenó `CONTRIBUTING.md` dejando primero la versión en español y después la inglesa, coherente con el español como idioma principal del paquete. El 2026-08-17 la sección Tests quedó además espejo ES/EN, con 5 bullets idénticos en ambos idiomas (incluyen `testthat (>= 3.3.0)` con edición 3, withr y `expect_snapshot()`).

### Documentación / sitio pkgdown

<summary> (ms02, ms03, ms08, ms09, ms10, ms11, ms12, ms14, ms18, ms19)</summary>

**ms02.** *"El enlace 'English' lleva al README en GitHub; mejor usar README.en.html del sitio."* El ítem "English" del navbar en `_pkgdown.yml` ahora apunta a `README.en.html` dentro del propio sitio, en vez de salir a GitHub. El 2026-08-17, `_pkgdown.yml` quedó además con navbar y articles espejo ES/EN (4+4), incorporando la nueva viñeta `languages.Rmd` (par EN de `idiomas.Rmd`).

**ms03.** *"Agrupar primero las cosas de código/datos y después las funcionalidades de transformación; consistencia al nombrar funciones (todas o ninguna)."* Se reordenó el párrafo de características en `README.Rmd` (primero catálogo oficial, siglas médicas y API CIE-11; después validación, búsqueda difusa, comorbilidades, SQL y expansión jerárquica) y se uniformó la mención de funciones. Se regeneraron `README.md` y `README.en.md`.

**ms08.** *"En el índice de referencia del sitio faltan tildes."* Barrido de tildes en los `@title` roxygen que alimentan el índice de referencia y en los títulos de sección de `_pkgdown.yml`. Índice de referencia verificado tras rebuild de pkgdown.

**ms09.** *"El footer es un poco raro; hice una PR: Rodotasso/ciecl#15."* El 2026-08-17 se mergeó la PR #15 en `main` (04660d5) y el fix se aplicó también en la rama de trabajo `dev` (`_pkgdown.yml`: `developed_by` pasa de bloque anidado a escalar, "Desarrollado por Rodolfo Tasso Suazo"). Gracias por la contribución directa. 😄

**ms10.** *"¿Por qué el enlace 'Inicio' apunta a index.html en vez de a una viñeta de introducción?"* Resuelto junto con ms18: la nueva viñeta de introducción quedó expuesta como entrada principal del sitio (mecanismo "Para empezar" de pkgdown con la viñeta principal del paquete), de modo que la navegación ya no aterriza en `index.html`.

**ms11.** *"La 'Guía de inicio' solo trata instalación; esperaba un caso de uso para principiantes."* Absorbido por ms18: la guía de inicio ahora presenta un caso de uso real con progresión de complejidad (ver ms18).

**ms12.** *"En la viñeta de instalación se puede conservar solo pak; dudo de su utilidad y es difícil de mantener."* Se simplificó la viñeta de instalación dejando **pak** como único método recomendado, eliminando las alternativas que encarecían el mantenimiento. La instalación queda también cubierta en el README y en la viñeta principal. El 2026-08-17 se restauraron además las secciones informativas de `instalacion.Rmd` / `installation.Rmd` (dependencias por funcionalidad, requisitos de sistema, verificación de la API, caché y troubleshooting) como espejo 1:1 ES/EN, manteniendo pak como método recomendado.

**ms14.** *"La mención de IA debería estar en CONTRIBUTING.md, con más detalle sobre qué considerar si alguien hace una PR con Claude."* Al 2026-08-17, la mención de IA vive solo en `CONTRIBUTING.md` (los READMEs no contienen declaración; `README.en.md` enlaza a CONTRIBUTING). Se reemplazó la oración inicial por una sección propia, **"Declaración de asistencia con IA"**, con el texto canónico del autor: herramientas usadas (Claude, Kimi, Gemini; web y CLI, mayo–agosto 2026), alcance (documentación del paquete y scripts de benchmark), revisión y responsabilidad humana completa, y la nota de que ninguna IA generó, recopiló ni analizó los datos del paquete (pipelines deterministas y reproducibles). Se conserva además la política para contribuciones externas asistidas por IA (es/en).
(Esto me encantó y me gusta el espíritu que refleja 🫶 ).

**ms18.** *"Este es mi comentario más importante… empezar con los datos, decir quién recibiría estos datos y con qué objetivo; mostrar también casos sin resultados."* Cerrada el 2026-08-17. Se reescribió la documentación de caso de uso como guía de inicio con una narrativa de usuario concreta (persona responsable de un reporte en un hospital), partiendo de los datos y su objetivo, e incluyendo búsquedas sin resultados y cómo interpretarlas. El caso de egresos quedó unificado como par ES/EN 1:1: `ciecl.Rmd` (ES, viñeta principal de get-started) ↔ `case-study-discharges.Rmd` (EN, reescrita como espejo). Es importante decir que `ciecl-es.Rmd` / `ciecl-en.Rmd` siguen existiendo como par de **introducción** (decisión intencional; ambas listadas en `_pkgdown.yml`), por lo que la simetría ES/EN del sitio no se reduce a un solo par.

**ms19.** *"Los ejemplos de cie11_search() con withr::with_envvar(...) no funcionan; sugiero usar vcr para grabar respuestas en ejemplos (vcr::insert_example_cassette())."* Implementado el 2026-08-17 adoptando el patrón oficial de vcr 2.x (que soporta httr2, el cliente HTTP de ciecl):

1. `vcr (>= 2.0.0)` agregado a `Suggests` en DESCRIPTION.
2. Ejemplos reescritos en `R/cie-api.R`: se eliminó el bloque `@examplesIf interactive()` con `withr::with_envvar()` y ahora el ejemplo real `cie11_search("depresion mayor")` corre siempre, envuelto en `\dontshow{}` con `vcr::insert_example_cassette("cie11_search", package = "ciecl", match_requests_on = c("method", "uri"))` y `vcr::eject_cassette()` al final. El `\dontshow` también fija una llave ficticia (`client_id:client_secret`) solo si `ICD_API_KEY` no existe (hace falta porque el argumento `api_key = get_icd_api_key()` se evalúa al llamar la función) y la restaura después. Además, el bloque `\dontshow` incluye un guard `requireNamespace("vcr")` para que los ejemplos no fallen si el paquete no está instalado.
3. Cassette grabado contra la API real de la OMS (una sola corrida con las credenciales del `.Renviron` local) en `inst/_vcr/cie11_search.yml`. Al grabarse con `match_requests_on = c("method", "uri")`, ni los headers ni el body de la petición quedan en el YAML, por lo que las credenciales nunca se persisten. Se verificó con grep que la llave completa, el `client_id` y el `client_secret` no aparecen en el cassette.
4. Como medida adicional, `tests/testthat/setup-vcr.R` (convención oficial) configura `vcr_configure(filter_sensitive_data = list("<ICD_API_KEY>" = Sys.getenv("ICD_API_KEY")))`, de modo que cualquier cassette futuro de tests reemplace la llave real por el placeholder.
5. Verificación offline: `devtools::run_examples()` corre todos los ejemplos sin internet ni llave (record mode `"none"` fuera de desarrollo, garantizado por `insert_example_cassette()`); suite completa FAIL 0 / PASS 1109. En CI los ejemplos ya no necesitan credenciales, y los tests contra la API real (`test-cie-api.R`) mantienen su `skip_if()` sin `ICD_API_KEY`, sin cambios. 📼 👍

### Código (detectados con el linter Jarl)
<summary> (ms04, ms05, ms15, ms16, ms17, ms20)</summary>
**ms04.** *"Usar anyNA() en vez de any(is.na(...)): test-cie-describe.R:17 y :77."* Corregido en `tests/testthat/test-cie-describe.R` (ambas ocurrencias) con el autofix de Jarl.

**ms05.** *"Simplificar uso_cl == TRUE en R/data.R (redundant_equals)."* Ya resuelto en el trabajo previo (ybs09): el ejemplo pasó a `subset(cie10_cl, uso_cl != "legado")`. Verificado que no queda ningún `== TRUE` residual en `R/data.R`.

**ms15.** *"Usar nzchar(trimws(text)) en vez de nchar(trimws(text)) == 0 en R/cie-api.R."* Corregido en `R/cie-api.R`.

**ms16.** *"La condición compuesta de max_results sería más legible con explaining variables."* Refactorizada la validación de `max_results` en `R/cie-api.R` extrayendo variables explicativas (tipo, escalaridad, positividad, enteridad), manteniendo el mismo mensaje de error. El 2026-08-17 los identificadores de `R/cie-api.R` pasaron además a inglés ASCII, siguiendo la convención del paquete (código en inglés, comentarios en español).

**ms17.** *"Poner api_key como argumento me parece peligroso: debería ser solo variable de entorno."* Resuelto:
Se planteó en el issue el patrón recomendado por la guía oficial de httr2 ("Wrapping APIs", sección "User-supplied key", `httr2/vignettes/articles/wrapping-apis.Rmd:479-498`): `api_key = get_api_key()`, donde el helper lee la variable de entorno y aborta con mensaje orientativo si falta, conservando el argumento "only ... in exceptional circumstances".

Implementado:

1. Firma alineada al idiom httr2: `api_key = get_icd_api_key()` en `cie11_search()`. El helper `get_icd_api_key()` (exportado) lee `ICD_API_KEY` en el momento de la llamada y aborta con el mismo error informativo de antes (clase `ciecl_api_error`), atribuido a la función llamante vía `rlang::caller_env()`. Comportamiento externo idéntico para quien usa la env var.
2. Riesgo documentado, tal como se solicitó: nueva subsección "Seguridad de la API key" en `@details` de `cie11_search()`. Indica no escribir nunca la llave literal en scripts (riesgo de exposición al compartir código o vía `.Rhistory`), recomienda la vía `ICD_API_KEY` con `usethis::edit_r_environ()` y reserva el argumento para circunstancias excepcionales (múltiples llaves, entornos sin env vars). La misma nota se agregó en las viñetas de instalación (`instalacion.Rmd` / `installation.Rmd`, es/en).
3. Deprecación descartada de común acuerdo: el argumento se conserva.

Verificación: tests de API FAIL 0 / PASS 77 (snapshot del error de llave faltante intacto); `man/cie11_search.Rd` regenerado con la nueva sección; `get_icd_api_key` agregada al índice de referencia de `_pkgdown.yml`.

**ms20.** *"En mensajes cli faltan tildes; hay que revisar todo el texto del paquete, no solo la documentación."* Barrido completo de mensajes `cli_*`/`message`/`warning`/`stop` en `R/`; las tildes van vía escapes `\uXXXX` según la convención del paquete (0 warnings de caracteres no-ASCII en R CMD check). Completado el 2026-08-17 con la restauración de tildes UTF-8 en la sección ES de `CODE_OF_CONDUCT.md` y en los comentarios de `R/cie-api.R` y `pkgcheck.yaml`.

### Tests

<summary> (punto sin número sobre :::, ms06, ms07)</summary>

**Punto sin número (`:::`).** *"No hace falta ::: para llamar funciones internas en los tests, porque testthat carga el paquete (9 casos)."* Eliminado el prefijo `ciecl:::` en los 9 lugares: `test-cie-sql.R` (6 ocurrencias de `get_cache_dir()`), `test-robustness.R` (2) y `test-utils-internal.R` (`cie_lookup_single`).

**ms06.** *"Usar expect_all_false() en vez de expect_true(all(!resultado)): test-edge-cases.R:321."* Corregido en `tests/testthat/test-edge-cases.R`. Como consecuencia, DESCRIPTION fija ahora `testthat (>= 3.3.0)` como piso, requerido por `expect_all_false()`.

**ms07.** *"8 expectativas más específicas (expect_named / expect_length / expect_type) fixables con jarl --fix."* Aplicadas las 8 correcciones: `expect_named` en `test-cie-comorbid.R`, `test-comorbid-validation.R` y `test-robustness.R`; `expect_length` en `test-cie-search.R` (2) y `test-integration.R` (2); `expect_type` en `test-cie-sql.R`.

Me gustó mucho como funciona Jarl, y de pasada revisé un poco el trabajo de Etienne (es impresionante), aún no he tenido tiempo de contrastar mis auditorías de código vs jarl. Las mías solo se basan en la práctica más actual de lo que estoy usando, pero Jarl ya paso a ser parte de mi trabajo 😃.

Muchas gracias a ambas por sus comentarios y correcciones, cuando quieran podemos continuar 🫶 .
