### 2026-08-18 #765 @yabellini
Hola @Rodotasso que bueno que ya pudiste empezar a trabajar en los comentarios de revision.

Me cuesta muchísimo seguir tus comentarios a mis comentarios porque no tienen formato, así que son muy pero muy difíciles de leer. ¿Podés arreglarlos?

Seria genial ademas que agregues un link al commit donde lo implementaste asi no tenemos que estar buscando todo "a mano" en todo el paquete.  Eso nos va a ahorrar un montón de tiempo. Gracias!
