### 2026-08-19 d5035f5
ci: regenerar codemeta.json desde DESCRIPTION [skip ci]
