### 2026-08-19 #765 @Rodotasso
Hola @yabellini, gracias por el aviso y si este formato es más cómodo aprovecharé de usarlo para las entregas posteriores.
Mis estructuras para ahorrar tiempo a veces son un desorden jaja.

Dejaré las respuestas un ítem por comentario (ybs08 a ybs36), con un resumen de tu punto, qué se implementó y el link al commit donde quedó el cambio. Todos los commits referenciados ya están en `main`.
No usaré los "###" ya que ocupan mucho espacio y no puedo controlar su tamaño jaja.
(Odio no conocer los keybindings o apretar de más, lamento si se recibió alguna notificación mientras estructuraba el mensaje)

ybs08 — Columna `uso_cl` documentada como binaria

<summarize></summarize>

**Comentario:** `uso_cl` estaba documentada como lógica, pero es character con categorías ("legado", "principal", "etiologico", "causa_externa", "causa_externa | principal") ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se corrigió el bloque `@format` en `R/data.R` documentando `uso_cl` como character categórica con sus cinco valores, y se regeneró `man/cie10_cl.Rd`.

**Ubicación:** [R/data.R#L31](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/data.R#L31)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs09 — Ejemplo `subset(cie10_cl, uso_cl == TRUE)` devuelve 0 filas
<summarize></summarize>

**Comentario:** el ejemplo "Códigos con uso clínico registrado" devolvía un conjunto vacío porque comparaba texto contra `TRUE` ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** el ejemplo pasó a `subset(cie10_cl, uso_cl != "legado") |> nrow()`, que devuelve datos reales y es consistente con el filtro interno `only_uso_cl` que usan `cie_lookup()` y `cie_search()`.

**Ubicación:** [R/data.R#L45](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/data.R#L45)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs10 — Ayuda escueta de `cie10_clear_cache()`
<summarize></summarize>

**Comentario:** no quedaba claro si la función era para el usuario o interna; faltaba explicar qué es la caché, qué significa forzar rebuild y cuándo es necesario, y había spanglish ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se reescribió la ayuda de `cie10_clear_cache()` explicando qué es la caché SQLite, cuándo conviene forzar la reconstrucción y por qué, con ejemplos explicados y en castellano.

**Ubicación:** [R/cie-sql.R#L381-L407](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-sql.R#L381-L407)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs11 — Ayuda de `cie10_disconnect()`
<summarize></summarize>

**Comentario:** mismo problema que la anterior; faltaba explicar qué es el bloqueo del archivo, por qué ocurre y qué pasa si no se libera, y el comentario "(Ejemplo omitido por usar internal environment)" no ayudaba ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se amplió la ayuda de `cie10_disconnect()` explicando el bloqueo (lock) del archivo `.db`, por qué puede ocurrir, qué pasa si no se libera y cuándo usar la función. Se reemplazó el comentario poco útil por una explicación real.

**Ubicación:** [R/cie-sql.R#L443-L476](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-sql.R#L443-L476)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs12 — Descripción de `cie10_sql()` insuficiente
<summarize></summarize>

**Comentario:** la descripción debía explicar que la función permite hacer consultas SQL sobre el conjunto de datos CIE-10 Chile incluido en el paquete ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se reescribió la descripción de `cie10_sql()` dejando claro que ejecuta consultas SQL de solo lectura sobre el dataset embebido `cie10_cl`.

**Ubicación:** [R/cie-sql.R#L276-L286](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-sql.R#L276-L286)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs13 — Sentencias SQL permitidas mal documentadas
<summarize></summarize>

**Comentario:** en el argumento `query` faltaba aclarar qué sentencias se permiten: además de SELECT, WHERE y JOIN, también FROM, ORDER BY, GROUP BY y HAVING ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se documentaron explícitamente en `@param query` todas las instrucciones permitidas: SELECT, FROM, WHERE, JOIN, ORDER BY, GROUP BY y HAVING.

**Ubicación:** [R/cie-sql.R#L287-L290](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-sql.R#L287-L290)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs14 — Terminología "query/keyword/statements" sin traducir
<summarize></summarize>

**Comentario:** los mensajes debían estar en un solo idioma: "query" como "consulta", "keyword" como "instrucción" o similar, "statements SQL" como "sentencias SQL" ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** los cuatro mensajes de `cie10_sql()` quedaron en castellano: "Solo consultas `SELECT` permitidas", "La consulta contiene una palabra clave no permitida", "Múltiples sentencias SQL no permitidas" y "Error al ejecutar la consulta SQL" (`R/cie-sql.R`). Los snapshots y tests que fijaban los textos antiguos se actualizaron en conjunto.

**Ubicación:** [R/cie-sql.R#L328](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-sql.R#L328) y [#L347-L372](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-sql.R#L347-L372)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs15 — Errores de `DBI::dbGetQuery()` sin capturar
<summarize></summarize>

**Comentario:** parte de los errores llegaba cruda desde R en inglés; sería bueno capturarlos y generar errores adecuados para el usuario ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** la llamada a `DBI::dbGetQuery()` quedó envuelta en `tryCatch()` que relanza con `cli::cli_abort()` en castellano, clase `ciecl_sql_error` y el error original como `parent`. Tests nuevos en `tests/testthat/test-cie-sql.R`.

**Ubicación:** [R/cie-sql.R#L367-L376](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-sql.R#L367-L376)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs16 — Falta el dataset en "See also"
<summarize></summarize>

**Tu comentario:** sin conocer el nombre de la tabla y los campos disponibles es difícil escribir una consulta SQL ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se agregó `@seealso [cie10_cl]` en la documentación de `cie10_sql()`.

**Ubicación:** [R/cie-sql.R#L296-L297](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-sql.R#L296-L297)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs17 — Chequeos básicos de `query` faltantes
<summarize></summarize>

**Comentario:** `cie10_sql()` no chequeaba tipo de dato, longitud ni valores faltantes, y producía errores crípticos de R base ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se agregó validación explícita de `query` antes de ejecutar: tipo character, longitud 1 y no `NA`, con mensajes `cli` en castellano. Tests nuevos que cubren número, vector largo, `NA` y argumento ausente en `tests/testthat/test-cie-sql.R`.

**Ubicación:** [R/cie-sql.R#L315-L321](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-sql.R#L315-L321) y [tests/testthat/test-cie-sql.R#L621-L646](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-cie-sql.R#L621-L646)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs18 — Tildes faltantes en la ayuda de `cie_norm()`
<summarize></summarize>

**Comentario:** la ayuda en castellano no tenía acentos ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se corrigió la ortografía completa del bloque roxygen de `cie_norm()` y se regeneró el `.Rd`.

**Ubicación:** [R/cie-utils.R#L1-L24](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-utils.R#L1-L24)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs19 — `cie_norm()` sin chequeo de parámetro obligatorio
<summarize></summarize>

**Comentario:** faltaba el chequeo de ejecutar la función sin valores en los parámetros ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se creó el helper interno `check_required_es()` en `R/globals.R`, que lanza `cli::cli_abort()` en castellano con clase `ciecl_invalid_input`, y se aplicó en `cie_norm()`. Detalle técnico: `rlang::check_required()` no se disparaba con defaults `lifecycle::deprecated()`; el helper usa `missing()`.

**Ubicación:** [R/globals.R#L27-L35](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/globals.R#L27-L35) y [R/cie-utils.R#L67](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-utils.R#L67)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs20 — "DB" en el mensaje de error y tildes en `cie_normalizar()`
<summarize></summarize>

**Comentario:** reemplazar "DB" por "Base de datos", agregar la tilde en "Códigos" usando el código unicode (con `stringi::stri_escape_unicode()`), y corregir la ortografía de la ayuda de `cie_normalizar()` aunque esté deprecated ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** en `cie_validate_vector()` el mensaje pasó a "Códigos no encontrados en base de datos MINSAL", con la tilde vía escape `\uXXXX` tal como sugeriste. Las tildes de la ayuda de `cie_normalizar()` también quedaron corregidas.

**Ubicación:** [R/cie-utils.R#L257-L259](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-utils.R#L257-L259) y [#L151-L155](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-utils.R#L151-L155)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs21 — Tildes faltantes en la ayuda de `cie_expand()`
<summarize></summarize>

**Comentario:** mismos errores de ortografía en la ayuda ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se corrigieron las tildes del bloque roxygen de `cie_expand()` y se regeneró el `.Rd`.

**Ubicación:** [R/cie-utils.R#L270-L275](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-utils.R#L270-L275)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs22 — `cie_expand()` da error en inglés sin `code`
<summarize></summarize>

**Comentario:** `argument "code" is missing, with no default`; considerar manejarlo con cli ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** `cie_expand()` ahora usa `check_required_es()`, que lanza un mensaje en castellano con clase `ciecl_invalid_input`.

**Ubicación:** [R/cie-utils.R#L291](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-utils.R#L291)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs23 — Tildes faltantes en la ayuda de `cie_table()`
<summarize></summarize>

**Comentario:** falta de acentos en la ayuda ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se corrigieron las tildes del bloque roxygen de `cie_table()` y se regeneró el `.Rd`.

**Ubicación:** [R/cie-table.R#L1-L4](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-table.R#L1-L4)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs24 — `cie_table()` da error en inglés sin `code`
<summarize></summarize>

**Comentario:** mismo error de argumento faltante sin manejar con cli ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** `cie_table()` ahora usa `check_required_es()` con mensaje en castellano.

**Ubicación:** [R/cie-table.R#L30](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-table.R#L30)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs25 — Tildes faltantes en la ayuda de `cie_siglas()`
<summarize></summarize>

**Comentario:** falta de acentos en la ayuda de la función de siglas ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se corrigieron "Listar siglas médicas (deprecated)" y "filtrar por categoría" en `R/cie-siglas.R`, y se regeneró `man/cie_siglas.Rd`.

**Ubicación:** [R/cie-siglas.R#L522-L526](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-siglas.R#L522-L526)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs26 — Tildes faltantes en `cie_short()` y `cie_search()`
<summarize></summarize>

**Comentario:** falta de acentos en las ayudas de ambas funciones ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se corrigieron las tildes en los bloques roxygen de `cie_short()` (`R/cie-siglas.R`) y `cie_search()` (`R/cie-search.R`), con sus `.Rd` regenerados.

**Ubicación:** [R/cie-siglas.R#L453](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-siglas.R#L453) y [R/cie-search.R#L32-L39](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-search.R#L32-L39)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs27 — Conector "and" en `cie_short()` y error en inglés en `cie_search()`
<summarize></summarize>

**Comentario:** el warning de categorías válidas mezclaba castellano con el conector "and", y `cie_search()` sin `text` daba "`text` is absent but must be supplied" en inglés ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** el "and" venía del colapso de vectores de `cli`, que usa ese conector por defecto en cualquier idioma. Se forzó el conector castellano con `cli::cli_vec(x, style = list("vec-last" = " y "))` en `cie_short()` y `cie_validate_vector()`. En `cie_search()` se reemplazó `rlang::check_required()` por `check_required_es()` (`R/cie-search.R`), con mensaje en castellano y clase `ciecl_invalid_input`.

**Ubicación:** [R/cie-siglas.R#L505-L508](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-siglas.R#L505-L508) y [R/cie-search.R#L123](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-search.R#L123)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs28 — Valores sin tildes/eñe en la lista y tests de las ramas else
<summarize></summarize>

**Comentario:** dos puntos: (a) por qué los valores de la lista evitan tildes y eñes, si eso no complica la búsqueda; (b) los tests nunca ejercitaban las dos ramas else de `cie_search()` ("sin palabras válidas tras sanitizar" y "fallback") ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** (a) los valores internos se mantienen sin tildes/eñe por diseño: son tokens usados para matching, igual que el resto de los identificadores del paquete; la prosa de la ayuda que los rodea sí quedó con ortografía completa. (b) se agregaron dos tests en `tests/testthat/test-coverage-paths.R` que ejercitan ambas ramas y verifican que la consulta de respaldo devuelve datos reales (filas, columnas y códigos existentes), no solo que no falla.

**Ubicación:** [tests/testthat/test-coverage-paths.R#L33-L48](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-coverage-paths.R#L33-L48)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs29 — Tildes faltantes en la ayuda de `cie_lookup()`
<summarize></summarize>

**Comentario:** falta de acentos en la ayuda ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se corrigieron las tildes del bloque roxygen de `cie_lookup()` y se regeneró el `.Rd`.

**Ubicación:** [R/cie-lookup.R#L36](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-lookup.R#L36)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs30 — Validaciones de parámetros faltantes en `cie_lookup()`
<summarize></summarize>

**Comentario:** había validaciones de valores no válidos ausentes, con errores poco explicativos en los condicionales del cuerpo de la función ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** `cie_lookup()` ahora valida sus parámetros al inicio, incluido `check_required_es()` para el argumento obligatorio, con mensajes `cli` en castellano que guían al usuario.

**Ubicación:** [R/cie-lookup.R#L130](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-lookup.R#L130)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs31 — `cie_guide()` poco visible
<summarize></summarize>

**Comentario:** la tabla comparativa es útil pero debería estar en las ayudas, una viñeta y el README, no solo como función ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se decidió mantener la función (sirve para consulta rápida en consola) y se agregó un enlace cruzado bilingüe a `cie_guide()` desde el README, quedando también listada en el sitio (`_pkgdown.yml`).

**Ubicación:** [README.Rmd#L164-L166](https://github.com/RodoTasso/ciecl/blob/24b5ecf/README.Rmd#L164-L166) y [_pkgdown.yml#L75](https://github.com/RodoTasso/ciecl/blob/24b5ecf/_pkgdown.yml#L75)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs32 — Tildes faltantes en la ayuda de `cie_guide()`
<summarize></summarize>

**Comentario:** falta de acentos en la ayuda ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se corrigieron las tildes del bloque roxygen de `cie_guide()` y se regeneró el `.Rd`.

**Ubicación:** [R/cie-guide.R#L4-L6](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-guide.R#L4-L6)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs33 — Tildes faltantes en la ayuda de `cie_describe()`
<summarize></summarize>

**Comentario:** falta de acentos en la ayuda ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018)).

**Implementado:** se corrigieron las tildes del bloque roxygen de `cie_describe()` y se regeneró el `.Rd`.

**Ubicación:** [R/cie-describe.R#L1-L4](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-describe.R#L1-L4)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ybs34 — `cie_map_comorbid()` clasifica entradas inválidas como "Otra" sin aviso
<summarize></summarize>

[…]
