### 2026-08-19 #765 @Rodotasso
Hola @maelle, siguiendo el mismo formato que acabo de hacer para @yabellini, dejaré las respuestas un ítem por comentario, con un resumen de tu punto, qué se implementó y el link al commit donde quedó el cambio.

Nota: tu review va de ms01 a ms20 pero no incluye ms13 (al parecer fue saltado o es el sin número), así que son 19 ítems numerados más el punto sin número sobre `:::` en tests, 20 entradas en total. Todos los commits referenciados ya están en `main`.

ms01 — CONTRIBUTING.md: español primero
<summarize></summarize>

**Comentario:** por qué no poner primero la versión en español del texto y después la versión en inglés ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** se reordenó `CONTRIBUTING.md` dejando primero el español y después el inglés, coherente con el español como idioma principal del paquete. La sección Tests quedó además espejo ES/EN, con los mismos 5 bullets en ambos idiomas.

**Ubicación:** [CONTRIBUTING.md#L1-L5](https://github.com/RodoTasso/ciecl/blob/24b5ecf/CONTRIBUTING.md#L1-L5)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694) y [`77a0470`](https://github.com/RodoTasso/ciecl/commit/77a0470d68966ce8fdfba20c5311248c23d28e42)

ms02 — Enlace "English" del navbar
<summarize></summarize>

**Comentario:** el enlace "English" del sitio pkgdown llevaba al README en GitHub; mejor usar `README.en.html` del propio sitio ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** en `_pkgdown.yml` el ítem "English" del navbar ahora apunta a `README.en.html` dentro del sitio. El navbar y la lista de artículos quedaron espejo ES/EN (4+4), incorporando la nueva viñeta `languages.Rmd`.

**Ubicación:** [_pkgdown.yml#L51](https://github.com/RodoTasso/ciecl/blob/24b5ecf/_pkgdown.yml#L51)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694) y [`77a0470`](https://github.com/RodoTasso/ciecl/commit/77a0470d68966ce8fdfba20c5311248c23d28e42)

ms03 — Orden y consistencia del párrafo de características del README
<summarize></summarize>

**Comentario:** agrupar primero las cosas de código/datos (catálogo, siglas, API CIE-11) y después las funcionalidades de transformación; y mencionar una función por elemento o ninguna ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** se reordenó el párrafo de características en `README.Rmd` con ese agrupamiento y se uniformó la mención de funciones. Se regeneraron `README.md` y `README.en.md`.

**Ubicación:** [README.Rmd#L61-L68](https://github.com/RodoTasso/ciecl/blob/24b5ecf/README.Rmd#L61-L68)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms04 — `anyNA()` en vez de `any(is.na(...))`
<summarize></summarize>

**Comentario:** Jarl marcó `any(is.na(...))` en `test-cie-describe.R:17` y `:77` como ineficiente ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** ambas ocurrencias corregidas a `anyNA()` en `tests/testthat/test-cie-describe.R` con el autofix de Jarl.

**Ubicación:** [tests/testthat/test-cie-describe.R#L17](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-cie-describe.R#L17) y [#L77](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-cie-describe.R#L77)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms05 — `uso_cl == TRUE` redundante en `data.R`
<summarize></summarize>

**Comentario:** Jarl marcó `uso_cl == TRUE` (redundant_equals) en `R/data.R:43` ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** ya estaba resuelto en el trabajo previo de la revisión de @yabellini (ybs09): el ejemplo pasó a `subset(cie10_cl, uso_cl != "legado")`. Se verificó que no queda ningún `== TRUE` residual en `R/data.R`.

**Ubicación:** [R/data.R#L45](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/data.R#L45)

**Commit:** [`8c181f7`](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)

ms06 — `expect_all_false()` en test-edge-cases.R
<summarize></summarize>

**Comentario:** en vez de `expect_true(all(!resultado))` se podía usar `expect_all_false()`, más fácil de leer ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** corregido en `tests/testthat/test-edge-cases.R`. Como `expect_all_false()` requiere testthat reciente, DESCRIPTION fijó el piso `testthat (>= 3.3.0)`.

**Ubicación:** [tests/testthat/test-edge-cases.R#L321](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-edge-cases.R#L321) y [DESCRIPTION#L45](https://github.com/RodoTasso/ciecl/blob/24b5ecf/DESCRIPTION#L45)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694) y [`77a0470`](https://github.com/RodoTasso/ciecl/commit/77a0470d68966ce8fdfba20c5311248c23d28e42)

ms07 — Expectativas más específicas (8 fixes de Jarl)
<summarize></summarize>

**Comentario:** Jarl encontró 8 expectativas que podían ser más específicas (`expect_named`, `expect_length`, `expect_type`), fixeables con `--fix` ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** se aplicaron las 8 correcciones: `expect_named` en `test-cie-comorbid.R`, `test-comorbid-validation.R` y `test-robustness.R`; `expect_length` en `test-cie-search.R` (2) y `test-integration.R` (2); `expect_type` en `test-cie-sql.R`.

**Ubicación:** [tests/testthat/test-cie-comorbid.R#L674](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-cie-comorbid.R#L674), [tests/testthat/test-cie-search.R#L29](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-cie-search.R#L29) y [tests/testthat/test-cie-sql.R#L517](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-cie-sql.R#L517)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms08 — Tildes faltantes en el índice de referencia del sitio
<summarize></summarize>

**Comentario:** en https://rodotasso.github.io/ciecl/reference/index.html faltaban tildes ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** barrido de tildes en los `@title` roxygen que alimentan el índice de referencia y en los títulos de sección de `_pkgdown.yml`. Índice verificado tras rebuild de pkgdown.

**Ubicación:** [_pkgdown.yml#L67-L76](https://github.com/RodoTasso/ciecl/blob/24b5ecf/_pkgdown.yml#L67-L76)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms09 — Footer del sitio (tu PR #15)
<summarize></summarize>

**Comentario:** el footer era un poco raro e hiciste una PR directa: Rodotasso/ciecl#15 ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** tu PR #15 se mergeó en `main` ([04660d5](https://github.com/RodoTasso/ciecl/commit/04660d57d606797c7b677c066a3f9cb75264ccb5)) y el fix quedó también aplicado en la rama de trabajo (`_pkgdown.yml`: `developed_by` pasa de bloque anidado a escalar, "Desarrollado por Rodolfo Tasso Suazo"). Gracias por la contribución directa.

**Ubicación:** [_pkgdown.yml#L104-L107](https://github.com/RodoTasso/ciecl/blob/24b5ecf/_pkgdown.yml#L104-L107)

**Commit:** [`04660d5`](https://github.com/RodoTasso/ciecl/commit/04660d57d606797c7b677c066a3f9cb75264ccb5) y [`77a0470`](https://github.com/RodoTasso/ciecl/commit/77a0470d68966ce8fdfba20c5311248c23d28e42)

ms10 — Enlace "Inicio" apuntaba a index.html
<summarize></summarize>

**Comentario:** por qué el enlace "Inicio" del navbar iba a index.html en vez de a una viñeta de introducción ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** resuelto junto con ms18: la nueva viñeta de introducción quedó expuesta como entrada principal del sitio (mecanismo "Para empezar" de pkgdown con la viñeta principal del paquete), de modo que la navegación ya no aterriza en `index.html`.

**Ubicación:** [vignettes/ciecl.Rmd#L1-L7](https://github.com/RodoTasso/ciecl/blob/24b5ecf/vignettes/ciecl.Rmd#L1-L7)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms11 — "Guía de inicio" solo trataba instalación
<summarize></summarize>

**Comentario:** esperabas algo como un caso de uso para principiantes, no solo instalación ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** absorbido por ms18: la guía de inicio ahora presenta un caso de uso real con progresión de complejidad (ver ms18).

**Ubicación:** [vignettes/ciecl.Rmd#L23-L31](https://github.com/RodoTasso/ciecl/blob/24b5ecf/vignettes/ciecl.Rmd#L23-L31)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms12 — Viñeta de instalación: conservar solo pak
<summarize></summarize>

**Comentario:** se podía conservar solo pak como método; la viñeta era difícil de mantener si cambian dependencias sugeridas ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** la viñeta de instalación quedó simplificada con pak como único método recomendado. Las secciones informativas (dependencias por funcionalidad, requisitos de sistema, verificación de la API, caché, troubleshooting) quedaron como espejo 1:1 ES/EN en `instalacion.Rmd` / `installation.Rmd`.

**Ubicación:** [vignettes/instalacion.Rmd#L22-L40](https://github.com/RodoTasso/ciecl/blob/24b5ecf/vignettes/instalacion.Rmd#L22-L40)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694) y [`77a0470`](https://github.com/RodoTasso/ciecl/commit/77a0470d68966ce8fdfba20c5311248c23d28e42)

ms14 — Mención de IA movida a CONTRIBUTING.md
<summarize></summarize>

**Comentario:** la mención de IA del README tendría que estar en CONTRIBUTING.md, con más detalle sobre qué considerar si alguien hace una PR con Claude ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** la mención de IA vive ahora solo en `CONTRIBUTING.md`, como sección propia "Declaración de asistencia con IA": herramientas usadas, alcance (documentación y scripts de benchmark), revisión y responsabilidad humana completa, y la nota de que ninguna IA generó ni analizó los datos del paquete. Se conserva la política para contribuciones externas asistidas por IA (es/en). Los READMEs ya no contienen la declaración.

**Ubicación:** [CONTRIBUTING.md#L47-L49](https://github.com/RodoTasso/ciecl/blob/24b5ecf/CONTRIBUTING.md#L47-L49)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms15 — `nzchar(trimws(text))` en cie-api.R
<summarize></summarize>

**Comentario:** en vez de `nchar(trimws(text)) == 0` se podía usar `nzchar(trimws(text))` ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** corregido en `R/cie-api.R`.

**Ubicación:** [R/cie-api.R#L83](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-api.R#L83)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms16 — Explaining variables en la validación de `max_results`
<summarize></summarize>

**Comentario:** la condición compuesta de `max_results` sería más legible con explaining variables ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** se refactorizó la validación en `R/cie-api.R` extrayendo variables explicativas (tipo, escalaridad, positividad, enteridad), manteniendo el mismo mensaje de error. Los identificadores de ese archivo pasaron a inglés ASCII, siguiendo la convención del paquete (código en inglés, comentarios en español).

**Ubicación:** [R/cie-api.R#L87-L91](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-api.R#L87-L91)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms17 — `api_key` como argumento (acuerdo del 2026-08-11)
<summarize></summarize>

**Comentario:** poner `api_key` como argumento te parecía peligroso; debería ser solo variable de entorno para evitar llaves escritas en scripts ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)). Tras mi propuesta del patrón oficial de httr2, respondiste el 2026-08-11: *"Entonces estoy de acuerdo con conservar el argumento, pero tu propuesta 2 me parece importante."* ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5249517131)).

**Implementado (incluye la propuesta 2):** la firma quedó alineada al código de httr2: `api_key = get_icd_api_key()` en `cie11_search()`, donde el helper exportado lee `ICD_API_KEY` al momento de la llamada y aborta con mensaje orientativo si falta. La propuesta 2 quedó aplicada como nueva subsección "Seguridad de la API key" en `@details` de `cie11_search()` (`R/cie-api.R`): no escribir nunca la llave literal en scripts (riesgo al compartir código o vía `.Rhistory`), recomendación de `ICD_API_KEY` con `usethis::edit_r_environ()`, y el argumento reservado para circunstancias excepcionales. La misma nota se agregó en las viñetas de instalación (`instalacion.Rmd` / `installation.Rmd`, es/en). La propuesta 3 (deprecación) quedó descartada de común acuerdo.

**Ubicación:** [R/cie-api.R#L14-L27](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-api.R#L14-L27) y [vignettes/instalacion.Rmd#L95-L98](https://github.com/RodoTasso/ciecl/blob/24b5ecf/vignettes/instalacion.Rmd#L95-L98)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms18 — Viñeta de caso de uso como guía de inicio.
<summarize></summarize>

**Comentario:** empezar con los datos, decir quién recibiría esos datos y con qué objetivo; hacerlo más concreto (una persona a cargo de un reporte de diabetes en un hospital), mostrar también búsquedas sin resultados ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** se reescribió la documentación de caso de uso como guía de inicio con una narrativa de usuario concreta (persona responsable de un reporte en un hospital), partiendo de los datos y su objetivo, e incluyendo búsquedas sin resultados y cómo interpretarlas. El caso de egresos quedó unificado como par ES/EN 1:1: `ciecl.Rmd` (ES, viñeta principal de get-started) y `case-study-discharges.Rmd` (EN, reescrita como espejo). Las viñetas `ciecl-es.Rmd` / `ciecl-en.Rmd` siguen existiendo como par de introducción, listadas en `_pkgdown.yml`.

**Ubicación:** [vignettes/ciecl.Rmd#L23-L31](https://github.com/RodoTasso/ciecl/blob/24b5ecf/vignettes/ciecl.Rmd#L23-L31) y [vignettes/case-study-discharges.Rmd#L23-L31](https://github.com/RodoTasso/ciecl/blob/24b5ecf/vignettes/case-study-discharges.Rmd#L23-L31)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

ms19 — Ejemplos de `cie11_search()` con vcr
<summarize></summarize>

**Comentario:** los ejemplos con `withr::with_envvar(...)` no funcionaban; sugeriste usar vcr para grabar respuestas en ejemplos (`vcr::insert_example_cassette()`) ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** se adoptó el patrón oficial de vcr 2.x (compatible con httr2): `vcr (>= 2.0.0)` en Suggests; el ejemplo real `cie11_search("depresion mayor")` corre siempre envuelto en `\dontshow{}` con `vcr::insert_example_cassette("cie11_search", package = "ciecl", match_requests_on = c("method", "uri"))`, con guard `requireNamespace("vcr")` y una llave ficticia temporal solo si `ICD_API_KEY` no existe. El cassette se grabó contra la API real de la OMS en `inst/_vcr/cie11_search.yml`; como se grabó con `match_requests_on = c("method", "uri")`, ni headers ni body quedan en el YAML, y se verificó con grep que las credenciales no aparecen. `tests/testthat/setup-vcr.R` configura `filter_sensitive_data` para que cualquier cassette futuro reemplace la llave real. Los ejemplos corren offline en R CMD check y en CI sin credenciales; los tests contra la API real mantienen su `skip_if()`.

**Ubicación:** [R/cie-api.R#L47-L49](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-api.R#L47-L49), [tests/testthat/setup-vcr.R#L10](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/setup-vcr.R#L10) y [inst/_vcr/cie11_search.yml](https://github.com/RodoTasso/ciecl/blob/24b5ecf/inst/_vcr/cie11_search.yml)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694) y [`77a0470`](https://github.com/RodoTasso/ciecl/commit/77a0470d68966ce8fdfba20c5311248c23d28e42)

ms20 — Tildes en mensajes cli de todo el paquete
<summarize></summarize>

**Comentario:** en mensajes cli faltaban tildes (ej. "codigos vacios que seran ignorados"); había que revisar todo el texto del paquete, no solo la documentación ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** barrido completo de mensajes `cli_*`/`message`/`warning`/`stop` en `R/`, con tildes vía escapes `\uXXXX` según la convención del paquete (0 warnings de caracteres no-ASCII en R CMD check). Se completó con la restauración de tildes UTF-8 en la sección ES de `CODE_OF_CONDUCT.md` y en comentarios de `R/cie-api.R`.

**Ubicación:** [R/cie-comorbid.R#L49-L56](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-comorbid.R#L49-L56)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694) y [`77a0470`](https://github.com/RodoTasso/ciecl/commit/77a0470d68966ce8fdfba20c5311248c23d28e42)

Punto sin número — `:::` en tests (9 casos)
<summarize></summarize>

**Comentario:** en los tests no hace falta `:::` para llamar funciones internas, porque testthat carga el paquete; Jarl marcó 9 casos ([comentario](https://github.com/ropensci/software-review/issues/765#issuecomment-5237964358)).

**Implementado:** se eliminó el prefijo `ciecl:::` en los 9 lugares: `test-cie-sql.R` (6 ocurrencias de `get_cache_dir()`), `test-robustness.R` (2) y `test-utils-internal.R` (`cie_lookup_single`).

**Ubicación:** [tests/testthat/test-cie-sql.R#L194](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-cie-sql.R#L194) y [tests/testthat/test-utils-internal.R#L269](https://github.com/RodoTasso/ciecl/blob/24b5ecf/tests/testthat/test-utils-internal.R#L269)

**Commit:** [`eb3ab6e`](https://github.com/RodoTasso/ciecl/commit/eb3ab6e9f6386f0fc67c761dbeaec78a6c97c694)

Espero que esta forma de entregar sus revisiones y sugerencias sean más útiles que las que se entregaron antes.
De todas formas si hay algo más que cambiar no tengo problema. 👍 🫰
