### 2026-08-19 #765 @maelle
Hola @Rodotasso, gracias por tu trabajo.

No pienso que es comodo hacer un clic para cada comentario, creo que preferria que envias tu respuesta sin lo de los "details". Además, en vez de reformular lo que se preguntó, mejor citar lo que se preguntó (si no, es un esfuerzo para las personas que comproban la respuesta, en particular reconozco mejor mis comentarios tal cual en vez de su versión reformulada). En vez de

```md
Comentario: falta de acentos en la ayuda de la función de siglas (https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018).

Implementado: se corrigieron "Listar siglas médicas (deprecated)" y "filtrar por categoría" en R/cie-siglas.R, y se regeneró man/cie_siglas.Rd.

Ubicación: [R/cie-siglas.R#L522-L526](https://github.com/RodoTasso/ciecl/blob/24b5ecf/R/cie-siglas.R#L522-L526)

Commit: [8c181f7](https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773)
```

`https://github.com/ropensci/software-review/issues/765#issuecomment-4769658018` es el enlace de la revisión de @yabellini, no hace falta repetir este enlace.

El commit `https://github.com/RodoTasso/ciecl/commit/8c181f7ea310f534311db889b2ce6d45745d0773` esta demasiado largo para ser útil. Y claro no seria tan bueno un enlace hasta un commit de la rama dev dado que después este commit no se encuentra tal cual en main. Asi que te aconsejeria solo usar el permalink, aunque aqui lo más útil seria un enlace hasta la pagina pkdown.

Además este comentario de Yani esta repetido para varias funciones, será más facil para nosotras que hagas un resumén asi:

```md
> ybs23, <otros codigos>: mismo comentario que las funciones anteriores sobre la falta de acentos en la ayuda.

Coregí este problema en las paginas de [`cie_table()`](https://rodotasso.github.io/ciecl/reference/cie_table.html), <otras-paginas>.
```

Para todos los elementos, es importante pensar en la persona que lee el texto, no seguir una formula general.

A ver si @yabellini esta de acuerdo conmigo, y si tiene otros comentarios. De verdad queremos ayudar 😸
