### 2026-08-19 #765 @maelle
> No dejaré mis propios comentarios por que es mucha info.

Qué quieres decir, cual tipo de comentarios?
