### 2026-08-19 #765 @Rodotasso
Hola @maelle, muchas gracias por tu sugerencia, entiendo y me parece una buena forma de hacer las cosas. Intentaré juntar la primera entrega de las respuestas (agrupadas) y los links. Lamento lo complejo de las respuestas, solo entorpece el proceso de revisión.

Si prefieren, puedo hacer otra vez la entrega de los cambios.

Sobre mi expresión de mis comentarios, por lo general siempre agrego más información de la necesaria, en este caso, `uso_cl` fue creado para aclarar el trabajo con 2 listas tabulares CIE-10 simultáneas (o una que contiene a la otra), por lo que debe aclararse en el README (por eso el número grande de códigos CIE-10, ~38k). O que debo pulir el trabajo de `cie_search` y las siglas médicas, al tener un threshold o una tolerancia al error tan baja (0.7), algunas siglas entregan el resultado incorrecto, y hay que decidir si se redirigen al resultado correcto o se sube la tolerancia. Ese tipo de comentarios, o procesos en los que me gusto trabajar.

Agregar este tipo de información a respuestas que ya son largas y llevaban tiempo hechas, solo nubla aún más el proceso, por eso intenté mantenerlas al margen y centrarme en la información que se necesitaba.

Nuevamente agradezco, y trataré de ser más claro con las respuestas. Dejando de usar details también (a mi me gustaba, pero ya estoy viejo jaja).
