### 2026-08-19 #765 @yabellini
Hola, gracias por esta charla. 😄
Prefiero respuestas cortas, que indiquen qué cambio se hizo (si se hizo alguno, si no se hizo una explicación de la razón) y un link donde yo pueda ver el cambio propiamente dicho, así puedo verificar (y preguntar/sugerir si es que tengo preguntas o sugerencias nuevas)
