### 2026-08-20 #765 @Rodotasso
Hola @yabellini y @maelle . Si me permiten, separaré en 2 post con las revisiones de cada una, así tampoco queda un mega post. 🫰 También intenté que se mantuviera alineado a los ejemplos de las guías de rOpenSci y considerar sus solicitudes 😸

**Dataset `cie10_cl`**

> **ybs08**  *"El campo uso_cl esta marcado como binario en la documentacion, pero en realidad tiene texto"*
> Hecho: `uso_cl` documentada como character categórica con sus 5 valores y qué significa `"legado"`. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/data.R#L31)

> **ybs09**  *"El ejemplo (...) devuelve un conjunto de datos vacío"*
> Hecho: ejemplo reemplazado por `subset(cie10_cl, uso_cl != "legado")`, consistente con el filtro interno. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/data.R#L45)

**Caché y conexión**

> **ybs10 / ybs11**  *"La ayuda sobre esta función es muy escueta"* / *"explique qué es el bloqueo del archivo (...) y por qué es necesario liberarlo"*
> Hecho: ayudas reescritas (qué es la caché SQLite, cuándo reconstruirla, qué es el bloqueo del `.db` y qué pasa si no se libera); el spanglish residual de `@returns` también quedó corregido. [cie10_clear_cache](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-sql.R#L381-L407) · [cie10_disconnect](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-sql.R#L443-L466)

**`cie10_sql()`**

> **ybs12 / ybs13 / ybs16**  *"dar mas información en la descripción"*, *"aclarar mejor qué sentencias SQL se permiten usar"*, *"Agregar el conjunto de datos a 'See also'"*
> Hecho: descripción reescrita (consultas de solo lectura sobre el dataset embebido), `@param query` lista SELECT/FROM/WHERE/JOIN/ORDER BY/GROUP BY/HAVING, y `@seealso` con `cie10_cl`. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-sql.R#L276-L297)

> **ybs14 / ybs15 / ybs17**  *"los mensajes de error o todos en castellano o todos en ingles"*, *"No se captura ningun error en el llamado a DBI::dbGetQuery"*, *"le faltan algunos chequeos basicos de parametros"*
> Hecho: validación de `query` (tipo, largo, `NA`), mensajes `cli` en castellano y `tryCatch` que relanza el error de DBI con mensaje propio. [Validación](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-sql.R#L315-L321) · [tryCatch](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-sql.R#L367-L375) · [tests](https://github.com/Rodotasso/ciecl/blob/d5035f5/tests/testthat/test-cie-sql.R#L621-L646)

**Tildes en las ayudas**

> **ybs18, ybs21, ybs23, ybs25, ybs26, ybs29, ybs32, ybs33, ybs35**  *"la ayuda está en castellano pero le faltan todos los acentos"* (mismo comentario en varias funciones)
> Hecho: corrección ortográfica completa de las ayudas de `cie_norm`, `cie_expand`, `cie_table`, `cie_siglas`, `cie_short`, `cie_search`, `cie_lookup`, `cie_guide`, `cie_describe` y `cie11_search` (la de `cie_normalizar` se cubre en ybs20). Ver en el sitio: [índice de referencia](https://rodotasso.github.io/ciecl/reference/index.html), ej. [cie_norm](https://rodotasso.github.io/ciecl/reference/cie_norm.html), [cie_search](https://rodotasso.github.io/ciecl/reference/cie_search.html).

**Chequeos de parámetros**

> **ybs19 / ybs22 / ybs24 / ybs30**  *"le falta el chequeo ejecutar la función sin pasarle valores a los parametros"* (mismo error en inglés en `cie_expand` y `cie_table`; validaciones ausentes en `cie_lookup`)
> Hecho: helper interno `check_required_es()` con mensaje `cli` en castellano, aplicado en las cuatro funciones; `cie_lookup()` además valida sus parámetros al inicio. [helper](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/globals.R#L27-L35) · [cie_lookup](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-lookup.R#L130)

**Spanglish en mensajes**

> **ybs20 / ybs27**  *"sugiero reemplazar 'DB' en el mensaje de error con 'Base de datos'"* y *"usa la palabra 'and' para decir 'y'"*
> Hecho: el mensaje ahora dice "Códigos no encontrados en la base de datos MINSAL" (tilde vía escape unicode, como sugeriste) y el conector castellano "y" quedó forzado con `cli::cli_vec(style = list("vec-last" = " y "))`. Las tildes de la ayuda de `cie_normalizar()` también se corrigieron. [mensaje](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-utils.R#L255-L261) · [conector](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-siglas.R#L505-L508)

**Puntos individuales**

> **ybs28**  *"nunca se chequea el codigo por dos else"* y *"cuál es la razón de evitar usar esos caracteres?"*
> Tests: dos tests nuevos ejercitan ambas ramas y verifican que devuelven datos reales del catálogo. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/tests/testthat/test-coverage-paths.R#L33-L48)
> Valores sin tildes/eñe: es por diseño, son tokens internos de matching; la prosa de la ayuda sí quedó con ortografía completa.

> **ybs31**  *"La comparativa me parece muy útil, pero mas para que esté en la ayuda de cada función y como una viñeta"*
> Hecho: se mantiene la función (consulta rápida en consola) y se agregó un enlace cruzado bilingüe desde el README; también está listada en el sitio. [README](https://github.com/Rodotasso/ciecl/blob/d5035f5/README.Rmd#L162-L168) · [sitio](https://github.com/Rodotasso/ciecl/blob/d5035f5/_pkgdown.yml#L75)

> **ybs34**  *"cuando se le pasa un valor que no existe, devuelve una tabla con la categoria 'Otra'"*
> Hecho: se agregó un warning que avisa cuántas entradas no tienen formato CIE-10 válido; los códigos válidos no mapeados siguen en "Otra", que es lo esperado en un índice de comorbilidad. [código](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-comorbid.R#L127-L138) · [tests](https://github.com/Rodotasso/ciecl/blob/d5035f5/tests/testthat/test-cie-comorbid.R#L630-L660)

> **ybs36**  *"En cie10_empty_tibble le falta el campo uso_cl"*
> Hecho: `uso_cl` agregada como `character(0)`. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/globals.R#L37-L51)

Este es el primero, continúo en el siguiente.
