### 2026-08-20 #765 @Rodotasso
Hola @maelle, acá va la respuesta a tu revisión. 👍

> **ms01**  *"Porqué no poner primero la versión en español del texto y después la version en inglés?"*
> Hecho: `CONTRIBUTING.md` ahora va primero en español y después en inglés. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/CONTRIBUTING.md#L1-L5)

> **ms14**  *"la mención de IA en el README tendria que ser en el CONTRIBUTING.md, con más detalles"*
> Hecho: movida a `CONTRIBUTING.md` como sección propia ("Declaración de asistencia con IA"), con herramientas, alcance y responsabilidad humana; los READMEs ya no la contienen. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/CONTRIBUTING.md#L47-L51)

**Sitio pkgdown y README**

> **ms02**  *"el enlace 'English' es hasta el README en GitHub"*
> Hecho: ahora apunta a `README.en.html` del propio sitio. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/_pkgdown.yml#L51)

> **ms03**  *"podrias poner las cosas de codigo juntas (...) mencionar una función para cada elemento de la lista, o mencionar ninguna"*
> Hecho: párrafo de características reordenado y uniformado (ninguna función por nombre). [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/README.Rmd#L61-L68)

> **ms08**  *"faltan tildes"* (índice de referencia)
> Hecho: barrido de tildes en los títulos roxygen y en `_pkgdown.yml`. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/_pkgdown.yml#L66-L76)

> **ms09**  *"El footer es un poco raro, hicé una PR"*
> Hecho: tu PR #15 quedó mergeada. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/_pkgdown.yml#L101-L107)

> **ms10 / ms11**  *"Porqué el enlace 'Inicio' (...) es para index.html en vez de una viñeta de introducción?"* / *"solo es sobre instalación, esperaba algo como un caso de uso para principiantes"*
> Hecho: la entrada principal del sitio ahora es la nueva guía de inicio (ver ms18).

> **ms12**  *"se puede conservar solo pak, seria más facil"*
> Hecho: viñeta de instalación simplificada, pak como único método recomendado. [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/vignettes/instalacion.Rmd#L22-L40)

> **ms18**  *"porqué no empezar con los datos y decir quién recibiria estos datos, y cual seria su objectivo?"*
> Hecho: la guía de inicio ahora parte de un caso concreto (persona a cargo del reporte de egresos por diabetes en un hospital), con progresión de complejidad y búsquedas sin resultados. [Ver viñeta](https://rodotasso.github.io/ciecl/articles/ciecl.html)

> **ms19**  *"no funcionan. Podrias usar vcr para recordar respuestas para ejemplos"*
> Hecho: ejemplo de `cie11_search()` con `vcr::insert_example_cassette()`, cassette grabado en `inst/_vcr/` (sin credenciales, verificado) y `filter_sensitive_data` configurado para cassettes futuros; los ejemplos corren offline en check y CI. [ejemplo](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-api.R#L45-L49) · [cassette](https://github.com/Rodotasso/ciecl/blob/d5035f5/inst/_vcr/cie11_search.yml)

**Código y tests**

> **ms04 / ms05 / ms15 / ms16**  *"que podés usar `anyNA()`"*, *"que podés simplificar código en `data.R`"*, *"podés usar `nzchar(trimws(text))`"*, *"podés usar 'explaining variables'"*
> Hecho: los cuatro fixes de Jarl aplicados; el de `data.R` ya quedó resuelto con ybs09. [anyNA](https://github.com/Rodotasso/ciecl/blob/d5035f5/tests/testthat/test-cie-describe.R#L17) · [nzchar](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-api.R#L83) · [variables explicativas](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-api.R#L86-L90)

> **ms17**  *"Me parece peligrosa la idea de poner `api_key` cómo argumento"*
> Resuelto el 2026-08-11: se conserva el argumento (lenguaje httr2) y el riesgo quedó documentado en la sección "Seguridad de la API key". [Ver](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-api.R#L14-L27)

> **ms20**  *"faltan tildes. Hay que revisar todo el texto del paquete, no solo en la documentación"*
> Hecho: barrido completo de mensajes `cli`/`warning`/`stop` en `R/`, con tildes vía escapes `\uXXXX` (0 notas de caracteres no-ASCII en check). [Ejemplo](https://github.com/Rodotasso/ciecl/blob/d5035f5/R/cie-comorbid.R#L47-L57)

> **ms06 / ms07 / `:::` en tests**  *"creo que podrias usar `expect_all_false()`?"*, *"otras funciones que podrian ser más especificas"*, *"En los tests no hace falta usar `:::` para llamar funciones internas del paquete"*
> Hecho: `expect_all_false()` aplicado (DESCRIPTION fija `testthat (>= 3.3.0)`), las 8 expectativas más específicas (`expect_named`/`expect_length`/`expect_type`) y `:::` eliminado en los 9 lugares. [expect_all_false](https://github.com/Rodotasso/ciecl/blob/d5035f5/tests/testthat/test-edge-cases.R#L321) · [expectativas](https://github.com/Rodotasso/ciecl/blob/d5035f5/tests/testthat/test-cie-comorbid.R#L674) · [:::](https://github.com/Rodotasso/ciecl/blob/d5035f5/tests/testthat/test-cie-sql.R#L194)

Espero sea mucho más amigable esta vez.
