### 2026-08-21 #765 @maelle
Gracias, pero podrias por favor reparar el formato para que se vea lo que es nuestro y lo que es tuyo?
