### 2026-08-21 #765 @Rodotasso
Hola @maelle, edite el comentario para que se note la diferencia. Si les parece bien, me gustaría pedir permiso para borrar los otros post con las otras entregas de las respuestas para editar solo uno (debí haber hecho eso en un principio 😝 ).
