### 2026-08-24 #765 @Rodotasso
Hola @maelle!, muchas gracias por el aviso, lo acabo de corregir. Creo que no se me da esto de entregar respuestas adecuadas jajaja. En cuanto tenga tiempo revisaré los comentarios de los otros threads y la estructura de la guía 👍
