### 2026-08-24 #765 @maelle
Gracias! Estoy muy contenta poder leer un guia de inicio. 😃

En https://rodotasso.github.io/ciecl/articles/ciecl.html#paso-2-agregar-las-descripciones-oficiales-con-cie_describe qué significa:

> por lo que se integra en un [mutate()](https://dplyr.tidyverse.org/reference/mutate.html) sin joins temporales:

Además,

> El resultado incluye un score de similitud para evaluar la confiabilidad de cada coincidencia. Ahora sabes que E11.9 y E14.9 de tu base corresponden a diabetes y puedes filtrar tu reporte con criterio.

Esos codigos no aparecen en lo que se ve de la tabla. Además, no se explica cómo se eligieron solo estos dos codigos de los codigos que devolvió la función.

Creo que se podria desarrollar un poquito lo que se haria con los datos del hospital después de conocer el codigo: filtrar los datos, resumirlos.

Otra pregunta: por el momento, tiene el paquete usuarios y usuarias? Dientro de tu equipo?
