### 2026-08-25 #765 @Rodotasso
Muchas gracias @maelle ! Responderé acá mientras espero los comentarios de Yanina.

> *"qué significa: por lo que se integra en un `mutate()` sin joins temporales"*

Es explicación innecesaria. Lo que quería decir es que `cie_describe()` devuelve un vector del mismo largo que la columna de códigos, así que es posible agregar las descripciones como otra columna a la tabla directamente, sin crear tablas intermedias ni cruzarlas después (a diferencia de `cie_lookup()`, que devuelve un tibble y sí requiere un `left_join()`, como se muestra justo debajo). [Ver](https://rodotasso.github.io/ciecl/articles/ciecl.html#paso-2-agregar-las-descripciones-oficiales-con-cie_describe)

Reescribiría la frase en lenguaje simple: (estoy probando los cuadros(cajas) de código jaja)
```
 "`cie_describe()` devuelve un vector de texto con una descripción por cada código, así que puedes agregarlo como una columna más de tu tabla, sin pasos intermedios."
```

> *"Esos códigos no aparecen en lo que se ve de la tabla. Además, no se explica cómo se eligieron solo estos dos codigos de los códigos que devolvió la función."*

Es correcto en ambos puntos. La lógica era implícita: los datos sintéticos de la guía se generaron con dos códigos de diabetes (`E119` y `E149`), y tras la búsqueda de `"diabetis"` (mal escrito), el cruce entre los resultados y los códigos presentes en la base daba exactamente `E11.9` y `E14.9`. Pero ese cruce nunca se mostró como código, y la tabla impresa se trunca antes de que esos códigos sean visibles.

Mi corrección en este caso sería haciendo explícito el cruce con código ejecutable:

```r
# ¿Qué códigos de diabetes están realmente en mi base?
codigos_diabetes <- intersect(
  resultados_busqueda$codigo,
  unique(egresos$DIAG1_NORM)
)
```

> *"Creo que se podría desarrollar un poquito lo que se haría con los datos del hospital después de conocer el código: filtrar los datos, resumirlos."*

De acuerdo, cierra mejor el caso de uso. Agregaría al final del Paso 3 el cierre del reporte: filtrar los egresos por los códigos identificados y resumirlos por tipo de diabetes (conteo por descripción), que además vuelve visible en el output por qué `E11.9` y `E14.9` son los códigos relevantes de la base.

> *"por el momento, tiene el paquete usuarios y usuarias? Dentro de tu equipo?"*

La verdad no tengo un equipo como tal, en el grupo de Ciencia de Datos para la Salud Pública si es usado, pero de forma esporádica para el filtrado y normalización de códigos en investigación (Esperé unas horas para que respondieran, no les había preguntado).
He recibido correos de proyectos de ONGs en Chile, los cuales preguntan si es posible utilizarla con bases de datos de atenciones de urgencia (me gustaría hacer un caso de uso), ya que esos códigos tienen un formato clúster o de rango.
Me han escrito algunos codificadores de GRD (Grupos de pacientes Relacionados por el Diagnóstico), DRG en Inglés, consultando por el proyecto y por la utilidad del paquete.
El más importante es el Departamento de Estadísticas de Información en Salud (DEIS) del Ministerio de Salud de Chile, quienes están probando el paquete y esperando una salida oficial (la revisión por pares es mi prioridad 😸 ).
Usuarios finales son investigadores, personal de salud y estadísticos que trabajen con bases de datos de salud y necesiten filtrar o normalizar por código, ya que en chile las bases de datos en salud pueden ser muy heterogéneas en su composición.
No he hecho publicidad como tal, solo ha tenido esta salida en prensa [Ver](https://saludpublica.uchile.cl/noticias/238027/nuevo-software-chileno-facilita-el-analisis-de-datos-en-salud)

Me extiendo demasiado en las respuestas, pero es mucha información, y tengo más información incluso jaja.
