### 2026-08-27 #765 @maelle
@Rodotasso podrías por favor registrar tu repuesta? https://devguide.ropensci.org/es/bot_cheatsheet.es.html#submit-response-to-reviewers
