### 2026-08-27 #765 @maelle
Algunos comentarios

- ybs10 / ybs11: porqué no renombras la función para que no se use la palabra "caché"? Además, cuando se actualiza el paquete, no debería estar forzado de manera automatica el rebuild la primera vez que se usa la nueva versión del paquete? Comparando la versión del paquete y la versión de los datos. Otra cosa: "—por ejemplo antes de [cie10_clear_cache()] " no podría la función del cache llamar ella misma la función para desconectar?

- No estoy segura de que se necesite una viñeta entera de instalacíon/configuración. Me pregunto si da miedo?

- ms19. El ejemplo en https://rodotasso.github.io/ciecl/reference/cie11_search.html#ref-examples tiene `if (.has_vcr)`, eso lo pondrias esconder mejor con `@examplesIf` o algo asi?

- los cambios en la viñeta que serian una consecuencia de https://github.com/ropensci/software-review/issues/765#issuecomment-5403444413, no los hiciste, verdad? Me gustaría leer la versión mejorada.

Gracias por tu trabajo. 🙏
