### 2026-08-27 1f3a2e7
docs: aplica mejoras de la guia de inicio solicitadas por maelle en rOpenSci #765
