### 2026-08-27 #765 @Rodotasso
Hola @maelle, muchas gracias por los comentarios. Respondo punto por punto. Te cuento también que los cambios van quedando en la rama `dev`: estoy esperando la devolución de @yabellini para integrar todo junto a `main`. Si prefieres que los integre inmediatamente me dices y lo hago. 😸

Respecto a el submit de la respuesta, lo hice unos post arriba para los dos comentarios (no sé si se superponen), debo hacerlo otra vez?.

> *"ybs10 / ybs11: ¿por qué no renombras la función para que no se use la palabra caché?"*

Lo pensé, pero si me permites preferiría mantener `cie10_clear_cache()` por dos razones. Es una función de uso excepcional (solo para una caché corrupta o para liberar disco), fuera del flujo normal de uso. Y con el rebuild automático al actualizar el paquete (tu siguiente punto, que voy a implementar) desaparece el caso que hoy obliga a llamarla a mano, así que la mayoría de los usuarios nunca la verá. Renombrar la función exportada implica un alias deprecado para evitar rupturas, y puede ser desproporcionado para una función de emergencia. Si aún así te parece mejor el rename, lo hago con la deprecación.

> *"cuando se actualiza el paquete, ¿no debería estar forzado de manera automática el rebuild la primera vez que se usa la nueva versión? Comparando la versión del paquete y la versión de los datos"*

De acuerdo, es lo correcto y elimina la fricción manual. Lo implementaría así: guardar la versión del paquete como metadata dentro del `cie10.db` al construirlo, y compararla con `utils::packageVersion("ciecl")` al abrir la conexión; si difieren, la caché se reconstruye sola en ese primer uso. Quedará en `dev` junto con el resto de los cambios.

> *"'por ejemplo antes de cie10_clear_cache()': ¿no podría la función del caché llamar ella misma la función para desconectar?"*

Sí, y de hecho ya lo hace 🙂: `cie10_clear_cache()` cierra internamente la conexión pooled antes de borrar el archivo ([código](https://github.com/Rodotasso/ciecl/blob/dev/R/cie-sql.R)). La frase de la ayuda de `cie10_disconnect()` quedó sugiriendo un paso manual innecesario; la voy a corregir para aclarar que la desconexión previa solo aplica si alguien borra el `.db` por fuera del paquete.

> *"No estoy segura de que se necesite una viñeta entera de instalación/configuración. Me pregunto si da miedo"*

Entiendo la duda. La mantendría por el perfil de algunos de los usuarios del paquete: muchas personas que trabajan con bases de datos de salud en Chile están o podrían estar en proceso de aprendizaje de R, y para ese público la instalación desde GitHub con `pak` y la configuración de la API key de la OMS no son obvias; sin esa guía, la barrera de entrada es real. Para que no intimide a quien ya tiene experiencia, puedo agregar un párrafo inicial que diga para quién es la viñeta y quién puede saltársela. No se si te parece bien esta solución.

> *"ms19: el ejemplo tiene `if (.has_vcr)`, ¿eso lo podrías esconder mejor con `@examplesIf` o algo así?"*

De acuerdo, lo cambio a `@examplesIf` para que la condición no aparezca en la documentación. Quedará en `dev` (hasta la incorporación de cambios de Yanina o tu preferencia de integrarlos ahora).

> *"los cambios en la viñeta que serían consecuencia de tu comentario anterior, ¿no los hiciste, verdad? Me gustaría leer la versión mejorada"*

Correcto, no estaban aplicados: en mi respuesta anterior había dicho que esperaría los comentarios de @yabellini antes de tocar las viñetas. Finalmente los apliqué igual en `dev` ([viñeta](https://github.com/Rodotasso/ciecl/blob/1f3a2e7/vignettes/ciecl.Rmd)):

- La frase de `mutate()` quedó en lenguaje simple: "`cie_describe()` "devuelve un vector de texto con una descripción por cada código, así que puedes agregarlo como una columna más de tu tabla, sin pasos intermedios".
- El cruce entre la búsqueda y la base ahora es código ejecutable con `intersect()`, así se ve de dónde salen `E11.9` y `E14.9`.
- La guía cierra el caso de uso: filtra los egresos por los códigos identificados y los resume por tipo de diabetes, con las glosas oficiales visibles en el output.

Quedo atento a lo que prefieras con los cambios, mientras aplicaré estas recomendaciones en `dev`.
Gracias de nuevo por la revisión 🙏
