### 2026-08-27 0bcda4e
docs: sincroniza gemela en ingles de la guia de inicio con las mejoras de rOpenSci #765
