### 2026-08-27 581c63d
fix: documenta rebuild automatico de cache con tests y esconde guard vcr con @examplesIf
