+++ man/cie11_search.Rd
@@ -56,28 +56,23 @@ una variable de entorno).
 # Ver parámetros disponibles
 args(cie11_search)
 
+\dontshow{if (requireNamespace("vcr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
 \dontshow{
 # El cassette grabado (inst/_vcr/) reproduce la respuesta de la API sin
 # conexión ni credenciales reales; la llave ficticia solo satisface el
 # parseo interno del argumento api_key y se restaura al final.
-# Guard: vcr está en Suggests; sin él se omite la llamada de ejemplo
-# (evita una request real a la API con llave ficticia)
-.has_vcr <- requireNamespace("vcr", quietly = TRUE)
-if (.has_vcr) {
-  .key_prev <- Sys.getenv("ICD_API_KEY", unset = NA)
-  if (is.na(.key_prev)) Sys.setenv(ICD_API_KEY = "client_id:client_secret")
-  vcr::insert_example_cassette("cie11_search", package = "ciecl",
-                               match_requests_on = c("method", "uri"))
-}
+.key_prev <- Sys.getenv("ICD_API_KEY", unset = NA)
+if (is.na(.key_prev)) Sys.setenv(ICD_API_KEY = "client_id:client_secret")
+vcr::insert_example_cassette("cie11_search", package = "ciecl",
+                             match_requests_on = c("method", "uri"))
 }
 # Requiere credenciales OMS gratuitas (https://icd.who.int/icdapi)
-if (.has_vcr) cie11_search("depresion mayor")
+cie11_search("depresion mayor")
 \dontshow{
-if (.has_vcr) {
-  vcr::eject_cassette()
-  if (is.na(.key_prev)) Sys.unsetenv("ICD_API_KEY")
+vcr::eject_cassette()
+if (is.na(.key_prev)) Sys.unsetenv("ICD_API_KEY")
 }
-}
+\dontshow{\}) # examplesIf}
 }
 \seealso{
 \code{\link[=cie_search]{cie_search()}}, \code{\link[=cie_lookup]{cie_lookup()}}, \code{\link[=cie_guide]{cie_guide()}}
