### 2026-08-27 d5b985b
docs: parrafo de audiencia en viñetas de instalacion ES/EN y NEWS 2026-08-27
