### 2026-08-28 #765 @maelle
Hola @Rodotasso! Gracias, miraré más en detalles la semana que viene pero me parece bien si pones los cambios de dev en main, asi Yani y yo podemos mirar la versión más mejorada la semana que viene. 😸
