### 2026-08-28 ad8b4b8
docs: nombra E11.9 y E14.9 en prosa del cruce de busqueda en viñetas guia ES/EN
