### 2026-08-28 5d391d4
ci: continue-on-error en pkgcheck por hallazgos cosmeticos de goodpractice
