+++ reference/cie10_clear_cache.md
@@ -12,11 +12,12 @@ consulta
 [`cie10_sql()`](https://rodotasso.github.io/ciecl/reference/cie10_sql.md),
 etc.) la reconstruya desde cero.
 
-Es necesario forzar el rebuild cuando: (1) se actualiza el paquete a una
-version con un dataset CIE-10 corregido y la caché vieja quedó
-desactualizada, (2) se sospecha que el archivo `.db` está corrupto
-(errores de lectura SQL inesperados), o (3) se quiere liberar el espacio
-en disco que ocupa la caché.
+No es necesario llamarla tras actualizar el paquete: la caché guarda la
+versión del paquete con que se construyó (tabla `cie10_meta`) y, si la
+versión cambió, se reconstruye automáticamente en el primer uso. Los
+casos en que conviene forzar el rebuild manual son: (1) se sospecha que
+el archivo `.db` está corrupto (errores de lectura SQL inesperados), o
+(2) se quiere liberar el espacio en disco que ocupa la caché.
 
 ## Usage
 
