### 2026-08-28 2fcd11f
ci: timeout-minutes por job en workflows para evitar cuelgues de runner
