+++ .github/workflows/codemeta.yaml
@@ -21,6 +21,7 @@ permissions:
 jobs:
   codemeta:
     runs-on: ubuntu-latest
+    timeout-minutes: 15
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
