+++ .github/workflows/urlchecker.yaml
@@ -14,6 +14,7 @@ permissions: read-all
 jobs:
   urlchecker:
     runs-on: ubuntu-latest
+    timeout-minutes: 15
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
