+++ .github/workflows/lint.yaml
@@ -12,6 +12,7 @@ permissions: read-all
 jobs:
   lint:
     runs-on: ubuntu-latest
+    timeout-minutes: 10
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
