+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   instalacion: instalacion.html
   installation: installation.html
   languages: languages.html
-last_built: 2026-08-28T19:18Z
+last_built: 2026-08-28T19:46Z
 urls:
   reference: https://rodotasso.github.io/ciecl/reference
   article: https://rodotasso.github.io/ciecl/articles
