### 2026-09-04 #765 @maelle
> Y con el rebuild automático al actualizar el paquete (tu siguiente punto, que voy a implementar) desaparece el caso que hoy obliga a llamarla a mano, así que la mayoría de los usuarios nunca la verá.

Fue implementado o aún no?
