### 2026-09-04 #765 @maelle
En tu repuesta https://github.com/ropensci/software-review/issues/765#issuecomment-5442619136 hay mucho futuro, pero veo que has puesto los cambios de dev en main. Podrías modificar tu repuesta para que tenga los enlaces y los cambios que ya has hecho? Si no, es muy dificil encontrar la información. Muchas gracias.
