### 2026-09-05 #765 @Rodotasso
Hola @maelle!, muchas gracias por tus mensajes!.

>*Fue implementado o aún no?*

Si, lamento la confusión. Quedó implementado en [f88b77f](https://github.com/Rodotasso/ciecl/commit/f88b77f)

>*En tu repuesta https://github.com/ropensci/software-review/issues/765#issuecomment-5442619136 hay mucho futuro, pero veo que has puesto los cambios de dev en main. Podrías modificar tu repuesta para que tenga los enlaces y los cambios que ya has hecho? Si no, es muy dificil encontrar la información. Muchas gracias.*

Me disculparán el lapsus, pero primero me ilusione pensando que el post era prometedor (*veo mucho futuro*), después me di cuenta que era porque el post solo ponía promesas sin avisar si estaban cumplidas o no 😄. Fuera del lapsus, el post ya está modificado, con todos los cambios aplicados y los links correspondientes. Me disculpo por no dejar los avisos, es un error que no debo dejar pasar otra vez.
