### 2026-09-05 #765 @Rodotasso
Aprovechando, @maelle y @yabellini, si me permiten, me gustaría hacer una limpieza de los test durante este fin de semana. Para dejar los esenciales y eliminar cosas que están duplicadas o que solo ocupan espacio.
