### 2026-09-05 6925959
test: poda y endurecimiento de suite (457→419 tests, cobertura 97.7→98.6%)
- Eliminar duplicados verificados: test-comorbid-validation.R entero,
  copias literales en robustness/coverage-paths/cie-api (C2/C4/C9/C10)
- Convertir 6 tests de red real OMS a mocks deterministicos; smoke
  unico con skip_on_cran/skip_if_offline (C11); fundir API real en cie-api
- Politica skip_on_cran consistente: 42 skips, canarios sin skip por rama
- Estabilizar tests de timing flaky con skip_on_ci (test-integration)
- Corregir tests tautologicos/enganosos: sigla_to_codigo, COVID-19,
  daga/cruz, encoding (titulos
[…]
