+++ tests/testthat/test-cie-sql-cache.R
@@ -18,6 +18,9 @@
 # --- build_fts: failsafe FTS5 missing -------------------------------------
 
 test_that("build_fts crea tabla FTS5 sobre conexion existente", {
+  # Politica CRAN: construccion FTS5 siempre con skip
+  skip_on_cran()
+
   con <- withr::local_db_connection(
     DBI::dbConnect(RSQLite::SQLite(), ":memory:")
   )
@@ -33,6 +36,8 @@ test_that("build_fts crea tabla FTS5 sobre conexion existente", {
 })
 
 # --- cache_is_current ------------------------------------------------------
+# canario CRAN: estos 4 tests corren sin skip en CRAN; usan SQLite
+# :memory: y no construyen el cache del paquete.
 
 test_that("cache_is_current retorna FALSE si no existe tabla cie10_meta", {
   con <- withr::local_db_connection(
@@ -84,6 +89,8 @@ test_that("cache_is_current retorna TRUE cuando version coincide", {
 # --- build_cache_atomic: ciclo completo en directorio aislado --------------
 
 test_that("build_cache_atomic crea cache_dir y construye DB completa", {
+  skip_on_cran()
+
   cache_dir <- withr::local_tempdir()
   db_path <- file.path(cache_dir, "test.db")
 
@@ -109,6 +116,8 @@ test_that("build_cache_atomic crea cache_dir y construye DB completa", {
 })
 
 test_that("build_cache_atomic crea cache_dir cuando no existe", {
+  skip_on_cran()
+
   parent <- withr::local_tempdir()
   cache_dir <- file.path(parent, "subdir_no_existe")
   db_path <- file.path(cache_dir, "test.db")
@@ -122,6 +131,8 @@ test_that("build_cache_atomic crea cache_dir cuando no existe", {
 })
 
 test_that("build_cache_atomic limpia .tmp residual antes de empezar", {
+  skip_on_cran()
+
   cache_dir <- withr::local_tempdir()
   db_path <- file.path(cache_dir, "test.db")
   tmp_path <- paste0(db_path, ".tmp")
@@ -136,40 +147,6 @@ test_that("build_cache_atomic limpia .tmp residual antes de empezar", {
   expect_false(file.exists(tmp_path))
 })
 
-# --- get_cie10_db: invalidacion por version mismatch ----------------------
-
-test_that("get_cie10_db reconstruye cache cuando version no coincide", {
-  cache_dir <- withr::local_tempdir()
-
-  # Aislar cache via env var (R_USER_CACHE_DIR alimenta R_user_dir)
-  withr::local_envvar(R_USER_CACHE_DIR = cache_dir)
-
-  # Asegurar conexion limpia y reset al final
-  cie10_disconnect()
-  withr::defer(cie10_disconnect())
-
-  # Construir cache: la primera llamada genera DB con version actual
-  con1 <- get_cie10_db()
-  expect_true(DBI::dbIsValid(con1))
-
-  # Sobrescribir la version a "0.0.0" para simular cache obsoleto
-  DBI::dbExecute(
-    con1,
-    "UPDATE cie10_meta SET value = '0.0.0' WHERE key = 'cache_version'"
-  )
-
-  # Reset estado interno para forzar reevaluacion
-  cie10_disconnect()
-
-  # Volver a pedir el DB: debe detectar version mismatch y reconstruir
-  con2 <- get_cie10_db()
-  v <- DBI::dbGetQuery(
-    con2,
-    "SELECT value FROM cie10_meta WHERE key = 'cache_version'"
-  )$value[1]
-  expect_equal(v, as.character(utils::packageVersion("ciecl")))
-})
-
 # ============================================================
[…]
