### 2026-09-06 272c122
test: fase 6A+6B - eliminaciones puras y parametrizaciones (419→384 tests)
- 6A: 17 replicas literales eliminadas (C6, C12, C7, cie-guide,
  robustness threshold 0, cie-table triviales)
- 6B: parametrizaciones comorbid (Elixhauser 4→1, EPOC absorbido,
  diabetes/infarto 2→1), cie-sql validaciones 4→1, dplyr 3→1
- Cobertura intacta 98.55% (0 lineas perdidas vs base covr fase 5)
- Auditado y aprobado (r-code-reviewer)
