+++ tests/testthat/test-cie-comorbid.R
@@ -20,34 +20,23 @@ test_that("cie_comorbid calcula Charlson", {
   expect_true("score_charlson" %in% names(resultado))
 })
 
-test_that("cie_comorbid Charlson con diabetes tipo 2", {
+test_that("cie_comorbid Charlson con diabetes tipo 1 y 2", {
   skip_on_cran()
   skip_if_not_installed("comorbidity")
 
   df <- data.frame(
-    id = 1,
-    diag = "E11.0"  # Diabetes mellitus tipo 2
+    id = c(1, 2),
+    diag = c("E11.0", "E10.9")  # Diabetes mellitus tipo 2 y tipo 1
   )
 
   resultado <- cie_comorbid(df, id = "id", code = "diag", map = "charlson")
   expect_s3_class(resultado, "tbl_df")
   expect_true("score_charlson" %in% names(resultado))
-  # Diabetes sin complicaciones = 1 punto en Charlson
-  expect_gte(resultado$score_charlson[1], 0)
-})
-
-test_that("cie_comorbid Charlson con diabetes tipo 1", {
-  skip_on_cran()
-  skip_if_not_installed("comorbidity")
-
-  df <- data.frame(
-    id = 1,
-    diag = "E10.9"  # Diabetes mellitus tipo 1
-  )
-
-  resultado <- cie_comorbid(df, id = "id", code = "diag", map = "charlson")
-  expect_s3_class(resultado, "tbl_df")
-  expect_true("score_charlson" %in% names(resultado))
+  # Diabetes sin complicaciones = 1 punto en Charlson (valores
+  # verificados con comorbidity 1.1.0, mapa charlson_icd10_quan).
+  # as.numeric: comorbidity::score() adjunta atributos map/weights
+  expect_equal(resultado$diab, c(1, 1))
+  expect_equal(as.numeric(resultado$score_charlson), c(1, 1))
 })
 
 test_that("cie_comorbid Charlson con infarto miocardio", {
@@ -62,6 +51,9 @@ test_that("cie_comorbid Charlson con infarto miocardio", {
   resultado <- cie_comorbid(df, id = "id", code = "diag", map = "charlson")
   expect_s3_class(resultado, "tbl_df")
   expect_true("score_charlson" %in% names(resultado))
+  # IAM = 1 punto en Charlson (verificado con comorbidity 1.1.0)
+  expect_equal(resultado$mi[1], 1)
+  expect_equal(as.numeric(resultado$score_charlson[1]), 1)
 })
 
 test_that("cie_comorbid Charlson con cancer", {
@@ -76,6 +68,9 @@ test_that("cie_comorbid Charlson con cancer", {
   resultado <- cie_comorbid(df, id = "id", code = "diag", map = "charlson")
   expect_s3_class(resultado, "tbl_df")
   expect_length(resultado$id, 2)
+  # Cancer sin metastasis = 2 puntos c/u (verificado comorbidity 1.1.0)
+  expect_equal(resultado$canc, c(1, 1))
+  expect_equal(as.numeric(resultado$score_charlson), c(2, 2))
 })
 
 test_that("cie_comorbid Charlson con insuficiencia cardiaca", {
@@ -89,6 +84,9 @@ test_that("cie_comorbid Charlson con insuficiencia cardiaca", {
 
   resultado <- cie_comorbid(df, id = "id", code = "diag", map = "charlson")
   expect_s3_class(resultado, "tbl_df")
+  # ICC = 1 punto en Charlson (verificado con comorbidity 1.1.0)
+  expect_equal(resultado$chf[1], 1)
+  expect_equal(as.numeric(resultado$score_charlson[1]), 1)
 })
 
[…]
