### 2026-09-07 84983ca
test: fase 6E + fixes auditoria rOpenSci (376->372 tests)
- 6E: elimina 5 flujos E2E redundantes de test-integration.R
- extrae test-cie-map-comorbid.R (comorbid bajo 600 lineas)
- canario CRAN ligero en test-integration.R
- test cie10_clear_cache verifica invisibilidad (withVisible)
- politica canario skip_on_cran documentada en setup.R
