+++ tests/testthat/test-cie-map-comorbid.R
@@ -1,0 +1,180 @@
+# ============================================================
+# PRUEBAS cie_map_comorbid()
+# ============================================================
+
+# ------------------------------------------------------------------------------
+# Pruebas basicas de mapeo
+# ------------------------------------------------------------------------------
+
+test_that("cie_map_comorbid categoriza correctamente", {
+  codigos <- c("E11.0", "I21.0", "INVALIDO")
+  expect_warning(resultado <- cie_map_comorbid(codigos), "formato CIE-10")
+
+  expect_equal(nrow(resultado), 3)
+  expect_equal(resultado$categoria[1], "Diabetes")
+})
+
+test_that("cie_map_comorbid retorna tibble", {
+  codigos <- c("E11.0", "I50.9")
+  resultado <- cie_map_comorbid(codigos)
+
+  expect_s3_class(resultado, "tbl_df")
+  expect_true("codigo" %in% names(resultado))
+  expect_true("categoria" %in% names(resultado))
+})
+
+test_that("cie_map_comorbid maneja vector vacio", {
+  resultado <- cie_map_comorbid(character(0))
+
+  expect_s3_class(resultado, "tbl_df")
+  expect_equal(nrow(resultado), 0)
+  expect_true("codigo" %in% names(resultado))
+  expect_true("categoria" %in% names(resultado))
+})
+
+# ------------------------------------------------------------------------------
+# Pruebas de mapeo Charlson
+# ------------------------------------------------------------------------------
+
+test_that("cie_map_comorbid categoriza diabetes E10 y E11", {
+  resultado <- cie_map_comorbid(c("E10.9", "E11.0"))
+  expect_equal(resultado$categoria, c("Diabetes", "Diabetes"))
+})
+
+test_that("cie_map_comorbid categoriza infarto miocardio I21 e I22", {
+  resultado <- cie_map_comorbid(c("I21.0", "I22.0"))
+  expect_true(all(resultado$categoria == "Infarto miocardio"))
+})
+
+test_that("cie_map_comorbid categoriza neoplasia maligna", {
+  codigos <- c("C50.9", "C34.9", "C18.9")
+  resultado <- cie_map_comorbid(codigos)
+  expect_true(all(resultado$categoria == "Neoplasia maligna"))
+})
+
+test_that("cie_map_comorbid categoriza todo espectro EPOC J40-J44", {
+  codigos <- c("J40", "J41.0", "J42", "J43.9", "J44.1", "J44.9")
+  resultado <- cie_map_comorbid(codigos)
+  expect_true(all(resultado$categoria == "EPOC"))
+})
+
+test_that("cie_map_comorbid categoriza enfermedad renal cronica", {
+  resultado <- cie_map_comorbid("N18.9")
+  expect_equal(resultado$categoria[1], "Enfermedad renal cronica")
+})
+
+test_that("cie_map_comorbid categoriza trastornos mentales", {
+  codigos <- c("F32.9", "F41.9", "F20.0")
+  resultado <- cie_map_comorbid(codigos)
+  expect_true(all(resultado$categoria == "Trastornos mentales"))
+})
+
+# ------------------------------------------------------------------------------
+# Pruebas de codigos sin categoria
+# ------------------------------------------------------------------------------
+
+test_that("cie_map_comorbid asigna Otra a codigos no mapeados", {
+  resultado <- cie_map_comorbid("Z00.0")
+  expect_equal(resultado$categoria[1], "Otra")
+})
+
+test_that("cie_map_comorbid asigna Otra a codigos invalidos y advierte", {
[…]
