+++ tests/testthat/test-integration.R
@@ -55,28 +55,6 @@ test_that("flujo: buscar categoria -> expandir -> calcular comorbilidad", {
   expect_true("score_charlson" %in% names(resultado_comorbid))
 })
 
-test_that("flujo: normalizar codigos -> buscar -> mapear comorbilidad", {
-  skip_on_cran()
-
-  # 1. Codigos en formato mixto (como vendrian de datos reales)
-  codigos_raw <- c("E110", "I509", "C509", "e11.0", " Z00 ")
-
-  # 2. Normalizar codigos
-  codigos_norm <- cie_norm(codigos_raw, search_db = FALSE)
-  expect_length(codigos_norm, length(codigos_raw))
-
-  # 3. Buscar detalles
-  suppressMessages({
-    detalles <- cie_lookup(codigos_norm)
-  })
-  expect_gt(nrow(detalles), 0)
-
-  # 4. Mapear a categorias de comorbilidad
-  mapa <- cie_map_comorbid(codigos_norm)
-  expect_length(codigos_norm, nrow(mapa))
-  expect_true("Diabetes" %in% mapa$categoria)
-})
-
 test_that("flujo: SQL personalizado -> procesamiento -> validacion", {
   skip_on_cran()
 
@@ -130,99 +108,9 @@ test_that("cie_expand y cie_lookup expandir dan mismos resultados", {
   expect_setequal(hijos_expand, hijos_lookup)
 })
 
-test_that("cie_norm y cie_lookup son coherentes", {
-  skip_on_cran()
-
-  # Codigo sin punto
-  codigo_raw <- "E110"
-
-  # Normalizar
-  codigo_norm <- cie_norm(codigo_raw, search_db = FALSE)
-  expect_equal(codigo_norm, "E11.0")
-
-  # Buscar con codigo raw y normalizado
-  resultado_raw <- cie_lookup(codigo_raw)
-  resultado_norm <- cie_lookup(codigo_norm)
-
-  # Deben dar mismo resultado
-  expect_equal(resultado_raw$codigo, resultado_norm$codigo)
-})
-
-test_that("cie_validate_vector y cie_lookup son coherentes", {
-  skip_on_cran()
-
-  codigos <- c("E11.0", "INVALIDO", "Z00")
-
-  # Validar
-  validacion <- cie_validate_vector(codigos)
-  expect_equal(validacion, c(TRUE, FALSE, TRUE))
-
-  # Buscar (solo los validos deben retornar)
-  suppressMessages({
-    resultado <- cie_lookup(codigos)
-  })
-
-  # Solo codigos validos en resultado
-  expect_true(all(resultado$codigo %in% codigos[validacion]))
-})
-
 # ============================================================
 # PRUEBAS DE ESCENARIOS REALES
 # ============================================================
-
-test_that("escenario: analisis de egreso hospitalario", {
-  skip_on_cran()
-  skip_if_not_installed("comorbidity")
-
-  # Simular datos de egresos hospitalarios
-  set.seed(42)
-  egresos <- data.frame(
-    id_egreso = 1:20,
-    codigo_principal = sample(c("E11.0", "I50.9", "C50.9", "J18.9", "K70.3"), 20, replace = TRUE),
-    codigo_secundario = sample(c("E11.9", "I10", "Z86.7", "N18.9", "F32.9"), 20, replace = TRUE)
-  )
-
-  # 1. Validar codigos principales
-  val_principal <- cie_validate_vector(egresos$codigo_principal)
-  expect_true(all(val_principal))
-
-  # 2. Validar codigos secundarios
-  val_secundario <- cie_validate_vector(egresos$codigo_secundario)
-  expect_true(all(val_secundario))
-
-  # 3. Preparar datos para comorbilidad (formato largo)
[…]
