### 2026-09-07 852d6f4
docs: NEWS de la poda y endurecimiento de la suite de tests (ES/EN)
