### 2026-09-09 1c2fe47
feat: ejemplos ejecutables en pkgdown con IN_PKGDOWN y cassette vcr para cie11_search (rOpenSci #765)
