+++ R/cie-comorbid.R
@@ -18,7 +18,7 @@ NULL
 #' # Ver documentacion de parametros
 #' args(cie_comorbid)
 #'
-#' @examplesIf interactive()
+#' @examplesIf rlang::is_interactive() || identical(Sys.getenv("IN_PKGDOWN"), "true")
 #' df <- data.frame(
 #'   id_pac = c(1, 1, 2, 2),
 #'   diag = c("E11.0", "I21.0", "C50.9", "E10.9")
