+++ DESCRIPTION
@@ -50,6 +50,6 @@ URL: https://rodotasso.github.io/ciecl/,
     https://github.com/Rodotasso/ciecl
 BugReports: https://github.com/Rodotasso/ciecl/issues
 Config/testthat/edition: 3
-Config/Needs/website: pkgdown
+Config/Needs/website: pkgdown, vcr
 Language: en, es
 Config/roxygen2/version: 8.0.0
