+++ reference/cie_lookup.md
@@ -115,23 +115,107 @@ cie_lookup("E11.0")
 #> # ℹ 4 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>,
 #> #   uso_cl <chr>
 
-if (FALSE) { # interactive()
 cie_lookup("E110") # Sin punto
+#> # A tibble: 1 × 11
+#>   codigo descripcion       categoria seccion capitulo_nombre inclusion exclusion
+#>   <chr>  <chr>             <chr>     <chr>   <chr>           <chr>     <chr>    
+#> 1 E11.0  Diabetes mellitu… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#> # ℹ 4 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>,
+#> #   uso_cl <chr>
 cie_lookup("E11") # Solo categoría
+#> # A tibble: 1 × 11
+#>   codigo descripcion       categoria seccion capitulo_nombre inclusion exclusion
+#>   <chr>  <chr>             <chr>     <chr>   <chr>           <chr>     <chr>    
+#> 1 E11    Diabetes mellitu… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#> # ℹ 4 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>,
+#> #   uso_cl <chr>
 cie_lookup("E11", expand = TRUE) # Todos E11.x
+#> # A tibble: 11 × 11
+#>    codigo descripcion      categoria seccion capitulo_nombre inclusion exclusion
+#>    <chr>  <chr>            <chr>     <chr>   <chr>           <chr>     <chr>    
+#>  1 E11    Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#>  2 E11.0  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#>  3 E11.1  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#>  4 E11.2  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#>  5 E11.3  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#>  6 E11.4  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#>  7 E11.5  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#>  8 E11.6  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#>  9 E11.7  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#> 10 E11.8  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#> 11 E11.9  Diabetes mellit… E11 DIAB… E08-E1… Cap.04  ENFERM… NA        NA       
+#> # ℹ 4 more variables: capitulo <chr>, es_daga <int>, es_cruz <int>,
+#> #   uso_cl <chr>
 # Vectorizado - múltiples códigos y formatos
 cie_lookup(c("E11.0", "Z00", "I10"))
+#> # A tibble: 3 × 11
+#>   codigo descripcion       categoria seccion capitulo_nombre inclusion exclusion
+#>   <chr>  <chr>             <chr>     <chr>   <chr>           <chr>     <chr>    
[…]
