+++ reference/cie11_search.md
@@ -100,4 +100,33 @@ cie11_search("depresion mayor")
 #>  8 6A71.7 Trastorno depresivo recurrente, actualmente en remisión comp… 06      
 #>  9 MB47.Y Otra alteración especificada del tono muscular y los reflejos 21      
 #> 10 6A70.Z Trastorno depresivo de episodio único, sin especificación     06      
+# Causas frecuentes de egreso en Chile
+cie11_search("diabetes mellitus")
+#> # A tibble: 10 × 3
+#>    codigo  titulo                                                       capitulo
+#>    <chr>   <chr>                                                        <chr>   
+#>  1 5A14    Diabetes mellitus, tipo no especificado                      05      
+#>  2 5A13    Diabetes mellitus, otro tipo especificado                    05      
+#>  3 5A13.Y  Otra diabetes mellitus especificada, otro tipo especificado  05      
+#>  4 JA63.Z  Diabetes mellitus en el embarazo, sin especificación         18      
+#>  5 JA63.2  Diabetes mellitus que aparece por primera vez en el embarazo 18      
+#>  6 KB60.2Z Diabetes mellitus neonatal, sin especificación               19      
+#>  7 KB60.2Y Otra diabetes mellitus neonatal especificada                 19      
+#>  8 5A10    Diabetes mellitus tipo 1                                     05      
+#>  9 5A11    Diabetes mellitus tipo 2                                     05      
+#> 10 5A12    Diabetes mellitus por malnutrición                           05      
+cie11_search("hipertension esencial")
+#> # A tibble: 10 × 3
+#>    codigo  titulo                                      capitulo
+#>    <chr>   <chr>                                       <chr>   
+#>  1 BA00.Z  Hipertensión esencial, sin especificación   11      
+#>  2 BA00.Y  Otra hipertensión esencial especificada     11      
+#>  3 KB45    Hipertensión neonatal                       19      
+#>  4 JA23    Hipertensión gestacional                    18      
+#>  5 BA04.Z  Hipertensión secundaria, sin especificación 11      
+#>  6 BA04.Y  Otra hipertension secundaria especificada   11      
+#>  7 9C61.01 Hipertensión ocular                         09      
+#>  8 DB98.7Z Hipertensión portal, sin especificación     13      
+#>  9 DB98.7Y Otra hipertensión portal especificada       13      
+#> 10 BA00.2  Hipertensión sistólica aislada              11      
 ```
