+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   instalacion: instalacion.html
   installation: installation.html
   languages: languages.html
-last_built: 2026-09-07T00:50Z
+last_built: 2026-09-09T15:05Z
 urls:
   reference: https://rodotasso.github.io/ciecl/reference
   article: https://rodotasso.github.io/ciecl/articles

[…]
