### 2026-09-14 759d434
test: reemplazar test artesanal de version-mismatch por mock de packageVersion (rOpenSci #765)
