### 2026-09-14 371c559
fix: robustez y consistencia en cie_lookup, cie_search, cie_expand y cie11_search
- cie_lookup(): rangos incluyen subcategorias del limite superior
  (E10-E14 ahora cubre E14.x; BETWEEN lexicografico las omitia)
- cie_lookup(): extract=TRUE valida contrato escalar; modo vectorial
  informa codigos invalidos/no encontrados en mensaje agregado
- cie_expand(): valida string escalar (NA aborta con error claro)
- cie_search(): valida tipo/NA de threshold y max_results; columna
  uso_cl presente en todos los caminos; only_uso_cl filtra antes
  de aplicar max_results
- cie11_search(): valida NA en ma
[…]
