### 2026-09-15 a356f44
fix: hardening SQL, cache y validaciones en cie_comorbid
- cie10_sql(): strip de literales/comentarios antes del blocklist
  (elimina falsos positivos; protecciones vigentes sin cambio)
- cache: CIECL_CACHE_DIR vacia cae a R_user_dir; retornos de
  file.rename/file.remove con warning informativo
- cie_norm(): E11.X ya no deja punto colgado
- cie_comorbid(): @returns corregido a tibble; validacion de
  data/id/code con clase ciecl_invalid_input
- cie_search(): entrada solo-simbolos retorna tibble vacio sin NaN
- limpieza de @importFrom sin uso (pull, rowwise, tribble, matches)
