+++ R/cie-sql.R
@@ -8,7 +8,14 @@
 #' @keywords internal
 #' @noRd
 get_cache_dir <- function() {
-  Sys.getenv("CIECL_CACHE_DIR", unset = tools::R_user_dir("ciecl", "data"))
+  # Una env var seteada pero vacia ("") cuenta como no definida:
+  # file.path("", "cie10.db") escribiria el cache en el directorio de
+  # trabajo, fuera del directorio designado del usuario (CRAN policy)
+  env <- Sys.getenv("CIECL_CACHE_DIR", unset = NA_character_)
+  if (is.na(env) || !nzchar(env)) {
+    return(tools::R_user_dir("ciecl", "data"))
+  }
+  env
 }
 
 #' Obtener conexion SQLite pooled CIE-10
@@ -190,10 +197,14 @@ build_cache_atomic <- function(cache_dir, db_path) {
       DBI::dbDisconnect(con)
 
       # Atomico: renombrar .tmp -> .db
-      if (file.exists(db_path)) {
-        file.remove(db_path)
+      # En Windows el rename/remove puede fallar si otro proceso tiene
+      # un lock sobre el .db: no silenciar el fallo
+      if (file.exists(db_path) && !file.remove(db_path)) {
+        cli::cli_warn("No se pudo eliminar el cache previo: {.path {db_path}}")
       }
-      file.rename(tmp_path, db_path)
+      if (!file.rename(tmp_path, db_path)) {
+        cli::cli_warn("No se pudo renombrar el cache temporal a {.path {db_path}}.")
+      }
 
       if (show_progress) {
         cli::cli_progress_done(id = progress_id)
@@ -328,7 +339,19 @@ cie10_sql <- function(query, close = lifecycle::deprecated()) {
     cli::cli_abort("Solo consultas {.code SELECT} permitidas (seguridad).", class = "ciecl_unsafe_query")
   }
 
-  # Bloquear keywords peligrosos (case-insensitive)
+  # Remover strings, comentarios de linea (--) y comentarios de bloque
+  # (/* */) ANTES de escanear: el blocklist y el chequeo de ";" deben
+  # operar sobre el SQL ejecutable, no sobre literales legitimos
+  # (ej. LIKE '%drop%') ni comentarios
+  query_sin_strings <- query_norm
+  query_sin_strings <- stringr::str_remove_all(query_sin_strings, "'[^']*'")
+  query_sin_strings <- stringr::str_remove_all(query_sin_strings, "--[^\n]*")
+  query_sin_strings <- stringr::str_remove_all(
+    query_sin_strings, "(?s)/\\*.*?\\*/"
+  )
+
+  # Bloquear keywords peligrosos (case-insensitive) sobre el SQL ya
+  # limpio de literales y comentarios
   keywords_peligrosos <- c(
     "\\bDROP\\b", "\\bDELETE\\b", "\\bUPDATE\\b", "\\bINSERT\\b",
     "\\bALTER\\b", "\\bCREATE\\b", "\\bTRUNCATE\\b", "\\bEXEC\\b",
@@ -338,7 +361,7 @@ cie10_sql <- function(query, close = lifecycle::deprecated()) {
 
   for (keyword in keywords_peligrosos) {
     keyword_found <- stringr::str_detect(
-      query_norm, stringr::regex(keyword, ignore_case = TRUE)
+      query_sin_strings, stringr::regex(keyword, ignore_case = TRUE)
     )
     if (keyword_found) {
       # Mostrar la palabra clave detectada (sin los anclajes \b del regex)
@@ -351,13 +374,6 @@ cie10_sql <- function(query, close = lifecycle::deprecated()) {
   }
 
[…]
