### 2026-09-15 7b55492
fix: nits en cie10_sql, cie_table y cie_describe
- cie10_sql(): valida ^SELECT tras remover comentarios/strings (queries con comentario inicial ya no abortan)
- cie_table(): valida code escalar (largo >1 aborta con error claro)
- cie_describe(): valida default (character escalar o NA) y coerciona a character
- docs: tildes en roxygen de get_siglas_medicas; NEWS actualizado
