+++ tests/testthat/test-cie-table.R
@@ -11,14 +11,6 @@ test_that("cie_table requiere gt instalado", {
   expect_s3_class(tabla, "gt_tbl")
 })
 
-test_that("cie_table funciona con multiples codigos", {
-  skip_if_not_installed("gt")
-  skip_on_cran()
-
-  tabla <- cie_table(c("E11", "I10"))
-  expect_s3_class(tabla, "gt_tbl")
-})
-
 test_that("cie_table maneja codigo invalido", {
   testthat::local_reproducible_output()
   skip_if_not_installed("gt")
@@ -82,4 +74,10 @@ test_that("cie_table error sin gt instalado", {
     cie_table("E11"),
     "gt"
   )
+})
+
+test_that("cie_table valida code escalar (F15)", {
+  expect_error(cie_table(c("E11", "I10")), class = "ciecl_invalid_input")
+  expect_error(cie_table(123), class = "ciecl_invalid_input")
+  expect_error(cie_table(NA_character_), class = "ciecl_invalid_code")
 })
