### 2026-09-17 69f973c
fix: validaciones escalares tipadas, ORDER BY determinista y limpieza de imports sin uso
- cie_short(): category valida como string escalar (ciecl_invalid_input)
- cie_table(): code con is_string; NA aborta como input invalido
- cie_lookup(): ORDER BY codigo en modo vector exacto (IN)
- sigla_to_codigo(): @returns corregido (primer codigo o NULL)
- tests: aserciones de warning rango invertido, errores tipados de
  threshold/max_results, coercion NA_real_/NaN en cie_describe, NULL
  del camino FTS de sigla_to_codigo, cie_table(NA)
- tests: literales non-ASCII a escapes \uXXXX en test-encoding
-
[…]
