### 2026-09-17 3988e1e
docs: sincroniza cran-comments, READMEs, NEWS, codemeta y vignettes con 0.9.8
- cran-comments: version 0.9.8, cifras medidas (1059 PASS, cobertura
  97.39%), sin referencia a vigneta inexistente, justificacion del
  ratio skip_on_cran (politica canario CRAN), notes etiquetadas como
  incoming-only
- README.Rmd: chunks de instalacion y comorbilidad parametrizados por
  idioma; README.en.md sin comentarios en espanol
- NEWS: fecha de corte 2026-09-16 y entradas de la tanda, fieles al
  codigo (Config/roxygen2/version se mantiene: document() la re-escribe)
- codemeta: dateModified 2026-09-16, wri
[…]
