+++ README.Rmd
@@ -98,11 +98,20 @@ The dataset is established by [Decree 356/2017](https://www.bcn.cl/leychile/nave
 cat(if (lang == "es") "## Instalación\n\n" else "## Installation\n\n")
 ```
 
-```r
+```{r install-es, eval=FALSE, echo=(params$lang == "es")}
 # CRAN
 install.packages("ciecl")
 
 # GitHub (desarrollo)
+# install.packages("pak")
+pak::pak("RodoTasso/ciecl")
+```
+
+```{r install-en, eval=FALSE, echo=(params$lang == "en")}
+# CRAN
+install.packages("ciecl")
+
+# GitHub (development)
 # install.packages("pak")
 pak::pak("RodoTasso/ciecl")
 ```
@@ -167,9 +176,14 @@ if (lang == "es") {
 }
 ```
 
-```{r comorb-example, eval=FALSE}
+```{r comorb-es, eval=FALSE, echo=(params$lang == "es")}
 # Comorbilidades (requiere: install.packages("comorbidity"))
 df |> cie_comorbid(id = "paciente", code = "diagnostico", map = "charlson")
+```
+
+```{r comorb-en, eval=FALSE, echo=(params$lang == "en")}
+# Comorbidities (requires: install.packages("comorbidity"))
+df |> cie_comorbid(id = "patient", code = "diagnosis", map = "charlson")
 ```
 
 ```{r api-intro, results='asis', echo=FALSE}
