+++ reference/cie_guia_busqueda.md
@@ -29,14 +29,14 @@ cie_guia_busqueda()
 #> Warning: `cie_guia_busqueda()` was deprecated in ciecl 0.9.8.
 #> ℹ Please use `cie_guide()` instead.
 #> # A tibble: 8 × 3
-#>   Tengo...                     `Usar funcion`                            Ejemplo
-#>   <chr>                        <chr>                                     <chr>  
-#> 1 Codigo exacto (E11.0)        cie_lookup()                              "cie_l…
-#> 2 Codigo sin punto (e110)      cie_lookup()                              "cie_l…
-#> 3 Codigo con espacios (E 11.0) cie_lookup()                              "cie_l…
-#> 4 Codigo con prefijos/sufijos  cie_lookup(extract = TRUE)                "cie_l…
-#> 5 Codigo categoria (E11)       cie_lookup() o cie_lookup(expandir = TRU… "cie_l…
-#> 6 Descripcion (diabetes)       cie_search()                              "cie_s…
-#> 7 Sigla medica (IAM, DM2)      cie_lookup(check_siglas = TRUE)           "cie_l…
-#> 8 No se que tengo              cie_search() con threshold bajo           "cie_s…
+#>   Tengo...                     `Usar funcion`                           Ejemplo 
+#>   <chr>                        <chr>                                    <chr>   
+#> 1 Codigo exacto (E11.0)        cie_lookup()                             "cie_lo…
+#> 2 Codigo sin punto (e110)      cie_lookup()                             "cie_lo…
+#> 3 Codigo con espacios (E 11.0) cie_lookup()                             "cie_lo…
+#> 4 Codigo con prefijos/sufijos  cie_lookup(extract = TRUE)               "cie_lo…
+#> 5 Codigo categoria (E11)       cie_lookup() o cie_lookup(expand = TRUE) "cie_lo…
+#> 6 Descripcion (diabetes)       cie_search()                             "cie_se…
+#> 7 Sigla medica (IAM, DM2)      cie_lookup(check_siglas = TRUE)          "cie_lo…
+#> 8 No se que tengo              cie_search() con threshold bajo          "cie_se…
 ```
