+++ news/index.md
@@ -1,8 +1,217 @@
 # Changelog
 
-## ciecl 0.9.8 (en desarrollo, 2026-04-25 → 2026-08-27)
+## ciecl 0.9.8 (en desarrollo, 2026-04-25 → 2026-09-16)
 
 *English summary below*
+
+### Correcciones menores — nits de auditoría (2026-09-16)
+
+Sin cambios en la API pública; dos cambios de comportamiento acotados:
+[`cie_short()`](https://rodotasso.github.io/ciecl/reference/cie_short.md)
+ahora valida `category` y `cie_table(NA)` aborta como input inválido en
+lugar de reportar “código no encontrado: NA”.
+
+- **Validaciones endurecidas** (clase `ciecl_invalid_input`):
+  [`cie_short()`](https://rodotasso.github.io/ciecl/reference/cie_short.md)
+  valida que `category` sea NULL o un string escalar no-NA (antes un
+  vector o numérico derivaba en un error duro no tipado);
+  [`cie_table()`](https://rodotasso.github.io/ciecl/reference/cie_table.md)
+  rechaza `code = NA` como input inválido (antes caía en “Código no
+  encontrado: NA”).
+- **Orden determinista en
+  [`cie_lookup()`](https://rodotasso.github.io/ciecl/reference/cie_lookup.md)**:
+  el modo vector exacto (`IN (...)`) ahora incluye `ORDER BY codigo`,
+  igual que los demás caminos de la función.
+- **Documentación interna corregida**: el `@returns` de
+  `sigla_to_codigo()` refleja que retorna el primer código (string de
+  longitud 1) o NULL.
+- **Limpieza de metadatos**: se completa la eliminación del
+  `@importFrom dplyr select` sin uso real (los usos en `R/` ya eran
+  calificados
+  [`dplyr::select()`](https://dplyr.tidyverse.org/reference/select.html));
+  `writexl` sale de `Suggests` (sin uso ejecutable en el paquete) y de
+  las tablas de instalación de las viñetas; se verificó que
+  `Config/roxygen2/version` en DESCRIPTION es el mecanismo estándar del
+  roxygen2 vigente (reemplaza a `RoxygenNote`; `document()` la
+  re-escribe automáticamente) y se mantiene con el valor correcto;
+  `codemeta.json` registra `dateModified`.
+- **README y viñetas bilingües**: los chunks de instalación y
+  comorbilidad del README ahora respetan la condicional de idioma (el
+  README en inglés ya no muestra comentarios en español); el encabezado
+  final de la viñeta en inglés se alinea al registro de la española
+  (“Collaboration and Support”).
+- **Tests**: nuevas aserciones para el warning de rango invertido en
+  [`cie_lookup()`](https://rodotasso.github.io/ciecl/reference/cie_lookup.md),
+  los errores tipados de `threshold`/`max_results` en
+  [`cie_search()`](https://rodotasso.github.io/ciecl/reference/cie_search.md),
+  la coerción de `default = NA_real_`/`NaN` en
+  [`cie_describe()`](https://rodotasso.github.io/ciecl/reference/cie_describe.md)
+  y el NULL del camino sin match de `sigla_to_codigo()`; literales
+  non-ASCII de `test-encoding.R` convertidos a escapes `\uXXXX`.
+
+*Input validation hardened (class `ciecl_invalid_input`):
+[`cie_short()`](https://rodotasso.github.io/ciecl/reference/cie_short.md)
+requires `category` to be NULL or a non-NA length-1 string, and
+`cie_table(NA)` now aborts as invalid input instead of reporting “code
+not found: NA”. Exact vector mode in
+[`cie_lookup()`](https://rodotasso.github.io/ciecl/reference/cie_lookup.md)
[…]
