+++ README.en.md
@@ -65,7 +65,7 @@ institutional decree.
 # CRAN
 install.packages("ciecl")
 
-# GitHub (desarrollo)
+# GitHub (development)
 # install.packages("pak")
 pak::pak("RodoTasso/ciecl")
 ```
@@ -146,8 +146,8 @@ and an example for each case.
 
 ``` r
 
-# Comorbilidades (requiere: install.packages("comorbidity"))
-df |> cie_comorbid(id = "paciente", code = "diagnostico", map = "charlson")
+# Comorbidities (requires: install.packages("comorbidity"))
+df |> cie_comorbid(id = "patient", code = "diagnosis", map = "charlson")
 ```
 
 ## ICD-11 API (optional)
